Compare categorical variables

The compare_category() compute information to examine the relationship between categorical variables.

compare_category(.data, ...)

# S3 method for data.frame
compare_category(.data, ...)

Arguments

.data
a data.frame or a tbl_df.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

Value

An object of the class as compare based list. The information to examine the relationship between categorical variables is as follows each components.

var1 : factor. The level of the first variable to compare. 'var1' is the name of the first variable to be compared.

var2 : factor. The level of the second variable to compare. 'var2' is the name of the second variable to be compared.

n : integer. frequency by var1 and var2.

rate : double. relative frequency.

first_rate : double. relative frequency in first variable.

second_rate : double. relative frequency in second variable.

Details

It is important to understand the relationship between categorical variables in EDA. compare_category() compares relations by pair combination of all categorical variables. and return compare_category class that based list object.

Attributes of return object

Attributes of compare_category class is as follows.

variables : character. List of variables selected for comparison.

combination : matrix. It consists of pairs of variables to compare.

See also

Examples

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

library(dplyr)
library(stringr)

# Compare the all categorical variables
all_var <- compare_category(carseats)

# Print compare_numeric class objects
all_var

#> $`ShelveLoc vs Urban`
#> # A tibble: 9 x 6
#>   ShelveLoc Urban     n   rate var1_rate var2_rate
#>   <fct>     <fct> <int>  <dbl>     <dbl>     <dbl>
#> 1 Bad       No       22 0.055     0.229      0.186
#> 2 Bad       Yes      73 0.182     0.760      0.264
#> 3 Bad       NA        1 0.0025    0.0104     0.2  
#> 4 Good      No       28 0.07      0.329      0.237
#> 5 Good      Yes      56 0.14      0.659      0.202
#> 6 Good      NA        1 0.0025    0.0118     0.2  
#> 7 Medium    No       68 0.17      0.311      0.576
#> 8 Medium    Yes     148 0.37      0.676      0.534
#> 9 Medium    NA        3 0.0075    0.0137     0.6  
#> 
#> $`ShelveLoc vs US`
#> # A tibble: 6 x 6
#>   ShelveLoc US        n  rate var1_rate var2_rate
#>   <fct>     <fct> <int> <dbl>     <dbl>     <dbl>
#> 1 Bad       No       34 0.085     0.354     0.239
#> 2 Bad       Yes      62 0.155     0.646     0.240
#> 3 Good      No       24 0.06      0.282     0.169
#> 4 Good      Yes      61 0.152     0.718     0.236
#> 5 Medium    No       84 0.21      0.384     0.592
#> 6 Medium    Yes     135 0.338     0.616     0.523
#> 
#> $`Urban vs US`
#> # A tibble: 6 x 6
#>   Urban US        n   rate var1_rate var2_rate
#>   <fct> <fct> <int>  <dbl>     <dbl>     <dbl>
#> 1 No    No       46 0.115      0.390   0.324  
#> 2 No    Yes      72 0.18       0.610   0.279  
#> 3 Yes   No       95 0.238      0.343   0.669  
#> 4 Yes   Yes     182 0.455      0.657   0.705  
#> 5 NA    No        1 0.0025     0.2     0.00704
#> 6 NA    Yes       4 0.01       0.8     0.0155 
#> 

# Compare the categorical variables that case of joint the US variable
all_var %>%
  "["(str_detect(names(all_var), "US"))

#> $`ShelveLoc vs US`
#> # A tibble: 6 x 6
#>   ShelveLoc US        n  rate var1_rate var2_rate
#>   <fct>     <fct> <int> <dbl>     <dbl>     <dbl>
#> 1 Bad       No       34 0.085     0.354     0.239
#> 2 Bad       Yes      62 0.155     0.646     0.240
#> 3 Good      No       24 0.06      0.282     0.169
#> 4 Good      Yes      61 0.152     0.718     0.236
#> 5 Medium    No       84 0.21      0.384     0.592
#> 6 Medium    Yes     135 0.338     0.616     0.523
#> 
#> $`Urban vs US`
#> # A tibble: 6 x 6
#>   Urban US        n   rate var1_rate var2_rate
#>   <fct> <fct> <int>  <dbl>     <dbl>     <dbl>
#> 1 No    No       46 0.115      0.390   0.324  
#> 2 No    Yes      72 0.18       0.610   0.279  
#> 3 Yes   No       95 0.238      0.343   0.669  
#> 4 Yes   Yes     182 0.455      0.657   0.705  
#> 5 NA    No        1 0.0025     0.2     0.00704
#> 6 NA    Yes       4 0.01       0.8     0.0155 
#> 

# Compare the two categorical variables
two_var <- compare_category(carseats, ShelveLoc, Urban)

# Print compare_numeric class objects
two_var

#> $`ShelveLoc vs Urban`
#> # A tibble: 9 x 6
#>   ShelveLoc Urban     n   rate var1_rate var2_rate
#>   <fct>     <fct> <int>  <dbl>     <dbl>     <dbl>
#> 1 Bad       No       22 0.055     0.229      0.186
#> 2 Bad       Yes      73 0.182     0.760      0.264
#> 3 Bad       NA        1 0.0025    0.0104     0.2  
#> 4 Good      No       28 0.07      0.329      0.237
#> 5 Good      Yes      56 0.14      0.659      0.202
#> 6 Good      NA        1 0.0025    0.0118     0.2  
#> 7 Medium    No       68 0.17      0.311      0.576
#> 8 Medium    Yes     148 0.37      0.676      0.534
#> 9 Medium    NA        3 0.0075    0.0137     0.6  
#> 

# Filtering the case of US included NA 
two_var %>%
  "[["(1) %>%
  filter(!is.na(Urban))

#> # A tibble: 6 x 6
#>   ShelveLoc Urban     n  rate var1_rate var2_rate
#>   <fct>     <fct> <int> <dbl>     <dbl>     <dbl>
#> 1 Bad       No       22 0.055     0.229     0.186
#> 2 Bad       Yes      73 0.182     0.760     0.264
#> 3 Good      No       28 0.07      0.329     0.237
#> 4 Good      Yes      56 0.14      0.659     0.202
#> 5 Medium    No       68 0.17      0.311     0.576
#> 6 Medium    Yes     148 0.37      0.676     0.534

# Summary the all case : Return a invisible copy of an object.
stat <- summary(all_var)

#> ── Contingency tables ───────────────────────────── Number of table is 3 ──
#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes
#>    Bad     22  73
#>    Good    28  56
#>    Medium  68 148
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes
#>    Bad     34  62
#>    Good    24  61
#>    Medium  84 135
#> 
#> $`Urban vs US`
#>      US
#> Urban  No Yes
#>   No   46  72
#>   Yes  95 182
#> 
#> ── Relative contingency tables ──────────────────── Number of table is 3 ──
#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc         No        Yes
#>    Bad    0.05569620 0.18481013
#>    Good   0.07088608 0.14177215
#>    Medium 0.17215190 0.37468354
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc     No    Yes
#>    Bad    0.0850 0.1550
#>    Good   0.0600 0.1525
#>    Medium 0.2100 0.3375
#> 
#> $`Urban vs US`
#>      US
#> Urban        No       Yes
#>   No  0.1164557 0.1822785
#>   Yes 0.2405063 0.4607595
#> 
#> ── Chi-squared contingency table tests ──────────── Number of table is 3 ──
#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 2.7917270 0.2476191  2
#> 2  ShelveLoc         US 2.7396674 0.2541492  2
#> 3      Urban         US 0.6009241 0.4382256  1

# Summary by returned objects
stat

#> $table
#> $table$`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes
#>    Bad     22  73
#>    Good    28  56
#>    Medium  68 148
#> 
#> $table$`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes
#>    Bad     34  62
#>    Good    24  61
#>    Medium  84 135
#> 
#> $table$`Urban vs US`
#>      US
#> Urban  No Yes
#>   No   46  72
#>   Yes  95 182
#> 
#> 
#> $relative
#> $relative$`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc         No        Yes
#>    Bad    0.05569620 0.18481013
#>    Good   0.07088608 0.14177215
#>    Medium 0.17215190 0.37468354
#> 
#> $relative$`ShelveLoc vs US`
#>          US
#> ShelveLoc     No    Yes
#>    Bad    0.0850 0.1550
#>    Good   0.0600 0.1525
#>    Medium 0.2100 0.3375
#> 
#> $relative$`Urban vs US`
#>      US
#> Urban        No       Yes
#>   No  0.1164557 0.1822785
#>   Yes 0.2405063 0.4607595
#> 
#> 
#> $chisq
#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 2.7917270 0.2476191  2
#> 2  ShelveLoc         US 2.7396674 0.2541492  2
#> 3      Urban         US 0.6009241 0.4382256  1
#> 

# component of table 
stat$table

#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes
#>    Bad     22  73
#>    Good    28  56
#>    Medium  68 148
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes
#>    Bad     34  62
#>    Good    24  61
#>    Medium  84 135
#> 
#> $`Urban vs US`
#>      US
#> Urban  No Yes
#>   No   46  72
#>   Yes  95 182
#> 

# component of chi-square test 
stat$chisq

#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 2.7917270 0.2476191  2
#> 2  ShelveLoc         US 2.7396674 0.2541492  2
#> 3      Urban         US 0.6009241 0.4382256  1

# component of chi-square test 
summary(all_var, "chisq")

#> ── Chi-squared contingency table tests ──────────── Number of table is 3 ──
#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 2.7917270 0.2476191  2
#> 2  ShelveLoc         US 2.7396674 0.2541492  2
#> 3      Urban         US 0.6009241 0.4382256  1

# component of chi-square test (first, third case)
summary(all_var, "chisq", pos = c(1, 3))

#> ── Chi-squared contingency table tests ──────────── Number of table is 2 ──
#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 2.7917270 0.2476191  2
#> 2      Urban         US 0.6009241 0.4382256  1

# component of relative frequency table 
summary(all_var, "relative")

#> ── Relative contingency tables ──────────────────── Number of table is 3 ──
#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc         No        Yes
#>    Bad    0.05569620 0.18481013
#>    Good   0.07088608 0.14177215
#>    Medium 0.17215190 0.37468354
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc     No    Yes
#>    Bad    0.0850 0.1550
#>    Good   0.0600 0.1525
#>    Medium 0.2100 0.3375
#> 
#> $`Urban vs US`
#>      US
#> Urban        No       Yes
#>   No  0.1164557 0.1822785
#>   Yes 0.2405063 0.4607595
#> 

# component of table without missing values 
summary(all_var, "table", na.rm = TRUE)

#> ── Contingency tables ───────────────────────────── Number of table is 3 ──
#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes
#>    Bad     22  73
#>    Good    28  56
#>    Medium  68 148
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes
#>    Bad     34  62
#>    Good    24  61
#>    Medium  84 135
#> 
#> $`Urban vs US`
#>      US
#> Urban  No Yes
#>   No   46  72
#>   Yes  95 182
#> 

# component of table include marginal value 
margin <- summary(all_var, "table", marginal = TRUE)

#> ── Contingency tables ───────────────────────────── Number of table is 3 ──
#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes <Total>
#>   Bad      22  73      95
#>   Good     28  56      84
#>   Medium   68 148     216
#>   <Total> 118 277     395
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes <Total>
#>   Bad      34  62      96
#>   Good     24  61      85
#>   Medium   84 135     219
#>   <Total> 142 258     400
#> 
#> $`Urban vs US`
#>          US
#> Urban      No Yes <Total>
#>   No       46  72     118
#>   Yes      95 182     277
#>   <Total> 141 254     395
#> 

margin

#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes <Total>
#>   Bad      22  73      95
#>   Good     28  56      84
#>   Medium   68 148     216
#>   <Total> 118 277     395
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes <Total>
#>   Bad      34  62      96
#>   Good     24  61      85
#>   Medium   84 135     219
#>   <Total> 142 258     400
#> 
#> $`Urban vs US`
#>          US
#> Urban      No Yes <Total>
#>   No       46  72     118
#>   Yes      95 182     277
#>   <Total> 141 254     395
#> 

# component of chi-square test 
summary(two_var, method = "chisq")

#> ── Chi-squared contingency table tests ──────────── Number of table is 1 ──
#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban  2.791727 0.2476191  2

# verbose is FALSE 
summary(all_var, "chisq", verbose = FALSE)

#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 2.7917270 0.2476191  2
#> 2  ShelveLoc         US 2.7396674 0.2541492  2
#> 3      Urban         US 0.6009241 0.4382256  1

#' # Using pipes & dplyr -------------------------
# If you want to use dplyr, set verbose to FALSE
summary(all_var, "chisq", verbose = FALSE) %>%
  filter(p.value < 0.26)

#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban  2.791727 0.2476191  2
#> 2  ShelveLoc         US  2.739667 0.2541492  2

# Extract component from list by index
summary(all_var, "table", na.rm = TRUE, verbose = FALSE) %>%
  "[["(1)

#>          Urban
#> ShelveLoc  No Yes
#>    Bad     22  73
#>    Good    28  56
#>    Medium  68 148

# Extract component from list by name
  summary(all_var, "table", na.rm = TRUE, verbose = FALSE) %>%
  "[["("ShelveLoc vs Urban")

#>          Urban
#> ShelveLoc  No Yes
#>    Bad     22  73
#>    Good    28  56
#>    Medium  68 148

# plot all pair of variables
# plot(all_var)

# plot a pair of variables
# plot(two_var)

# plot all pair of variables by prompt
# plot(all_var, prompt = TRUE)

# plot a pair of variables
# plot(two_var, las = 1)