Compare numerical variables

The compare_numeric() compute information to examine the relationship between numerical variables.

compare_numeric(.data, ...)

# S3 method for data.frame
compare_numeric(.data, ...)

Arguments

.data
a data.frame or a tbl_df.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

Value

An object of the class as compare based list. The information to examine the relationship between numerical variables is as follows each components. - correlation component : Pearson's correlation coefficient.

var1 : factor. The level of the first variable to compare. 'var1' is the name of the first variable to be compared.

var2 : factor. The level of the second variable to compare. 'var2' is the name of the second variable to be compared.

coef_corr : double. Pearson's correlation coefficient.

- linear component : linear model summaries

var1 : factor. The level of the first variable to compare. 'var1' is the name of the first variable to be compared.

var2 : factor.The level of the second variable to compare. 'var2' is the name of the second variable to be compared.

r.squared : double. The percent of variance explained by the model.

adj.r.squared : double. r.squared adjusted based on the degrees of freedom.

sigma : double. The square root of the estimated residual variance.

statistic : double. F-statistic.

p.value : double. p-value from the F test, describing whether the full regression is significant.

df : integer degrees of freedom.

logLik : double. the log-likelihood of data under the model.

AIC : double. the Akaike Information Criterion.

BIC : double. the Bayesian Information Criterion.

deviance : double. deviance.

df.residual : integer residual degrees of freedom.

Details

It is important to understand the relationship between numerical variables in EDA. compare_numeric() compares relations by pair combination of all numerical variables. and return compare_numeric class that based list object.

Attributes of return object

Attributes of compare_numeric class is as follows.

raw : a data.frame or a tbl_df. Data containing variables to be compared. Save it for visualization with plot.compare_numeric().

variables : character. List of variables selected for comparison.

combination : matrix. It consists of pairs of variables to compare.

See also

Examples

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

library(dplyr)

carseats <- carseats %>%
  select(CompPrice, Sales, Price)

# Compare the all numerical variables
all_var <- compare_numeric(carseats)

# Print compare_numeric class object
all_var

#> $correlation
#> # A tibble: 3 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Sales    0.0641
#> 2 CompPrice Price    0.585 
#> 3 Sales     Price   -0.445 
#> 
#> $linear
#> # A tibble: 3 x 14
#>   var1  var2  r.squared adj.r.squared sigma statistic  p.value    df logLik
#>   <chr> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl>
#> 1 Comp… Sales   0.00411       0.00160 15.3       1.64 2.01e- 1     1 -1658.
#> 2 Comp… Price   0.342         0.340   12.5     207.   4.50e-38     1 -1575.
#> 3 Sales Price   0.198         0.196    2.53     98.2  7.62e-21     1  -938.
#> # … with 5 more variables: AIC <dbl>, BIC <dbl>, deviance <dbl>,
#> #   df.residual <int>, nobs <int>
#> 

# Compare the correlation that case of joint the Price variable
all_var %>%
  "$"(correlation) %>%
  filter(var1 == "Price" | var2 == "Price") %>%
  arrange(desc(abs(coef_corr)))

#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Price     0.585
#> 2 Sales     Price    -0.445

# Compare the correlation that case of abs(coef_corr) > 0.3
all_var %>%
  "$"(correlation) %>%
  filter(abs(coef_corr) > 0.3)

#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Price     0.585
#> 2 Sales     Price    -0.445

# Compare the linear model that case of joint the Price variable  
all_var %>%
  "$"(linear) %>%
  filter(var1 == "Price" | var2 == "Price") %>%
  arrange(desc(r.squared))

#> # A tibble: 2 x 14
#>   var1  var2  r.squared adj.r.squared sigma statistic  p.value    df logLik
#>   <chr> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl>
#> 1 Comp… Price     0.342         0.340 12.5      207.  4.50e-38     1 -1575.
#> 2 Sales Price     0.198         0.196  2.53      98.2 7.62e-21     1  -938.
#> # … with 5 more variables: AIC <dbl>, BIC <dbl>, deviance <dbl>,
#> #   df.residual <int>, nobs <int>

# Compare the two numerical variables
two_var <- compare_numeric(carseats, Price, CompPrice)

# Print compare_numeric class objects
two_var

#> $correlation
#> # A tibble: 1 x 3
#>   var1  var2      coef_corr
#>   <chr> <chr>         <dbl>
#> 1 Price CompPrice     0.585
#> 
#> $linear
#> # A tibble: 1 x 14
#>   var1  var2  r.squared adj.r.squared sigma statistic  p.value    df logLik
#>   <chr> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl>
#> 1 Price Comp…     0.342         0.340  19.2      207. 4.50e-38     1 -1749.
#> # … with 5 more variables: AIC <dbl>, BIC <dbl>, deviance <dbl>,
#> #   df.residual <int>, nobs <int>
#> 

# Summary the all case : Return a invisible copy of an object.
stat <- summary(all_var)

#> ── Correlation check : abs(r) > 0.3 ───────────── Number of pairs is 2/3 ──
#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Price     0.585
#> 2 Sales     Price    -0.445
#> ── R.squared check : R^2 > 0.1 ────────────────── Number of pairs is 2/3 ──
#> # A tibble: 2 x 14
#>   var1  var2  r.squared adj.r.squared sigma statistic  p.value    df logLik
#>   <chr> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl>
#> 1 Comp… Price     0.342         0.340 12.5      207.  4.50e-38     1 -1575.
#> 2 Sales Price     0.198         0.196  2.53      98.2 7.62e-21     1  -938.
#> # … with 5 more variables: AIC <dbl>, BIC <dbl>, deviance <dbl>,
#> #   df.residual <int>, nobs <int>

# Just correlation
summary(all_var, method = "correlation")

#> ── Correlation check : abs(r) > 0.3 ───────────── Number of pairs is 2/3 ──
#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Price     0.585
#> 2 Sales     Price    -0.445

# Just correlation condition by r > 0.2
summary(all_var, method = "correlation", thres_corr = 0.2)

#> ── Correlation check : abs(r) > 0.2 ───────────── Number of pairs is 2/3 ──
#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Price     0.585
#> 2 Sales     Price    -0.445

# linear model summaries condition by R^2 > 0.05
summary(all_var, thres_rs = 0.05)

#> ── Correlation check : abs(r) > 0.3 ───────────── Number of pairs is 2/3 ──
#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Price     0.585
#> 2 Sales     Price    -0.445
#> ── R.squared check : R^2 > 0.05 ───────────────── Number of pairs is 2/3 ──
#> # A tibble: 2 x 14
#>   var1  var2  r.squared adj.r.squared sigma statistic  p.value    df logLik
#>   <chr> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl>
#> 1 Comp… Price     0.342         0.340 12.5      207.  4.50e-38     1 -1575.
#> 2 Sales Price     0.198         0.196  2.53      98.2 7.62e-21     1  -938.
#> # … with 5 more variables: AIC <dbl>, BIC <dbl>, deviance <dbl>,
#> #   df.residual <int>, nobs <int>

# verbose is FALSE 
summary(all_var, verbose = FALSE)

#> $correlation
#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Price     0.585
#> 2 Sales     Price    -0.445
#> 
#> $linear
#> # A tibble: 2 x 14
#>   var1  var2  r.squared adj.r.squared sigma statistic  p.value    df logLik
#>   <chr> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl>
#> 1 Comp… Price     0.342         0.340 12.5      207.  4.50e-38     1 -1575.
#> 2 Sales Price     0.198         0.196  2.53      98.2 7.62e-21     1  -938.
#> # … with 5 more variables: AIC <dbl>, BIC <dbl>, deviance <dbl>,
#> #   df.residual <int>, nobs <int>
#> 

  
# plot all pair of variables
# plot(all_var)

# plot a pair of variables
# plot(two_var)

# plot all pair of variables by prompt
# plot(all_var, prompt = TRUE)

# plot a pair of variables not focuses on typographic elements
# plot(two_var, typographic = FALSE)