Compute the correlation coefficient between two numerical data

The correlate() compute Pearson's the correlation coefficient of the numerical(INTEGER, NUMBER, etc.) column of the DBMS table through tbl_dbi.

# S3 method for tbl_dbi
correlate(
  .data,
  ...,
  in_database = FALSE,
  collect_size = Inf,
  method = c("pearson", "kendall", "spearman")
)

Arguments

.data
a tbl_dbi.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. If the first expression is negative, correlate() will automatically start with all variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

in_database
Specifies whether to perform in-database operations. If TRUE, most operations are performed in the DBMS. if FALSE, table data is taken in R and operated in-memory. Not yet supported in_database = TRUE.

collect_size
a integer. The number of data samples from the DBMS to R. Applies only if in_database = FALSE.

method
a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.

See vignette("EDA") for an introduction to these concepts.

Details

This function is useful when used with the group_by() function of the dplyr package. If you want to compute by level of the categorical data you are interested in, rather than the whole observation, you can use grouped_df as the group_by() function. This function is computed stats::cor() function by use = "pairwise.complete.obs" option.

Correlation coefficient information

The information derived from the numerical data compute is as follows.

var1 : names of numerical variable

var2 : name of the corresponding numeric variable

coef_corr : Pearson's correlation coefficient

See also

Examples

# \donttest{
library(dplyr)

# connect DBMS
con_sqlite <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# copy heartfailure to the DBMS with a table named TB_HEARTFAILURE
copy_to(con_sqlite, heartfailure, name = "TB_HEARTFAILURE", overwrite = TRUE)

# Using pipes ---------------------------------
# Correlation coefficients of all numerical variables
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  correlate()

#> # A tibble: 42 x 3
#>    var1              var2       coef_corr
#>    <fct>             <fct>          <dbl>
#>  1 cpk_enzyme        age          -0.0814
#>  2 ejection_fraction age           0.0602
#>  3 platelets         age          -0.0525
#>  4 creatinine        age           0.159 
#>  5 sodium            age          -0.0459
#>  6 time              age          -0.224 
#>  7 age               cpk_enzyme   -0.0814
#>  8 ejection_fraction cpk_enzyme   -0.0441
#>  9 platelets         cpk_enzyme    0.0245
#> 10 creatinine        cpk_enzyme   -0.0164
#> # … with 32 more rows

# Positive values select variables
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  correlate(platelets, sodium)

#> # A tibble: 12 x 3
#>    var1      var2              coef_corr
#>    <fct>     <fct>                 <dbl>
#>  1 platelets age                 -0.0525
#>  2 sodium    age                 -0.0459
#>  3 platelets cpk_enzyme           0.0245
#>  4 sodium    cpk_enzyme           0.0596
#>  5 platelets ejection_fraction    0.0722
#>  6 sodium    ejection_fraction    0.176 
#>  7 sodium    platelets            0.0621
#>  8 platelets creatinine          -0.0412
#>  9 sodium    creatinine          -0.189 
#> 10 platelets sodium               0.0621
#> 11 platelets time                 0.0105
#> 12 sodium    time                 0.0876

# Negative values to drop variables, and In-memory mode and collect size is 200
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  correlate(-platelets, -sodium, collect_size = 200)

#> # A tibble: 30 x 3
#>    var1              var2              coef_corr
#>    <fct>             <fct>                 <dbl>
#>  1 cpk_enzyme        age                -0.0946 
#>  2 ejection_fraction age                 0.0903 
#>  3 creatinine        age                 0.208  
#>  4 time              age                -0.199  
#>  5 age               cpk_enzyme         -0.0946 
#>  6 ejection_fraction cpk_enzyme         -0.0564 
#>  7 creatinine        cpk_enzyme          0.0122 
#>  8 time              cpk_enzyme          0.00421
#>  9 age               ejection_fraction   0.0903 
#> 10 cpk_enzyme        ejection_fraction  -0.0564 
#> # … with 20 more rows

# Positions values select variables
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  correlate(1)

#> # A tibble: 6 x 3
#>   var1  var2              coef_corr
#>   <fct> <fct>                 <dbl>
#> 1 age   cpk_enzyme          -0.0814
#> 2 age   ejection_fraction    0.0602
#> 3 age   platelets           -0.0525
#> 4 age   creatinine           0.159 
#> 5 age   sodium              -0.0459
#> 6 age   time                -0.224 

# Negative values to drop variables
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  correlate(-1, -2, -3, -5, -6)

#> # A tibble: 24 x 3
#>    var1       var2              coef_corr
#>    <fct>      <fct>                 <dbl>
#>  1 platelets  age                -0.0525 
#>  2 creatinine age                 0.159  
#>  3 sodium     age                -0.0459 
#>  4 time       age                -0.224  
#>  5 platelets  cpk_enzyme          0.0245 
#>  6 creatinine cpk_enzyme         -0.0164 
#>  7 sodium     cpk_enzyme          0.0596 
#>  8 time       cpk_enzyme         -0.00935
#>  9 platelets  ejection_fraction   0.0722 
#> 10 creatinine ejection_fraction  -0.0113 
#> # … with 14 more rows

# ---------------------------------------------
# Correlation coefficient
# that eliminates redundant combination of variables
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  correlate() %>%
  filter(as.integer(var1) > as.integer(var2))

#> # A tibble: 21 x 3
#>    var1              var2       coef_corr
#>    <fct>             <fct>          <dbl>
#>  1 cpk_enzyme        age          -0.0814
#>  2 ejection_fraction age           0.0602
#>  3 platelets         age          -0.0525
#>  4 creatinine        age           0.159 
#>  5 sodium            age          -0.0459
#>  6 time              age          -0.224 
#>  7 ejection_fraction cpk_enzyme   -0.0441
#>  8 platelets         cpk_enzyme    0.0245
#>  9 creatinine        cpk_enzyme   -0.0164
#> 10 sodium            cpk_enzyme    0.0596
#> # … with 11 more rows

con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  correlate(platelets, sodium) %>%
  filter(as.integer(var1) > as.integer(var2))

#> # A tibble: 8 x 3
#>   var1      var2              coef_corr
#>   <fct>     <fct>                 <dbl>
#> 1 platelets age                 -0.0525
#> 2 sodium    age                 -0.0459
#> 3 platelets cpk_enzyme           0.0245
#> 4 sodium    cpk_enzyme           0.0596
#> 5 platelets ejection_fraction    0.0722
#> 6 sodium    ejection_fraction    0.176 
#> 7 sodium    platelets            0.0621
#> 8 sodium    creatinine          -0.189 

# Using pipes & dplyr -------------------------
# Compute the correlation coefficient of creatinine variable by 'hblood_pressure'
# and 'death_event' variables. And extract only those with absolute
# value of correlation coefficient is greater than 0.2
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  group_by(hblood_pressure, death_event) %>%
  correlate(creatinine) %>%
  filter(abs(coef_corr) >= 0.2)

#> # A tibble: 5 x 5
#>   hblood_pressure death_event var1       var2              coef_corr
#>   <chr>           <chr>       <fct>      <fct>                 <dbl>
#> 1 No              No          creatinine sodium               -0.252
#> 2 Yes             No          creatinine cpk_enzyme           -0.217
#> 3 Yes             Yes         creatinine ejection_fraction     0.323
#> 4 Yes             Yes         creatinine sodium               -0.217
#> 5 Yes             Yes         creatinine time                  0.222

# extract only those with 'hblood_pressure' variable level is "Yes",
# and compute the correlation coefficient of 'creatinine' variable
# by 'sex' and 'death_event' variables.
# And the correlation coefficient is negative and smaller than -0.3
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  filter(hblood_pressure == "Yes") %>%
  group_by(sex, death_event) %>%
  correlate(creatinine) %>%
  filter(coef_corr < 0) %>%
  filter(abs(coef_corr) > 0.3)

#> # A tibble: 2 x 5
#>   sex    death_event var1       var2       coef_corr
#>   <chr>  <chr>       <fct>      <fct>          <dbl>
#> 1 Female No          creatinine cpk_enzyme    -0.399
#> 2 Male   Yes         creatinine sodium        -0.437

# Disconnect DBMS   
DBI::dbDisconnect(con_sqlite)
# }