Compute descriptive statistic

The describe() compute descriptive statistic of numeric variable for exploratory data analysis.

describe(.data, ...)

# S3 method for data.frame
describe(.data, ...)

# S3 method for grouped_df
describe(.data, ...)

Arguments

.data
a data.frame or a tbl_df or a grouped_df.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. If the first expression is negative, describe() will automatically start with all variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

See vignette("EDA") for an introduction to these concepts.

Value

An object of the same class as .data.

Details

This function is useful when used with the group_by function of the dplyr package. If you want to calculate the statistic by level of the categorical data you are interested in, rather than the whole statistic, you can use grouped_df as the group_by() function.

Descriptive statistic information

The information derived from the numerical data describe is as follows.

n : number of observations excluding missing values

na : number of missing values

mean : arithmetic average

sd : standard deviation

se_mean : standard error mean. sd/sqrt(n)

IQR : interquartile range (Q3-Q1)

skewness : skewness

kurtosis : kurtosis

p25 : Q1. 25% percentile

p50 : median. 50% percentile

p75 : Q3. 75% percentile

p01, p05, p10, p20, p30 : 1%, 5%, 20%, 30% percentiles

p40, p60, p70, p80 : 40%, 60%, 70%, 80% percentiles

p90, p95, p99, p100 : 90%, 95%, 99%, 100% percentiles

See also

Examples

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# Describe descriptive statistics of numerical variables
describe(carseats)

#> # A tibble: 8 x 26
#>   variable     n    na   mean     sd se_mean    IQR skewness kurtosis   p00
#>   <chr>    <int> <int>  <dbl>  <dbl>   <dbl>  <dbl>    <dbl>    <dbl> <dbl>
#> 1 Sales      400     0   7.50   2.82   0.141   3.93   0.186   -0.0809     0
#> 2 CompPri…   400     0 125.    15.3    0.767  20     -0.0428   0.0417    77
#> 3 Income     380    20  68.5   27.9    1.43   47.5    0.0488  -1.08      21
#> 4 Adverti…   400     0   6.64   6.65   0.333  12      0.640   -0.545      0
#> 5 Populat…   400     0 265.   147.     7.37  260.    -0.0512  -1.20      10
#> 6 Price      400     0 116.    23.7    1.18   31     -0.125    0.452     24
#> 7 Age        400     0  53.3   16.2    0.810  26.2   -0.0772  -1.13      25
#> 8 Educati…   400     0  13.9    2.62   0.131   4      0.0440  -1.30      10
#> # … with 16 more variables: p01 <dbl>, p05 <dbl>, p10 <dbl>, p20 <dbl>,
#> #   p25 <dbl>, p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>, p70 <dbl>,
#> #   p75 <dbl>, p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>, p100 <dbl>

# Select the variable to describe
describe(carseats, Sales, Price)

#> # A tibble: 2 x 26
#>   variable     n    na   mean    sd se_mean   IQR skewness kurtosis   p00    p01
#>   <chr>    <int> <int>  <dbl> <dbl>   <dbl> <dbl>    <dbl>    <dbl> <dbl>  <dbl>
#> 1 Sales      400     0   7.50  2.82   0.141  3.93    0.186  -0.0809     0  0.906
#> 2 Price      400     0 116.   23.7    1.18  31      -0.125   0.452     24 55.0  
#> # … with 15 more variables: p05 <dbl>, p10 <dbl>, p20 <dbl>, p25 <dbl>,
#> #   p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>, p70 <dbl>, p75 <dbl>,
#> #   p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>, p100 <dbl>

describe(carseats, -Sales, -Price)

#> # A tibble: 6 x 26
#>   variable     n    na   mean     sd se_mean   IQR skewness kurtosis   p00   p01
#>   <chr>    <int> <int>  <dbl>  <dbl>   <dbl> <dbl>    <dbl>    <dbl> <dbl> <dbl>
#> 1 CompPri…   400     0 125.    15.3    0.767  20    -0.0428   0.0417    77  89.0
#> 2 Income     380    20  68.5   27.9    1.43   47.5   0.0488  -1.08      21  21.8
#> 3 Adverti…   400     0   6.64   6.65   0.333  12     0.640   -0.545      0   0  
#> 4 Populat…   400     0 265.   147.     7.37  260.   -0.0512  -1.20      10  16.0
#> 5 Age        400     0  53.3   16.2    0.810  26.2  -0.0772  -1.13      25  25  
#> 6 Educati…   400     0  13.9    2.62   0.131   4     0.0440  -1.30      10  10  
#> # … with 15 more variables: p05 <dbl>, p10 <dbl>, p20 <dbl>, p25 <dbl>,
#> #   p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>, p70 <dbl>, p75 <dbl>,
#> #   p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>, p100 <dbl>

describe(carseats, 5)

#> # A tibble: 1 x 26
#>   variable     n    na  mean    sd se_mean   IQR skewness kurtosis   p00   p01
#>   <chr>    <int> <int> <dbl> <dbl>   <dbl> <dbl>    <dbl>    <dbl> <dbl> <dbl>
#> 1 Populat…   400     0  265.  147.    7.37  260.  -0.0512    -1.20    10  16.0
#> # … with 15 more variables: p05 <dbl>, p10 <dbl>, p20 <dbl>, p25 <dbl>,
#> #   p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>, p70 <dbl>, p75 <dbl>,
#> #   p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>, p100 <dbl>

# Using dplyr::grouped_dt
library(dplyr)

gdata <- group_by(carseats, ShelveLoc, US)
describe(gdata, "Income")

#> # A tibble: 6 x 28
#>   variable ShelveLoc US        n    na  mean    sd se_mean   IQR skewness
#>   <chr>    <fct>     <fct> <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>    <dbl>
#> 1 Income   Bad       No       33     1  66.0  24.9    4.34  31     0.160 
#> 2 Income   Bad       Yes      58     4  75.6  26.5    3.49  40.2  -0.214 
#> 3 Income   Good      No       24     0  65.8  30.1    6.14  60.2  -0.0127
#> 4 Income   Good      Yes      58     3  68.2  27.5    3.61  46    -0.0410
#> 5 Income   Medium    No       79     5  65.4  28.9    3.25  44.5   0.170 
#> 6 Income   Medium    Yes     128     7  68.6  28.5    2.52  45.2   0.133 
#> # … with 18 more variables: kurtosis <dbl>, p00 <dbl>, p01 <dbl>, p05 <dbl>,
#> #   p10 <dbl>, p20 <dbl>, p25 <dbl>, p30 <dbl>, p40 <dbl>, p50 <dbl>,
#> #   p60 <dbl>, p70 <dbl>, p75 <dbl>, p80 <dbl>, p90 <dbl>, p95 <dbl>,
#> #   p99 <dbl>, p100 <dbl>

# Using pipes ---------------------------------
# Positive values select variables
carseats %>%
 describe(Sales, CompPrice, Income)

#> # A tibble: 3 x 26
#>   variable     n    na   mean    sd se_mean   IQR skewness kurtosis   p00    p01
#>   <chr>    <int> <int>  <dbl> <dbl>   <dbl> <dbl>    <dbl>    <dbl> <dbl>  <dbl>
#> 1 Sales      400     0   7.50  2.82   0.141  3.93   0.186   -0.0809     0  0.906
#> 2 CompPri…   400     0 125.   15.3    0.767 20     -0.0428   0.0417    77 89.0  
#> 3 Income     380    20  68.5  27.9    1.43  47.5    0.0488  -1.08      21 21.8  
#> # … with 15 more variables: p05 <dbl>, p10 <dbl>, p20 <dbl>, p25 <dbl>,
#> #   p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>, p70 <dbl>, p75 <dbl>,
#> #   p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>, p100 <dbl>

# Negative values to drop variables
carseats %>%
 describe(-Sales, -CompPrice, -Income)

#> # A tibble: 5 x 26
#>   variable     n    na   mean     sd se_mean   IQR skewness kurtosis   p00   p01
#>   <chr>    <int> <int>  <dbl>  <dbl>   <dbl> <dbl>    <dbl>    <dbl> <dbl> <dbl>
#> 1 Adverti…   400     0   6.64   6.65   0.333  12     0.640    -0.545     0   0  
#> 2 Populat…   400     0 265.   147.     7.37  260.   -0.0512   -1.20     10  16.0
#> 3 Price      400     0 116.    23.7    1.18   31    -0.125     0.452    24  55.0
#> 4 Age        400     0  53.3   16.2    0.810  26.2  -0.0772   -1.13     25  25  
#> 5 Educati…   400     0  13.9    2.62   0.131   4     0.0440   -1.30     10  10  
#> # … with 15 more variables: p05 <dbl>, p10 <dbl>, p20 <dbl>, p25 <dbl>,
#> #   p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>, p70 <dbl>, p75 <dbl>,
#> #   p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>, p100 <dbl>

# Using pipes & dplyr -------------------------
# Find the statistic of all numerical variables by 'ShelveLoc' and 'US',
# and extract only those with 'ShelveLoc' variable level is "Good".
carseats %>%
 group_by(ShelveLoc, US) %>%
   describe() %>%
   filter(ShelveLoc == "Good")

#> # A tibble: 16 x 28
#>    variable ShelveLoc US        n    na    mean      sd se_mean    IQR skewness
#>    <chr>    <fct>     <fct> <dbl> <dbl>   <dbl>   <dbl>   <dbl>  <dbl>    <dbl>
#>  1 Sales    Good      No       24     0 9.15e+0   2.93   0.599    3.60  0.363  
#>  2 Sales    Good      Yes      61     0 1.06e+1   2.20   0.281    2.97  0.00369
#>  3 CompPri… Good      No       24     0 1.27e+2  15.8    3.23    21    -0.0748 
#>  4 CompPri… Good      Yes      61     0 1.25e+2  14.7    1.89    23     0.0435 
#>  5 Income   Good      No       24     0 6.58e+1  30.1    6.14    60.2  -0.0127 
#>  6 Income   Good      Yes      58     3 6.82e+1  27.5    3.61    46    -0.0410 
#>  7 Adverti… Good      No       24     0 4.17e-2   0.204  0.0417   0     4.90   
#>  8 Adverti… Good      Yes      61     0 1.02e+1   5.91   0.757    7     0.0647 
#>  9 Populat… Good      No       24     0 2.35e+2 117.    23.9    117.    0.0115 
#> 10 Populat… Good      Yes      61     0 2.80e+2 130.    16.6    180    -0.250  
#> 11 Price    Good      No       24     0 1.18e+2  29.6    6.04    41.2  -0.503  
#> 12 Price    Good      Yes      61     0 1.18e+2  23.4    3.00    27    -0.0332 
#> 13 Age      Good      No       24     0 5.23e+1  17.2    3.52    26    -0.158  
#> 14 Age      Good      Yes      61     0 5.27e+1  14.8    1.90    22    -0.136  
#> 15 Educati… Good      No       24     0 1.32e+1   2.24   0.457    2.5   0.635  
#> 16 Educati… Good      Yes      61     0 1.40e+1   2.68   0.343    5    -0.229  
#> # … with 18 more variables: kurtosis <dbl>, p00 <dbl>, p01 <dbl>, p05 <dbl>,
#> #   p10 <dbl>, p20 <dbl>, p25 <dbl>, p30 <dbl>, p40 <dbl>, p50 <dbl>,
#> #   p60 <dbl>, p70 <dbl>, p75 <dbl>, p80 <dbl>, p90 <dbl>, p95 <dbl>,
#> #   p99 <dbl>, p100 <dbl>

# extract only those with 'Urban' variable level is "Yes",
# and find 'Sales' statistics by 'ShelveLoc' and 'US'
carseats %>%
 filter(Urban == "Yes") %>%
 group_by(ShelveLoc, US) %>%
 describe(Sales)

#> # A tibble: 6 x 28
#>   variable ShelveLoc US        n    na  mean    sd se_mean   IQR skewness
#>   <chr>    <fct>     <fct> <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>    <dbl>
#> 1 Sales    Bad       No       23     0  5.36  1.91   0.398  2.32  -0.275 
#> 2 Sales    Bad       Yes      50     0  5.62  2.58   0.365  3.71   0.139 
#> 3 Sales    Good      No       18     0  9.21  2.97   0.700  3.71   0.0291
#> 4 Sales    Good      Yes      39     0 10.9   2.32   0.372  3.12  -0.0738
#> 5 Sales    Medium    No       53     0  7.04  2.05   0.282  3.09   0.235 
#> 6 Sales    Medium    Yes      96     0  7.55  2.19   0.224  3.39   0.190 
#> # … with 18 more variables: kurtosis <dbl>, p00 <dbl>, p01 <dbl>, p05 <dbl>,
#> #   p10 <dbl>, p20 <dbl>, p25 <dbl>, p30 <dbl>, p40 <dbl>, p50 <dbl>,
#> #   p60 <dbl>, p70 <dbl>, p75 <dbl>, p80 <dbl>, p90 <dbl>, p95 <dbl>,
#> #   p99 <dbl>, p100 <dbl>