Compute descriptive statistic

The describe() compute descriptive statistic of numerical(INTEGER, NUMBER, etc.) column of the DBMS table through tbl_dbi for exploratory data analysis.

# S3 method for tbl_dbi
describe(.data, ..., in_database = FALSE, collect_size = Inf)

Arguments

.data
a tbl_dbi.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. If the first expression is negative, describe() will automatically start with all variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

in_database
Specifies whether to perform in-database operations. If TRUE, most operations are performed in the DBMS. if FALSE, table data is taken in R and operated in-memory. Not yet supported in_database = TRUE.

collect_size
a integer. The number of data samples from the DBMS to R. Applies only if in_database = FALSE.

See vignette("EDA") for an introduction to these concepts.

Value

An object of the same class as .data.

Details

This function is useful when used with the group_by function of the dplyr package. If you want to calculate the statistic by level of the categorical data you are interested in, rather than the whole statistic, you can use grouped_df as the group_by() function.

Descriptive statistic information

The information derived from the numerical data describe is as follows.

n : number of observations excluding missing values

na : number of missing values

mean : arithmetic average

sd : standard deviation

se_mean : standard error mean. sd/sqrt(n)

IQR : interquartile range (Q3-Q1)

skewness : skewness

kurtosis : kurtosis

p25 : Q1. 25% percentile

p50 : median. 50% percentile

p75 : Q3. 75% percentile

p01, p05, p10, p20, p30 : 1%, 5%, 20%, 30% percentiles

p40, p60, p70, p80 : 40%, 60%, 70%, 80% percentiles

p90, p95, p99, p100 : 90%, 95%, 99%, 100% percentiles

See also

Examples

library(dplyr)

# connect DBMS
con_sqlite <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# copy heartfailure to the DBMS with a table named TB_HEARTFAILURE
copy_to(con_sqlite, heartfailure, name = "TB_HEARTFAILURE", overwrite = TRUE)

# Using pipes ---------------------------------
# Positive values select variables
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  describe(platelets, creatinine, sodium)

#> # A tibble: 3 x 26
#>   variable      n    na    mean      sd se_mean    IQR skewness kurtosis     p00
#>   <chr>     <int> <int>   <dbl>   <dbl>   <dbl>  <dbl>    <dbl>    <dbl>   <dbl>
#> 1 platelets   299     0  2.63e5  9.78e4 5.66e+3 9.1e+4     1.46     6.21 25100  
#> 2 creatini…   299     0  1.39e0  1.03e0 5.98e-2 5  e-1     4.46    25.8      0.5
#> 3 sodium      299     0  1.37e2  4.41e0 2.55e-1 6  e+0    -1.05     4.12   113  
#> # … with 16 more variables: p01 <dbl>, p05 <dbl>, p10 <dbl>, p20 <dbl>,
#> #   p25 <dbl>, p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>, p70 <dbl>,
#> #   p75 <dbl>, p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>, p100 <dbl>

# Negative values to drop variables, and In-memory mode and collect size is 200
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  describe(-platelets, -creatinine, -sodium, collect_size = 200)

#> # A tibble: 4 x 26
#>   variable      n    na  mean     sd se_mean   IQR skewness kurtosis   p00   p01
#>   <chr>     <int> <int> <dbl>  <dbl>   <dbl> <dbl>    <dbl>    <dbl> <dbl> <dbl>
#> 1 age         200     0  62.0   12.2   0.859  18      0.442   -0.209    40 41.0 
#> 2 cpk_enzy…   200     0 598.  1099.   77.7   468.     4.39    22.2      23 46.8 
#> 3 ejection…   200     0  38.0   12.6   0.894  15      0.509   -0.232    14 15   
#> 4 time        200     0  84.5   48.5   3.43   72.2    0.192   -0.732     4  6.99
#> # … with 15 more variables: p05 <dbl>, p10 <dbl>, p20 <dbl>, p25 <dbl>,
#> #   p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>, p70 <dbl>, p75 <dbl>,
#> #   p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>, p100 <dbl>

# Using pipes & dplyr -------------------------
# Find the statistic of all numerical variables by 'smoking' and 'death_event',
# and extract only those with 'smoking' variable level is "Yes".
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  group_by(smoking, death_event) %>%
  describe() %>%
  filter(smoking == "Yes")

#> # A tibble: 7 x 27
#>   variable  smoking     n    na    mean     sd se_mean     IQR skewness kurtosis
#>   <chr>     <chr>   <dbl> <dbl>   <dbl>  <dbl>   <dbl>   <dbl>    <dbl>    <dbl>
#> 1 age       Yes        96     0  6.12e1 1.14e1 1.16e+0 1.8 e+1    0.220  -0.462 
#> 2 cpk_enzy… Yes        96     0  5.85e2 1.01e3 1.04e+2 4.61e+2    3.65   14.2   
#> 3 ejection… Yes        96     0  3.69e1 1.11e1 1.13e+0 1   e+1    0.580  -0.0141
#> 4 platelets Yes        96     0  2.67e5 1.08e5 1.10e+4 1.02e+5    2.06    8.79  
#> 5 creatini… Yes        96     0  1.35e0 1.04e0 1.06e-1 4   e-1    5.41   38.1   
#> 6 sodium    Yes        96     0  1.37e2 4.40e0 4.49e-1 4   e+0   -1.63    7.90  
#> 7 time      Yes        96     0  1.28e2 7.77e1 7.93e+0 1.24e+2    0.276  -1.04  
#> # … with 17 more variables: p00 <dbl>, p01 <dbl>, p05 <dbl>, p10 <dbl>,
#> #   p20 <dbl>, p25 <dbl>, p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>,
#> #   p70 <dbl>, p75 <dbl>, p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>,
#> #   p100 <dbl>

# extract only those with 'sex' variable level is "Male",
# and find 'sodium' statistics by 'smoking' and 'death_event'
con_sqlite %>%
  tbl("TB_HEARTFAILURE") %>%
  filter(sex == "Male") %>%
  group_by(smoking, death_event) %>%
  describe(sodium)

#> # A tibble: 2 x 27
#>   variable smoking     n    na  mean    sd se_mean   IQR skewness kurtosis   p00
#>   <chr>    <chr>   <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>    <dbl>    <dbl> <dbl>
#> 1 sodium   No        102     0  136.  3.84   0.380     5    0.207    0.769   126
#> 2 sodium   Yes        92     0  137.  4.45   0.464     4   -1.64     7.87    113
#> # … with 16 more variables: p01 <dbl>, p05 <dbl>, p10 <dbl>, p20 <dbl>,
#> #   p25 <dbl>, p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>, p70 <dbl>,
#> #   p75 <dbl>, p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>, p100 <dbl>

# Disconnect DBMS   
DBI::dbDisconnect(con_sqlite)