Reporting the information of data diagnosis for table of the DBMS with html

The diagnose_web_report() report the information for diagnosing the quality of the DBMS table through tbl_dbi

# S3 method for tbl_dbi
diagnose_web_report(
  .data,
  output_file = NULL,
  output_dir = tempdir(),
  browse = TRUE,
  title = "Data Diagnosis",
  subtitle = deparse(substitute(.data)),
  author = "dlookr",
  title_color = "gray",
  thres_uniq_cat = 0.5,
  thres_uniq_num = 5,
  logo_img = NULL,
  create_date = Sys.time(),
  theme = c("orange", "blue")[1],
  sample_percent = 100,
  in_database = FALSE,
  collect_size = Inf,
  as_factor = TRUE,
  ...
)

Arguments

.data

a tbl_dbi.

output_file

name of generated file. default is NULL.

output_dir

name of directory to generate report file. default is tempdir().

browse

logical. choose whether to output the report results to the browser.

title

character. title of report. default is "Data Diagnosis Report".

subtitle

character. subtitle of report. default is name of data.

author

character. author of report. default is "dlookr".

title_color

character. color of title. default is "gray".

thres_uniq_cat

numeric. threshold to use for "Unique Values - Categorical Variables". default is 0.5.

thres_uniq_num

numeric. threshold to use for "Unique Values - Numerical Variables". default is 5.

logo_img

character. name of logo image on top right.

create_date

Date or POSIXct, character. The date on which the report is generated. The default value is the result of Sys.time().

theme

character. name of theme for report. support "orange" and "blue". default is "orange".

sample_percent

numeric. Sample percent of data for performing Diagnosis. It has a value between (0, 100]. 100 means all data, and 5 means 5% of sample data. This is useful for data with a large number of observations.

in_database

Specifies whether to perform in-database operations. If TRUE, most operations are performed in the DBMS. if FALSE, table data is taken in R and operated in-memory. Not yet supported in_database = TRUE.

collect_size

a integer. The number of data samples from the DBMS to R. Applies only if in_database = FALSE.

as_factor

logical. whether to convert to factor when importing a character type variable from DBMS table into R.

...

arguments to be passed to methods.

Value

No return value. This function only generates a report.

Details

Generate generalized data diagnostic reports automatically. This is useful for diagnosing a data frame with a large number of variables than data with a small number of variables.

Reported information

Reported from the data diagnosis is as follows.

Overview

Data Structures

Data Structures

Data Types

Job Informations

Warnings

Variables

Missing Values

Top Ranks

Numerical Variable Diagnosis

List of Missing Values

Visualization

Unique Values

Categorical Variables

Numerical Variables

Outliers

Samples

Duplicated

Heads

Tails

See also

Examples

# If you have the 'DBI' and 'RSQLite' packages installed, perform the code block:
if (FALSE) {
library(dplyr)

# Generate data for the example
heartfailure2 <- heartfailure
heartfailure2[sample(seq(NROW(heartfailure2)), 20), "platelets"] <- NA
heartfailure2[sample(seq(NROW(heartfailure2)), 5), "smoking"] <- NA

# connect DBMS
con_sqlite <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# copy heartfailure2 to the DBMS with a table named TB_HEARTFAILURE
copy_to(con_sqlite, heartfailure2, name = "TB_HEARTFAILURE", overwrite = TRUE)

# reporting the diagnosis information -------------------------
# create pdf file. file name is Diagnosis_Report.html
con_sqlite %>% 
  tbl("TB_HEARTFAILURE") %>% 
  diagnose_web_report()
  
# create pdf file. file name is Diagn.html, and collect size is 250
con_sqlite %>% 
  tbl("TB_HEARTFAILURE") %>% 
  diagnose_web_report(collect_size = 250, output_file = "Diagn.html")
  
# Disconnect DBMS   
DBI::dbDisconnect(con_sqlite)
}