plot_correlate {dlookr}	R Documentation

Deprecated functions in package ‘dlookr’

Description

These functions are provided for compatibility with older versions of ‘dlookr’ only, and will be defunct at the next release.

Usage

plot_correlate(.data, ...)

Details

The following functions are deprecated and will be made defunct; use the replacement indicated below:

plot_correlate: plot.correlate

[Package dlookr version 0.6.5 Index]