dlookr-package {dlookr}	R Documentation

dlookr: Tools for Data Diagnosis, Exploration, Transformation

Description

dlookr provides data diagnosis, data exploration and transformation of variables during data analysis.

Details

It has three main goals:

When data is acquired, it is possible to judge whether data is erroneous or to select a variable to be corrected or removed through data diagnosis.

Understand the distribution of data in the EDA process. It can also understand the relationship between target variables and predictor variables for the prediction model.

Imputes missing value and outlier to standardization and resolving skewness. And, To convert a continuous variable to a categorical variable, bin the continuous variables.

To learn more about dlookr, start with the vignettes: 'browseVignettes(package = "dlookr")'

Author(s)

Maintainer: Choonghyun Ryu choonghyun.ryu@gmail.com

See Also

Useful links:

https://github.com/choonghyunryu/dlookr/

https://choonghyunryu.github.io/dlookr/

Report bugs at https://github.com/choonghyunryu/dlookr/issues

[Package dlookr version 0.6.5 Index]