Reporting the information of EDA with html

The eda_web_report() report the information of exploratory data analysis for object inheriting from data.frame.

eda_web_report(.data, ...)

# S3 method for data.frame
eda_web_report(
  .data,
  target = NULL,
  output_file = NULL,
  output_dir = tempdir(),
  browse = TRUE,
  title = "EDA",
  subtitle = deparse(substitute(.data)),
  author = "dlookr",
  title_color = "gray",
  logo_img = NULL,
  create_date = Sys.time(),
  theme = c("orange", "blue"),
  sample_percent = 100,
  is_tbl_dbi = FALSE,
  base_family = NULL,
  ...
)

Arguments

.data
a data.frame or a tbl_df.

...
arguments to be passed to methods.

target
character. target variable.

output_file
name of generated file. default is NULL.

output_dir
name of directory to generate report file. default is tempdir().

browse
logical. choose whether to output the report results to the browser.

title
character. title of report. default is "EDA".

subtitle
character. subtitle of report. default is name of data.

author
character. author of report. default is "dlookr".

title_color
character. color of title. default is "gray".

logo_img
character. name of logo image file on top left.

create_date
Date or POSIXct, character. The date on which the report is generated. The default value is the result of Sys.time().

theme
character. name of theme for report. support "orange" and "blue". default is "orange".

sample_percent
numeric. Sample percent of data for performing EDA. It has a value between (0, 100]. 100 means all data, and 5 means 5% of sample data. This is useful for data with a large number of observations.

is_tbl_dbi
logical. whether .data is a tbl_dbi object.

base_family
character. The name of the base font family to use for the visualization. If not specified, the font defined in dlookr is applied. (See details)

Details

Generate generalized EDA report automatically. This feature is useful for EDA of data with many variables, rather than data with fewer variables.

Reported information

Reported from the EDA is as follows.

Overview

Data Structures

Data Types

Job Informations

Univariate Analysis

Descriptive Statistics

Normality Test

Bivariate Analysis

Compare Numerical Variables

Compare Categorical Variables

Multivariate Analysis

Correlation Analysis

Correlation Matrix

Correlation Plot

Target based Analysis

Grouped Numerical Variables

Grouped Categorical Variables

Grouped Correlation

The base_family is selected from "Roboto Condensed", "Liberation Sans Narrow", "NanumSquare", "Noto Sans Korean". If you want to use a different font, use it after loading the Google font with import_google_font().

See also

Examples

if (FALSE) {
# create the dataset
heartfailure2 <- dlookr::heartfailure
heartfailure2[sample(seq(NROW(heartfailure2)), 20), "sodium"] <- NA
heartfailure2[sample(seq(NROW(heartfailure2)), 5), "smoking"] <- NA

# create html file. file name is EDA_Report.html
# eda_web_report(heartfailure2)

# file name is EDA.html. and change logo image
# logo <- file.path(system.file(package = "dlookr"), "report", "R_logo_html.svg")
# eda_web_report(heartfailure2, logo_img = logo, title_color = "black",
#   output_file = "EDA.html")

# file name is ./EDA_heartfailure.html, "blue" theme and not browse
# eda_web_report(heartfailure2, target = "death_event", output_dir = ".", 
#   author = "Choonghyun Ryu", output_file = "EDA_heartfailure.html", 
#   theme = "blue", browse = FALSE)
}