Extract bins from "bins"

The extract() extract binned variable from "bins", "optimal_bins" class object.

extract(x)

# S3 method for bins
extract(x)

Arguments

x
a bins class or optimal_bins class.

Value

factor.

Details

The "bins" and "optimal_bins" class objects use the summary() and plot() functions to diagnose the performance of binned results. This function is used to extract the binned result if you are satisfied with the result.

See also

Examples

# Generate data for the example
library(dplyr)

carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# optimal binning
bin <- binning_by(carseats, "US", "Advertising")

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.
#> 'Yes' changed to 1 (positive) and 'No' changed to 0 (negative).

bin

#> binned type: optimal
#> number of bins: 3
#> x
#> [-1,0]  (0,6] (6,29] 
#>    144     69    187 

# summary optimal_bins class
summary(bin)

#> ── Binning Table ──────────────────────── Several Metrics ──
#>      Bin CntRec CntPos CntNeg RatePos RateNeg    Odds      WoE      IV     JSD     AUC
#> 1 [-1,0]    144     19    125 0.07364 0.88028  0.1520 -2.48101 2.00128 0.20093 0.03241
#> 2  (0,6]     69     54     15 0.20930 0.10563  3.6000  0.68380 0.07089 0.00869 0.01883
#> 3 (6,29]    187    185      2 0.71705 0.01408 92.5000  3.93008 2.76272 0.21861 0.00903
#> 4  Total    400    258    142 1.00000 1.00000  1.8169       NA 4.83489 0.42823 0.06028
#> 
#> ── General Metrics ─────────────────────────────────────────
#> ● Gini index                       :  -0.87944
#> ● IV (Jeffrey)                     :  4.83489
#> ● JS (Jensen-Shannon) Divergence   :  0.42823
#> ● Kolmogorov-Smirnov Statistics    :  0.80664
#> ● HHI (Herfindahl-Hirschman Index) :  0.37791
#> ● HHI (normalized)                 :  0.06687
#> ● Cramer's V                       :  0.81863
#> 
#> ── Significance Tests ──────────────────── Chisquare Test ──
#>    Bin A  Bin B statistics      p_value
#> 1 [-1,0]  (0,6]   87.67064 7.731349e-21
#> 2  (0,6] (6,29]   34.73349 3.780706e-09
#> 

# visualize all information for optimal_bins class
# plot(bin)

# extract binning result
# extract(bin) %>% 
#   head(20)