get_percentile {dlookr}	R Documentation

Finding percentile

Description

Find the percentile of the value specified in numeric vector.

Usage

get_percentile(x, value, from = 0, to = 1, eps = 1e-06)

Arguments

x
numeric. a numeric vector.

value
numeric. a scalar to find percentile value from vector x.

from
numeric. Start interval in logic to find percentile value. default to 0.

to
numeric. End interval in logic to find percentile value. default to 1.

eps
numeric. Threshold value for calculating the approximate value in recursive calling logic to find the percentile value. (epsilon). default to 1e-06.

Value

list. Components of list. is as follows.

percentile : numeric. Percentile position of value. It has a value between [0, 100].

is_outlier : logical. Whether value is an outlier.

Examples

carat <- ggplot2::diamonds$carat

quantile(carat)

get_percentile(carat, value = 0.5)
get_percentile(carat, value = median(carat))
get_percentile(carat, value = 1)
get_percentile(carat, value = 7)

[Package dlookr version 0.6.5 Index]