overview {dlookr}	R Documentation

Describe overview of data

Description

Inquire basic information to understand the data in general.

Usage

overview(.data)

Arguments

.data
a data.frame or a tbl_df.

Details

overview() creates an overview class. The 'overview' class includes general information such as the size of the data, the degree of missing values, and the data types of variables.

Value

An object of overview class. The overview class contains data.frame and two attributes. data.frame has the following 3 variables.: data.frame is as follow.:

division : division of information.

size : indicators of related to data capacity

duplicated : indicators of related to duplicated value

missing : indicators of related to missing value

data_type : indicators of related to data type

metrics : name of metrics.

observations : number of observations (number of rows)

variables : number of variables (number of columns)

values : number of values (number of cells. rows * columns)

memory size : an estimate of the memory that is being used to store an R object.

duplicate observation: number of duplicate cases(observations).

complete observation : number of complete cases(observations). i.e., have no missing values.

missing observation : number of observations that has missing values.

missing variables : number of variables that has missing values.

missing values : number of values(cells) that has missing values.

numerics : number of variables that is data type is numeric.

integers : number of variables that is data type is integer.

factors : number of variables that is data type is factor.

characters : number of variables that is data type is character.

Dates : number of variables that is data type is Date.

POSIXcts : number of variables that is data type is POSIXct.

others : number of variables that is not above.

value : value of metrics.

Attributes of overview class is as follows.:

duplicated : the index of duplicated observations.

na_col : the data type of predictor to replace missing value.

info_class : data.frame. variable name and class name that describe the data type of variables.

data.frame has a two variables.

variable : variable names

class : data type

See Also

summary.overview, plot.overview.

Examples

ov <- overview(jobchange)
ov

summary(ov)

plot(ov)

[Package dlookr version 0.6.5 Index]