Reporting the information of transformation

The transformation_report() report the information of transform numerical variables for object inheriting from data.frame.

transformation_report(
  .data,
  target = NULL,
  output_format = c("pdf", "html"),
  output_file = NULL,
  output_dir = tempdir(),
  font_family = NULL,
  browse = TRUE
)

Arguments

.data
a data.frame or a tbl_df.

target
target variable. If the target variable is not specified, the method of using the target variable information is not performed when the missing value is imputed. and Optimal binning is not performed if the target variable is not a binary class.

output_format
report output type. Choose either "pdf" and "html". "pdf" create pdf file by knitr::knit(). "html" create html file by rmarkdown::render().

output_file
name of generated file. default is NULL.

output_dir
name of directory to generate report file. default is tempdir().

font_family
character. font family name for figure in pdf.

browse
logical. choose whether to output the report results to the browser.

Details

Generate transformation reports automatically. You can choose to output to pdf and html files. This is useful for Binning a data frame with a large number of variables than data with a small number of variables. For pdf output, Korean Gothic font must be installed in Korean operating system.

Reported information

The transformation process will report the following information:

Imputation

Missing Values

* Variable names including missing value

Outliers

* Variable names including outliers

Resolving Skewness

Skewed variables information

* Variable names with an absolute value of skewness greater than or equal to 0.5

Binning

Numerical Variables for Binning

Binning

Numeric variable names

Optimal Binning

Numeric variable names

See vignette("transformation") for an introduction to these concepts.

Examples

# \donttest{
# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# reporting the Binning information -------------------------
# create pdf file. file name is Transformation_Report.pdf & No target variable
transformation_report(carseats)

#> 
#> 
#> processing file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_Report.Rnw

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |...........................                                                     |  33%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.....................................................                           |  67%
#> label: child-section-application (with options) 
#> List of 1
#>  $ child: chr "03_Transformation.Rnw"
#> 

#> 
#> 
#> processing file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/03_Transformation.Rnw

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |...                                                                             |   3%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......                                                                          |   7%
#> label: init (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 

#> Loading required package: sqldf

#> Loading required package: gsubfn

#> Loading required package: proto

#> Loading required package: partykit

#> Loading required package: grid

#> Loading required package: libcoin

#> Loading required package: mvtnorm

#> Loading required package: Formula

#>   |                                                                                        |........                                                                        |  10%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...........                                                                     |  14%
#> label: nalist (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |..............                                                                  |  17%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.................                                                               |  21%
#> label: impute_function (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |...................                                                             |  24%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......................                                                          |  28%
#> label: imputate_na (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |.........................                                                       |  31%
#>   ordinary text without R code
#> 
#>   |                                                                                        |............................                                                    |  34%
#> label: outlist (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |..............................                                                  |  38%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.................................                                               |  41%
#> label: impute_outliers (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |....................................                                            |  45%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.......................................                                         |  48%
#> label: summary_transform (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |.........................................                                       |  52%
#>   ordinary text without R code
#> 
#>   |                                                                                        |............................................                                    |  55%
#> label: skewlist (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |...............................................                                 |  59%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..................................................                              |  62%
#> label: resolve_skewness (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |....................................................                            |  66%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.......................................................                         |  69%
#> label: var_numeric (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |..........................................................                      |  72%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................................                   |  76%
#> label: target (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |...............................................................                 |  79%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..................................................................              |  83%
#> label: binning (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |.....................................................................           |  86%
#>   ordinary text without R code
#> 
#>   |                                                                                        |........................................................................        |  90%
#> label: optimal (with options) 
#> List of 12
#>  $ tidy      : logi TRUE
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ tidy.opts : language list(width.cutoff = 65)
#>  $ size      : chr "small"
#>  $ fig.height: num 4
#>  $ fig.width : num 8
#>  $ fig.align : chr "center"
#>  $ fig.pos   : chr "!ht"
#>  $ results   : chr "asis"
#> 
#>   |                                                                                        |..........................................................................      |  93%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................................................   |  97%
#> label: option_undo (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_Report.tex

#> Warning: LaTeX Warning: `!h' float specifier changed to `!ht'.
#> LaTeX Warning: `!h' float specifier changed to `!ht'.

# create pdf file. file name is Transformation_Report.pdf
transformation_report(carseats, US)

#> 
#> 
#> processing file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_Report.Rnw

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |...........................                                                     |  33%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.....................................................                           |  67%
#> label: child-section-application (with options) 
#> List of 1
#>  $ child: chr "03_Transformation.Rnw"
#> 

#> 
#> 
#> processing file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/03_Transformation.Rnw

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |...                                                                             |   3%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......                                                                          |   7%
#> label: init (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |........                                                                        |  10%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...........                                                                     |  14%
#> label: nalist (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |..............                                                                  |  17%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.................                                                               |  21%
#> label: impute_function (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |...................                                                             |  24%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......................                                                          |  28%
#> label: imputate_na (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |.........................                                                       |  31%
#>   ordinary text without R code
#> 
#>   |                                                                                        |............................                                                    |  34%
#> label: outlist (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |..............................                                                  |  38%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.................................                                               |  41%
#> label: impute_outliers (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |....................................                                            |  45%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.......................................                                         |  48%
#> label: summary_transform (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |.........................................                                       |  52%
#>   ordinary text without R code
#> 
#>   |                                                                                        |............................................                                    |  55%
#> label: skewlist (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |...............................................                                 |  59%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..................................................                              |  62%
#> label: resolve_skewness (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |....................................................                            |  66%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.......................................................                         |  69%
#> label: var_numeric (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |..........................................................                      |  72%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................................                   |  76%
#> label: target (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |...............................................................                 |  79%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..................................................................              |  83%
#> label: binning (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |.....................................................................           |  86%
#>   ordinary text without R code
#> 
#>   |                                                                                        |........................................................................        |  90%
#> label: optimal (with options) 
#> List of 12
#>  $ tidy      : logi TRUE
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ tidy.opts : language list(width.cutoff = 65)
#>  $ size      : chr "small"
#>  $ fig.height: num 4
#>  $ fig.width : num 8
#>  $ fig.align : chr "center"
#>  $ fig.pos   : chr "!ht"
#>  $ results   : chr "asis"
#> 

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#>   |                                                                                        |..........................................................................      |  93%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................................................   |  97%
#> label: option_undo (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_Report.tex

#> Warning: LaTeX Warning: `!h' float specifier changed to `!ht'.
#> LaTeX Warning: `!h' float specifier changed to `!ht'.

# create pdf file. file name is Transformation_carseats.pdf
transformation_report(carseats, "US", output_file = "Transformation_carseats.pdf")

#> 
#> 
#> processing file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_Report.Rnw

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |...........................                                                     |  33%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.....................................................                           |  67%
#> label: child-section-application (with options) 
#> List of 1
#>  $ child: chr "03_Transformation.Rnw"
#> 

#> 
#> 
#> processing file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/03_Transformation.Rnw

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |...                                                                             |   3%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......                                                                          |   7%
#> label: init (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |........                                                                        |  10%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...........                                                                     |  14%
#> label: nalist (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |..............                                                                  |  17%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.................                                                               |  21%
#> label: impute_function (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |...................                                                             |  24%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......................                                                          |  28%
#> label: imputate_na (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |.........................                                                       |  31%
#>   ordinary text without R code
#> 
#>   |                                                                                        |............................                                                    |  34%
#> label: outlist (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |..............................                                                  |  38%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.................................                                               |  41%
#> label: impute_outliers (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |....................................                                            |  45%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.......................................                                         |  48%
#> label: summary_transform (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |.........................................                                       |  52%
#>   ordinary text without R code
#> 
#>   |                                                                                        |............................................                                    |  55%
#> label: skewlist (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |...............................................                                 |  59%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..................................................                              |  62%
#> label: resolve_skewness (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |....................................................                            |  66%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.......................................................                         |  69%
#> label: var_numeric (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |..........................................................                      |  72%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................................                   |  76%
#> label: target (with options) 
#> List of 7
#>  $ tidy     : logi TRUE
#>  $ echo     : logi FALSE
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#>  $ comment  : chr ""
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#> 
#>   |                                                                                        |...............................................................                 |  79%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..................................................................              |  83%
#> label: binning (with options) 
#> List of 6
#>  $ tidy   : logi TRUE
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#>  $ results: chr "asis"
#> 

#>   |                                                                                        |.....................................................................           |  86%
#>   ordinary text without R code
#> 
#>   |                                                                                        |........................................................................        |  90%
#> label: optimal (with options) 
#> List of 12
#>  $ tidy      : logi TRUE
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ tidy.opts : language list(width.cutoff = 65)
#>  $ size      : chr "small"
#>  $ fig.height: num 4
#>  $ fig.width : num 8
#>  $ fig.align : chr "center"
#>  $ fig.pos   : chr "!ht"
#>  $ results   : chr "asis"
#> 

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#>   |                                                                                        |..........................................................................      |  93%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................................................   |  97%
#> label: option_undo (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_carseats.tex

#> Warning: LaTeX Warning: `!h' float specifier changed to `!ht'.
#> LaTeX Warning: `!h' float specifier changed to `!ht'.

# create html file. file name is Transformation_Report.html
transformation_report(carseats, "US", output_format = "html")

#> 
#> 
#> processing file: Transformation_Report.Rmd

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |..                                                                              |   3%
#>    inline R code fragments
#> 
#>   |                                                                                        |.....                                                                           |   6%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                                        |.......                                                                         |   9%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..........                                                                      |  12%
#> label: environment (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |............                                                                    |  15%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...............                                                                 |  18%
#> label: init (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |.................                                                               |  21%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...................                                                             |  24%
#> label: nalist (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |......................                                                          |  27%
#>   ordinary text without R code
#> 
#>   |                                                                                        |........................                                                        |  30%
#> label: Imputate_function (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |...........................                                                     |  33%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................                                                   |  36%
#> label: Imputate_na (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 5
#>  $ fig.width : num 7
#>  $ results   : chr "asis"
#> 

#>   |                                                                                        |................................                                                |  39%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..................................                                              |  42%
#> label: outlist (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |....................................                                            |  45%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.......................................                                         |  48%
#> label: Imputate_outliers (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 5
#>  $ fig.width : num 7
#>  $ results   : chr "asis"
#> 

#>   |                                                                                        |.........................................                                       |  52%
#>   ordinary text without R code
#> 
#>   |                                                                                        |............................................                                    |  55%
#> label: unnamed-chunk-1 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |..............................................                                  |  58%
#>   ordinary text without R code
#> 
#>   |                                                                                        |................................................                                |  61%
#> label: skewlist (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |...................................................                             |  64%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.....................................................                           |  67%
#> label: resolve_skewness (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 4
#>  $ fig.width : num 8
#>  $ results   : chr "asis"
#> 

#>   |                                                                                        |........................................................                        |  70%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..........................................................                      |  73%
#> label: unnamed-chunk-2 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |.............................................................                   |  76%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...............................................................                 |  79%
#> label: target (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |.................................................................               |  82%
#>   ordinary text without R code
#> 
#>   |                                                                                        |....................................................................            |  85%
#> label: binning (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 5
#>  $ fig.width : num 6
#>  $ results   : chr "asis"
#> 

#>   |                                                                                        |......................................................................          |  88%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.........................................................................       |  91%
#> label: optimal (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 6
#>  $ fig.width : num 7
#>  $ results   : chr "asis"
#> 

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#>   |                                                                                        |...........................................................................     |  94%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..............................................................................  |  97%
#> label: unnamed-chunk-3 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: Transformation_Report.knit.md

#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS Transformation_Report.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_Report.html --lua-filter /Library/Frameworks/R.framework/Versions/4.0/Resources/library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Library/Frameworks/R.framework/Versions/4.0/Resources/library/rmarkdown/rmarkdown/lua/latex-div.lua --email-obfuscation none --self-contained --standalone --section-divs --table-of-contents --toc-depth 3 --template /Library/Frameworks/R.framework/Versions/4.0/Resources/library/prettydoc/resources/templates/cayman.html --highlight-style pygments --number-sections --include-in-header /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/rmarkdown-str145b26558b559.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --css /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_Report_files/style.css 

#> 
#> Output created: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_Report.html

# create html file. file name is Transformation_carseats
transformation_report(carseats, US, output_format = "html",
                      output_file = "Transformation_carseats.html")

#> 
#> 
#> processing file: Transformation_Report.Rmd

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |..                                                                              |   3%
#>    inline R code fragments
#> 
#>   |                                                                                        |.....                                                                           |   6%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                                        |.......                                                                         |   9%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..........                                                                      |  12%
#> label: environment (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |............                                                                    |  15%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...............                                                                 |  18%
#> label: init (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |.................                                                               |  21%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...................                                                             |  24%
#> label: nalist (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |......................                                                          |  27%
#>   ordinary text without R code
#> 
#>   |                                                                                        |........................                                                        |  30%
#> label: Imputate_function (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |...........................                                                     |  33%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................                                                   |  36%
#> label: Imputate_na (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 5
#>  $ fig.width : num 7
#>  $ results   : chr "asis"
#> 

#>   |                                                                                        |................................                                                |  39%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..................................                                              |  42%
#> label: outlist (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |....................................                                            |  45%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.......................................                                         |  48%
#> label: Imputate_outliers (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 5
#>  $ fig.width : num 7
#>  $ results   : chr "asis"
#> 

#>   |                                                                                        |.........................................                                       |  52%
#>   ordinary text without R code
#> 
#>   |                                                                                        |............................................                                    |  55%
#> label: unnamed-chunk-1 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |..............................................                                  |  58%
#>   ordinary text without R code
#> 
#>   |                                                                                        |................................................                                |  61%
#> label: skewlist (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |...................................................                             |  64%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.....................................................                           |  67%
#> label: resolve_skewness (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 4
#>  $ fig.width : num 8
#>  $ results   : chr "asis"
#> 

#>   |                                                                                        |........................................................                        |  70%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..........................................................                      |  73%
#> label: unnamed-chunk-2 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |.............................................................                   |  76%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...............................................................                 |  79%
#> label: target (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |.................................................................               |  82%
#>   ordinary text without R code
#> 
#>   |                                                                                        |....................................................................            |  85%
#> label: binning (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 5
#>  $ fig.width : num 6
#>  $ results   : chr "asis"
#> 

#>   |                                                                                        |......................................................................          |  88%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.........................................................................       |  91%
#> label: optimal (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 6
#>  $ fig.width : num 7
#>  $ results   : chr "asis"
#> 

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.

#> Warning: No significant splits

#>   |                                                                                        |...........................................................................     |  94%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..............................................................................  |  97%
#> label: unnamed-chunk-3 (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: Transformation_Report.knit.md

#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS Transformation_Report.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_carseats.html --lua-filter /Library/Frameworks/R.framework/Versions/4.0/Resources/library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Library/Frameworks/R.framework/Versions/4.0/Resources/library/rmarkdown/rmarkdown/lua/latex-div.lua --email-obfuscation none --self-contained --standalone --section-divs --table-of-contents --toc-depth 3 --template /Library/Frameworks/R.framework/Versions/4.0/Resources/library/prettydoc/resources/templates/cayman.html --highlight-style pygments --number-sections --include-in-header /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/rmarkdown-str145b26445134c.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --css /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_carseats_files/style.css 

#> 
#> Output created: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//RtmpVQxIIX/Transformation_carseats.html

# }