transformation_web_report {dlookr}	R Documentation

Reporting the information of transformation with html

Description

The transformation_web_report() report the information of transform numerical variables for object inheriting from data.frame.

Usage

transformation_web_report(
  .data,
  target = NULL,
  output_file = NULL,
  output_dir = tempdir(),
  browse = TRUE,
  title = "Transformation",
  subtitle = deparse(substitute(.data)),
  author = "dlookr",
  title_color = "gray",
  logo_img = NULL,
  create_date = Sys.time(),
  theme = c("orange", "blue"),
  sample_percent = 100,
  base_family = NULL,
  ...
)

Arguments

.data
a data.frame or a tbl_df.

target
character. target variable.

output_file
name of generated file. default is NULL.

output_dir
name of directory to generate report file. default is tempdir().

browse
logical. choose whether to output the report results to the browser.

title
character. title of report. default is "EDA Report".

subtitle
character. subtitle of report. default is name of data.

author
character. author of report. default is "dlookr".

title_color
character. color of title. default is "gray".

logo_img
character. name of logo image file on top left.

create_date
Date or POSIXct, character. The date on which the report is generated. The default value is the result of Sys.time().

theme
character. name of theme for report. support "orange" and "blue". default is "orange".

sample_percent
numeric. Sample percent of data for performing EDA. It has a value between (0, 100]. 100 means all data, and 5 means 5% of sample data. This is useful for data with a large number of observations.

base_family
character. The name of the base font family to use for the visualization. If not specified, the font defined in dlookr is applied. (See details)

...
arguments to be passed to methods.

Details

Generate transformation reports automatically. This is useful for Binning a data frame with a large number of variables than data with a small number of variables.

Value

No return value. This function only generates a report.

Reported information

The transformation process will report the following information:

Overview

Data Structures

Data Types

Job Information

Imputation

Missing Values

Outliers

Resolving Skewness

Binning

Optimal Binning

The base_family is selected from "Roboto Condensed", "Liberation Sans Narrow", "NanumSquare", "Noto Sans Korean". If you want to use a different font, use it after loading the Google font with import_google_font().

Examples

if (FALSE) {
# create html file. file name is Transformation_Report.html
transformation_web_report(heartfailure)

# file name is Transformation.html. and change logo image
logo <- file.path(system.file(package = "dlookr"), "report", "R_logo_html.svg")
transformation_web_report(heartfailure, logo_img = logo, title_color = "black",
  output_file = "Transformation.html")

# file name is ./Transformation.html, "blue" theme and not browse
transformation_web_report(heartfailure, output_dir = ".", target = "death_event", 
  author = "Choonghyun Ryu", output_file = "Transformation.html", 
  theme = "blue", browse = FALSE)
}

[Package dlookr version 0.6.5 Index]