Optimal Binning for Scoring Modeling

The binning_by() finding intervals for numerical variable using optical binning. Optimal binning categorizes a numeric characteristic into bins for ulterior usage in scoring modeling.

binning_by(.data, y, x, p = 0.05, ordered = TRUE, labels = NULL)

Arguments

.data
a data frame.

y
character. name of binary response variable(0, 1). The variable must contain only the integers 0 and 1 as element. However, in the case of factor having two levels, it is performed while type conversion is performed in the calculation process.

x
character. name of continuous characteristic variable. At least 5 different values. and Inf is not allowed.

p
numeric. percentage of records per bin. Default 5% (0.05). This parameter only accepts values greater that 0.00 (0%) and lower than 0.50 (50%).

ordered
logical. whether to build an ordered factor or not.

labels
character. the label names to use for each of the bins.

Value

an object of "optimal_bins" class. Attributes of "optimal_bins" class is as follows.

class : "optimal_bins".

type : binning type, "optimal".

breaks : numeric. the number of intervals into which x is to be cut.

levels : character. levels of binned value.

raw : numeric. raw data, x argument value.

ivtable : data.frame. information value table.

iv : numeric. information value.

target : integer. binary response variable.

Details

This function is useful when used with the mutate/transmute function of the dplyr package. And this function is implemented using smbinning() function of smbinning package.

attributes of "optimal_bins" class

Attributes of the "optimal_bins" class that is as follows.

class : "optimal_bins".

levels : character. factor or ordered factor levels

type : character. binning method

breaks : numeric. breaks for binning

raw : numeric. before the binned the raw data

ivtable : data.frame. information value table

iv : numeric. information value

target : integer. binary response variable

See vignette("transformation") for an introduction to these concepts.

See also

Examples

# Generate data for the example
library(dplyr)

carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# optimal binning using character
bin <- binning_by(carseats, "US", "Advertising")

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.
#> 'Yes' changed to 1 (positive) and 'No' changed to 0 (negative).

# optimal binning using name
bin <- binning_by(carseats, US, Advertising)

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.
#> 'Yes' changed to 1 (positive) and 'No' changed to 0 (negative).

bin

#> binned type: optimal
#> number of bins: 3
#> x
#> [-1,0]  (0,6] (6,29] 
#>    144     69    187 

# performance table
attr(bin, "performance")

#>      Bin CntRec CntPos CntNeg CntCumPos CntCumNeg RatePos RateNeg RateCumPos
#> 1 [-1,0]    144     19    125        19       125 0.07364 0.88028    0.07364
#> 2  (0,6]     69     54     15        73       140 0.20930 0.10563    0.28295
#> 3 (6,29]    187    185      2       258       142 0.71705 0.01408    1.00000
#> 4  Total    400    258    142        NA        NA 1.00000 1.00000         NA
#>   RateCumNeg    Odds   LnOdds      WoE      IV     JSD     AUC
#> 1    0.88028  0.1520 -1.88387 -2.48101 2.00128 0.20093 0.03241
#> 2    0.98592  3.6000  1.28093  0.68380 0.07089 0.00869 0.01883
#> 3    1.00000 92.5000  4.52721  3.93008 2.76272 0.21861 0.00903
#> 4         NA  1.8169  0.59713       NA 4.83489 0.42823 0.06028

# summary optimal_bins class
summary(bin)

#> ── Binning Table ──────────────────────── Several Metrics ──
#>      Bin CntRec CntPos CntNeg RatePos RateNeg    Odds      WoE      IV     JSD
#> 1 [-1,0]    144     19    125 0.07364 0.88028  0.1520 -2.48101 2.00128 0.20093
#> 2  (0,6]     69     54     15 0.20930 0.10563  3.6000  0.68380 0.07089 0.00869
#> 3 (6,29]    187    185      2 0.71705 0.01408 92.5000  3.93008 2.76272 0.21861
#> 4  Total    400    258    142 1.00000 1.00000  1.8169       NA 4.83489 0.42823
#>       AUC
#> 1 0.03241
#> 2 0.01883
#> 3 0.00903
#> 4 0.06028
#> 
#> ── General Metrics ─────────────────────────────────────────
#> ● Gini index                       :  -0.87944
#> ● IV (Jeffrey)                     :  4.83489
#> ● JS (Jensen-Shannon) Divergence   :  0.42823
#> ● Kolmogorov-Smirnov Statistics    :  0.80664
#> ● HHI (Herfindahl-Hirschman Index) :  0.37791
#> ● HHI (normalized)                 :  0.06687
#> ● Cramer's V                       :  0.81863
#> 
#> ── Significance Tests ──────────────────── Chisquare Test ──
#>    Bin A  Bin B statistics      p_value
#> 1 [-1,0]  (0,6]   87.67064 7.731349e-21
#> 2  (0,6] (6,29]   34.73349 3.780706e-09
#> 

# visualize all information for optimal_bins class
# plot(bin)

# visualize WoE information for optimal_bins class
# plot(bin, type = "WoE")

# visualize all information without typographic
# plot(bin, typographic = FALSE)

# extract binned results
# extract(bin) %>% 
#   head(20)