Compute the correlation coefficient between two numerical data

The correlate() compute Pearson's the correlation coefficient of the numerical data.

correlate(.data, ...)

# S3 method for data.frame
correlate(.data, ..., method = c("pearson", "kendall", "spearman"))

Arguments

.data
a data.frame or a tbl_df.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. If the first expression is negative, correlate() will automatically start with all variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

See vignette("EDA") for an introduction to these concepts.

method
a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.

Details

This function is useful when used with the group_by() function of the dplyr package. If you want to compute by level of the categorical data you are interested in, rather than the whole observation, you can use grouped_df as the group_by() function. This function is computed stats::cor() function by use = "pairwise.complete.obs" option.

Correlation coefficient information

The information derived from the numerical data compute is as follows.

var1 : names of numerical variable

var2 : name of the corresponding numeric variable

coef_corr : Pearson's correlation coefficient

See also

Examples

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# Correlation coefficients of all numerical variables
correlate(carseats)

#> # A tibble: 56 x 3
#>    var1        var2      coef_corr
#>    <fct>       <fct>         <dbl>
#>  1 CompPrice   Sales        0.0641
#>  2 Income      Sales        0.151 
#>  3 Advertising Sales        0.270 
#>  4 Population  Sales        0.0505
#>  5 Price       Sales       -0.445 
#>  6 Age         Sales       -0.232 
#>  7 Education   Sales       -0.0520
#>  8 Sales       CompPrice    0.0641
#>  9 Income      CompPrice   -0.0586
#> 10 Advertising CompPrice   -0.0242
#> # … with 46 more rows

# Select the variable to compute
correlate(carseats, Sales, Price)

#> # A tibble: 14 x 3
#>    var1  var2        coef_corr
#>    <fct> <fct>           <dbl>
#>  1 Price Sales         -0.445 
#>  2 Sales CompPrice      0.0641
#>  3 Price CompPrice      0.585 
#>  4 Sales Income         0.151 
#>  5 Price Income        -0.0386
#>  6 Sales Advertising    0.270 
#>  7 Price Advertising    0.0445
#>  8 Sales Population     0.0505
#>  9 Price Population    -0.0121
#> 10 Sales Price         -0.445 
#> 11 Sales Age           -0.232 
#> 12 Price Age           -0.102 
#> 13 Sales Education     -0.0520
#> 14 Price Education      0.0117

correlate(carseats, -Sales, -Price)

#> # A tibble: 42 x 3
#>    var1        var2      coef_corr
#>    <fct>       <fct>         <dbl>
#>  1 CompPrice   Sales        0.0641
#>  2 Income      Sales        0.151 
#>  3 Advertising Sales        0.270 
#>  4 Population  Sales        0.0505
#>  5 Age         Sales       -0.232 
#>  6 Education   Sales       -0.0520
#>  7 Income      CompPrice   -0.0586
#>  8 Advertising CompPrice   -0.0242
#>  9 Population  CompPrice   -0.0947
#> 10 Age         CompPrice   -0.100 
#> # … with 32 more rows

correlate(carseats, "Sales", "Price")

#> # A tibble: 14 x 3
#>    var1  var2        coef_corr
#>    <fct> <fct>           <dbl>
#>  1 Price Sales         -0.445 
#>  2 Sales CompPrice      0.0641
#>  3 Price CompPrice      0.585 
#>  4 Sales Income         0.151 
#>  5 Price Income        -0.0386
#>  6 Sales Advertising    0.270 
#>  7 Price Advertising    0.0445
#>  8 Sales Population     0.0505
#>  9 Price Population    -0.0121
#> 10 Sales Price         -0.445 
#> 11 Sales Age           -0.232 
#> 12 Price Age           -0.102 
#> 13 Sales Education     -0.0520
#> 14 Price Education      0.0117

correlate(carseats, 1)

#> # A tibble: 7 x 3
#>   var1  var2        coef_corr
#>   <fct> <fct>           <dbl>
#> 1 Sales CompPrice      0.0641
#> 2 Sales Income         0.151 
#> 3 Sales Advertising    0.270 
#> 4 Sales Population     0.0505
#> 5 Sales Price         -0.445 
#> 6 Sales Age           -0.232 
#> 7 Sales Education     -0.0520

# Non-parametric correlation coefficient by kendall method
correlate(carseats, Sales, method = "kendall")

#> # A tibble: 7 x 3
#>   var1  var2        coef_corr
#>   <fct> <fct>           <dbl>
#> 1 Sales CompPrice      0.0459
#> 2 Sales Income         0.101 
#> 3 Sales Advertising    0.195 
#> 4 Sales Population     0.0298
#> 5 Sales Price         -0.280 
#> 6 Sales Age           -0.159 
#> 7 Sales Education     -0.0225

# Using dplyr::grouped_dt
library(dplyr)

gdata <- group_by(carseats, ShelveLoc, US)
correlate(gdata, "Sales")

#> # A tibble: 42 x 5
#>    ShelveLoc US    var1  var2        coef_corr
#>    <fct>     <fct> <fct> <fct>           <dbl>
#>  1 Bad       No    Sales CompPrice      0.206 
#>  2 Bad       No    Sales Income        -0.128 
#>  3 Bad       No    Sales Advertising    0.0573
#>  4 Bad       No    Sales Population     0.381 
#>  5 Bad       No    Sales Price         -0.527 
#>  6 Bad       No    Sales Age           -0.210 
#>  7 Bad       No    Sales Education      0.0491
#>  8 Bad       Yes   Sales CompPrice      0.0428
#>  9 Bad       Yes   Sales Income         0.371 
#> 10 Bad       Yes   Sales Advertising    0.267 
#> # … with 32 more rows

correlate(gdata)

#> # A tibble: 336 x 5
#>    ShelveLoc US    var1        var2      coef_corr
#>    <fct>     <fct> <fct>       <fct>         <dbl>
#>  1 Bad       No    CompPrice   Sales        0.206 
#>  2 Bad       No    Income      Sales       -0.128 
#>  3 Bad       No    Advertising Sales        0.0573
#>  4 Bad       No    Population  Sales        0.381 
#>  5 Bad       No    Price       Sales       -0.527 
#>  6 Bad       No    Age         Sales       -0.210 
#>  7 Bad       No    Education   Sales        0.0491
#>  8 Bad       No    Sales       CompPrice    0.206 
#>  9 Bad       No    Income      CompPrice    0.0194
#> 10 Bad       No    Advertising CompPrice   -0.0213
#> # … with 326 more rows

# Using pipes ---------------------------------
# Correlation coefficients of all numerical variables
carseats %>%
 correlate()

#> # A tibble: 56 x 3
#>    var1        var2      coef_corr
#>    <fct>       <fct>         <dbl>
#>  1 CompPrice   Sales        0.0641
#>  2 Income      Sales        0.151 
#>  3 Advertising Sales        0.270 
#>  4 Population  Sales        0.0505
#>  5 Price       Sales       -0.445 
#>  6 Age         Sales       -0.232 
#>  7 Education   Sales       -0.0520
#>  8 Sales       CompPrice    0.0641
#>  9 Income      CompPrice   -0.0586
#> 10 Advertising CompPrice   -0.0242
#> # … with 46 more rows

# Positive values select variables
carseats %>%
 correlate(Sales, Price)

#> # A tibble: 14 x 3
#>    var1  var2        coef_corr
#>    <fct> <fct>           <dbl>
#>  1 Price Sales         -0.445 
#>  2 Sales CompPrice      0.0641
#>  3 Price CompPrice      0.585 
#>  4 Sales Income         0.151 
#>  5 Price Income        -0.0386
#>  6 Sales Advertising    0.270 
#>  7 Price Advertising    0.0445
#>  8 Sales Population     0.0505
#>  9 Price Population    -0.0121
#> 10 Sales Price         -0.445 
#> 11 Sales Age           -0.232 
#> 12 Price Age           -0.102 
#> 13 Sales Education     -0.0520
#> 14 Price Education      0.0117

# Negative values to drop variables
carseats %>%
 correlate(-Sales, -Price)

#> # A tibble: 42 x 3
#>    var1        var2      coef_corr
#>    <fct>       <fct>         <dbl>
#>  1 CompPrice   Sales        0.0641
#>  2 Income      Sales        0.151 
#>  3 Advertising Sales        0.270 
#>  4 Population  Sales        0.0505
#>  5 Age         Sales       -0.232 
#>  6 Education   Sales       -0.0520
#>  7 Income      CompPrice   -0.0586
#>  8 Advertising CompPrice   -0.0242
#>  9 Population  CompPrice   -0.0947
#> 10 Age         CompPrice   -0.100 
#> # … with 32 more rows

# Positions values select variables
carseats %>%
 correlate(1)

#> # A tibble: 7 x 3
#>   var1  var2        coef_corr
#>   <fct> <fct>           <dbl>
#> 1 Sales CompPrice      0.0641
#> 2 Sales Income         0.151 
#> 3 Sales Advertising    0.270 
#> 4 Sales Population     0.0505
#> 5 Sales Price         -0.445 
#> 6 Sales Age           -0.232 
#> 7 Sales Education     -0.0520

# Positions values select variables
carseats %>%
 correlate(-1, -2, -3, -5, -6)

#> # A tibble: 21 x 3
#>    var1        var2        coef_corr
#>    <fct>       <fct>           <dbl>
#>  1 Advertising Sales         0.270  
#>  2 Age         Sales        -0.232  
#>  3 Education   Sales        -0.0520 
#>  4 Advertising CompPrice    -0.0242 
#>  5 Age         CompPrice    -0.100  
#>  6 Education   CompPrice     0.0252 
#>  7 Advertising Income        0.0690 
#>  8 Age         Income       -0.00862
#>  9 Education   Income       -0.0772 
#> 10 Age         Advertising  -0.00456
#> # … with 11 more rows

# Non-parametric correlation coefficient by spearman method
carseats %>%
 correlate(Sales, Price, method = "spearman")

#> # A tibble: 14 x 3
#>    var1  var2        coef_corr
#>    <fct> <fct>           <dbl>
#>  1 Price Sales        -0.408  
#>  2 Sales CompPrice     0.0676 
#>  3 Price CompPrice     0.542  
#>  4 Sales Income        0.153  
#>  5 Price Income       -0.0304 
#>  6 Sales Advertising   0.275  
#>  7 Price Advertising   0.0356 
#>  8 Sales Population    0.0419 
#>  9 Price Population   -0.00602
#> 10 Sales Price        -0.408  
#> 11 Sales Age          -0.236  
#> 12 Price Age          -0.118  
#> 13 Sales Education    -0.0345 
#> 14 Price Education     0.0244 

# ---------------------------------------------
# Correlation coefficient
# that eliminates redundant combination of variables
carseats %>%
 correlate() %>%
 filter(as.integer(var1) > as.integer(var2))

#> # A tibble: 28 x 3
#>    var1        var2      coef_corr
#>    <fct>       <fct>         <dbl>
#>  1 CompPrice   Sales        0.0641
#>  2 Income      Sales        0.151 
#>  3 Advertising Sales        0.270 
#>  4 Population  Sales        0.0505
#>  5 Price       Sales       -0.445 
#>  6 Age         Sales       -0.232 
#>  7 Education   Sales       -0.0520
#>  8 Income      CompPrice   -0.0586
#>  9 Advertising CompPrice   -0.0242
#> 10 Population  CompPrice   -0.0947
#> # … with 18 more rows

carseats %>%
 correlate(Sales, Price) %>%
 filter(as.integer(var1) > as.integer(var2))

#> # A tibble: 5 x 3
#>   var1  var2        coef_corr
#>   <fct> <fct>           <dbl>
#> 1 Price Sales         -0.445 
#> 2 Price CompPrice      0.585 
#> 3 Price Income        -0.0386
#> 4 Price Advertising    0.0445
#> 5 Price Population    -0.0121

# Using pipes & dplyr -------------------------
# Compute the correlation coefficient of Sales variable by 'ShelveLoc'
# and 'US' variables. And extract only those with absolute
# value of correlation coefficient is greater than 0.5
carseats %>%
 group_by(ShelveLoc, US) %>%
 correlate(Sales) %>%
 filter(abs(coef_corr) >= 0.5)

#> # A tibble: 6 x 5
#>   ShelveLoc US    var1  var2  coef_corr
#>   <fct>     <fct> <fct> <fct>     <dbl>
#> 1 Bad       No    Sales Price    -0.527
#> 2 Bad       Yes   Sales Price    -0.583
#> 3 Good      No    Sales Price    -0.811
#> 4 Good      Yes   Sales Price    -0.603
#> 5 Medium    No    Sales Price    -0.610
#> 6 Medium    Yes   Sales Price    -0.538

# extract only those with 'ShelveLoc' variable level is "Good",
# and compute the correlation coefficient of 'Sales' variable
# by 'Urban' and 'US' variables.
# And the correlation coefficient is negative and smaller than 0.5
carseats %>%
 filter(ShelveLoc == "Good") %>%
 group_by(Urban, US) %>%
 correlate(Sales) %>%
 filter(coef_corr < 0) %>%
 filter(abs(coef_corr) > 0.5)

#> # A tibble: 10 x 5
#>    Urban US    var1  var2        coef_corr
#>    <fct> <fct> <fct> <fct>           <dbl>
#>  1 No    No    Sales Population     -0.530
#>  2 No    No    Sales Price          -0.838
#>  3 No    Yes   Sales Price          -0.644
#>  4 Yes   No    Sales Price          -0.833
#>  5 Yes   No    Sales Age            -0.649
#>  6 Yes   Yes   Sales Price          -0.607
#>  7 NA    Yes   Sales Income         -1    
#>  8 NA    Yes   Sales Advertising    -1    
#>  9 NA    Yes   Sales Population     -1    
#> 10 NA    Yes   Sales Price          -1