Compute the correlation coefficient between two numerical data

The correlate() compute Pearson's the correlation coefficient of the numerical data.

correlate(.data, ...)

# S3 method for data.frame
correlate(.data, ..., method = c("pearson", "kendall", "spearman"))

# S3 method for grouped_df
correlate(.data, ..., method = c("pearson", "kendall", "spearman"))

Arguments

.data
a data.frame or a tbl_df.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. If the first expression is negative, correlate() will automatically start with all variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

See vignette("EDA") for an introduction to these concepts.

method
a character string indicating which correlation coefficient (or covariance) is to be computed. One of "pearson" (default), "kendall", or "spearman": can be abbreviated.

Details

This function is useful when used with the group_by() function of the dplyr package. If you want to compute by level of the categorical data you are interested in, rather than the whole observation, you can use grouped_df as the group_by() function. This function is computed stats::cor() function by use = "pairwise.complete.obs" option.

Correlation coefficient information

The information derived from the numerical data compute is as follows.

var1 : names of numerical variable

var2 : name of the corresponding numeric variable

coef_corr : Pearson's correlation coefficient

See also

Examples

# \donttest{
# Correlation coefficients of all numerical variables
correlate(heartfailure)

#> # A tibble: 42 x 3
#>    var1              var2       coef_corr
#>    <fct>             <fct>          <dbl>
#>  1 cpk_enzyme        age          -0.0814
#>  2 ejection_fraction age           0.0602
#>  3 platelets         age          -0.0525
#>  4 creatinine        age           0.159 
#>  5 sodium            age          -0.0459
#>  6 time              age          -0.224 
#>  7 age               cpk_enzyme   -0.0814
#>  8 ejection_fraction cpk_enzyme   -0.0441
#>  9 platelets         cpk_enzyme    0.0245
#> 10 creatinine        cpk_enzyme   -0.0164
#> # … with 32 more rows

# Select the variable to compute
correlate(heartfailure, creatinine, sodium)

#> # A tibble: 12 x 3
#>    var1       var2              coef_corr
#>    <fct>      <fct>                 <dbl>
#>  1 creatinine age                  0.159 
#>  2 sodium     age                 -0.0459
#>  3 creatinine cpk_enzyme          -0.0164
#>  4 sodium     cpk_enzyme           0.0596
#>  5 creatinine ejection_fraction   -0.0113
#>  6 sodium     ejection_fraction    0.176 
#>  7 creatinine platelets           -0.0412
#>  8 sodium     platelets            0.0621
#>  9 sodium     creatinine          -0.189 
#> 10 creatinine sodium              -0.189 
#> 11 creatinine time                -0.149 
#> 12 sodium     time                 0.0876

correlate(heartfailure, -creatinine, -sodium)

#> # A tibble: 30 x 3
#>    var1              var2              coef_corr
#>    <fct>             <fct>                 <dbl>
#>  1 cpk_enzyme        age                -0.0814 
#>  2 ejection_fraction age                 0.0602 
#>  3 platelets         age                -0.0525 
#>  4 time              age                -0.224  
#>  5 age               cpk_enzyme         -0.0814 
#>  6 ejection_fraction cpk_enzyme         -0.0441 
#>  7 platelets         cpk_enzyme          0.0245 
#>  8 time              cpk_enzyme         -0.00935
#>  9 age               ejection_fraction   0.0602 
#> 10 cpk_enzyme        ejection_fraction  -0.0441 
#> # … with 20 more rows

correlate(heartfailure, "creatinine", "sodium")

#> # A tibble: 12 x 3
#>    var1       var2              coef_corr
#>    <fct>      <fct>                 <dbl>
#>  1 creatinine age                  0.159 
#>  2 sodium     age                 -0.0459
#>  3 creatinine cpk_enzyme          -0.0164
#>  4 sodium     cpk_enzyme           0.0596
#>  5 creatinine ejection_fraction   -0.0113
#>  6 sodium     ejection_fraction    0.176 
#>  7 creatinine platelets           -0.0412
#>  8 sodium     platelets            0.0621
#>  9 sodium     creatinine          -0.189 
#> 10 creatinine sodium              -0.189 
#> 11 creatinine time                -0.149 
#> 12 sodium     time                 0.0876

correlate(heartfailure, 1)

#> # A tibble: 6 x 3
#>   var1  var2              coef_corr
#>   <fct> <fct>                 <dbl>
#> 1 age   cpk_enzyme          -0.0814
#> 2 age   ejection_fraction    0.0602
#> 3 age   platelets           -0.0525
#> 4 age   creatinine           0.159 
#> 5 age   sodium              -0.0459
#> 6 age   time                -0.224 

# Non-parametric correlation coefficient by kendall method
correlate(heartfailure, creatinine, method = "kendall")

#> # A tibble: 6 x 3
#>   var1       var2              coef_corr
#>   <fct>      <fct>                 <dbl>
#> 1 creatinine age                  0.191 
#> 2 creatinine cpk_enzyme          -0.0351
#> 3 creatinine ejection_fraction   -0.130 
#> 4 creatinine platelets           -0.0357
#> 5 creatinine sodium              -0.223 
#> 6 creatinine time                -0.110 

# Using dplyr::grouped_dt
library(dplyr)

gdata <- group_by(heartfailure, smoking, death_event)
correlate(gdata, "creatinine")

#> # A tibble: 24 x 5
#>    smoking death_event var1       var2              coef_corr
#>    <fct>   <fct>       <fct>      <fct>                 <dbl>
#>  1 No      No          creatinine age                 0.199  
#>  2 No      No          creatinine cpk_enzyme         -0.0339 
#>  3 No      No          creatinine ejection_fraction  -0.0732 
#>  4 No      No          creatinine platelets           0.00896
#>  5 No      No          creatinine sodium             -0.183  
#>  6 No      No          creatinine time               -0.0971 
#>  7 No      Yes         creatinine age                -0.00114
#>  8 No      Yes         creatinine cpk_enzyme         -0.00566
#>  9 No      Yes         creatinine ejection_fraction   0.298  
#> 10 No      Yes         creatinine platelets          -0.134  
#> # … with 14 more rows

correlate(gdata)

#> # A tibble: 168 x 5
#>    smoking death_event var1              var2       coef_corr
#>    <fct>   <fct>       <fct>             <fct>          <dbl>
#>  1 No      No          cpk_enzyme        age          -0.0393
#>  2 No      No          ejection_fraction age           0.0749
#>  3 No      No          platelets         age          -0.0579
#>  4 No      No          creatinine        age           0.199 
#>  5 No      No          sodium            age          -0.0427
#>  6 No      No          time              age          -0.0193
#>  7 No      No          age               cpk_enzyme   -0.0393
#>  8 No      No          ejection_fraction cpk_enzyme   -0.0819
#>  9 No      No          platelets         cpk_enzyme    0.0610
#> 10 No      No          creatinine        cpk_enzyme   -0.0339
#> # … with 158 more rows

# Using pipes ---------------------------------
# Correlation coefficients of all numerical variables
heartfailure %>%
 correlate()

#> # A tibble: 42 x 3
#>    var1              var2       coef_corr
#>    <fct>             <fct>          <dbl>
#>  1 cpk_enzyme        age          -0.0814
#>  2 ejection_fraction age           0.0602
#>  3 platelets         age          -0.0525
#>  4 creatinine        age           0.159 
#>  5 sodium            age          -0.0459
#>  6 time              age          -0.224 
#>  7 age               cpk_enzyme   -0.0814
#>  8 ejection_fraction cpk_enzyme   -0.0441
#>  9 platelets         cpk_enzyme    0.0245
#> 10 creatinine        cpk_enzyme   -0.0164
#> # … with 32 more rows

# Positive values select variables
heartfailure %>%
 correlate(creatinine, sodium)

#> # A tibble: 12 x 3
#>    var1       var2              coef_corr
#>    <fct>      <fct>                 <dbl>
#>  1 creatinine age                  0.159 
#>  2 sodium     age                 -0.0459
#>  3 creatinine cpk_enzyme          -0.0164
#>  4 sodium     cpk_enzyme           0.0596
#>  5 creatinine ejection_fraction   -0.0113
#>  6 sodium     ejection_fraction    0.176 
#>  7 creatinine platelets           -0.0412
#>  8 sodium     platelets            0.0621
#>  9 sodium     creatinine          -0.189 
#> 10 creatinine sodium              -0.189 
#> 11 creatinine time                -0.149 
#> 12 sodium     time                 0.0876

# Negative values to drop variables
heartfailure %>%
 correlate(-creatinine, -sodium)

#> # A tibble: 30 x 3
#>    var1              var2              coef_corr
#>    <fct>             <fct>                 <dbl>
#>  1 cpk_enzyme        age                -0.0814 
#>  2 ejection_fraction age                 0.0602 
#>  3 platelets         age                -0.0525 
#>  4 time              age                -0.224  
#>  5 age               cpk_enzyme         -0.0814 
#>  6 ejection_fraction cpk_enzyme         -0.0441 
#>  7 platelets         cpk_enzyme          0.0245 
#>  8 time              cpk_enzyme         -0.00935
#>  9 age               ejection_fraction   0.0602 
#> 10 cpk_enzyme        ejection_fraction  -0.0441 
#> # … with 20 more rows

# Positions values select variables
heartfailure %>%
 correlate(1)

#> # A tibble: 6 x 3
#>   var1  var2              coef_corr
#>   <fct> <fct>                 <dbl>
#> 1 age   cpk_enzyme          -0.0814
#> 2 age   ejection_fraction    0.0602
#> 3 age   platelets           -0.0525
#> 4 age   creatinine           0.159 
#> 5 age   sodium              -0.0459
#> 6 age   time                -0.224 

# Positions values select variables
heartfailure %>%
 correlate(-1, -3, -5, -7)

#> # A tibble: 18 x 3
#>    var1       var2              coef_corr
#>    <fct>      <fct>                 <dbl>
#>  1 creatinine age                 0.159  
#>  2 sodium     age                -0.0459 
#>  3 time       age                -0.224  
#>  4 creatinine cpk_enzyme         -0.0164 
#>  5 sodium     cpk_enzyme          0.0596 
#>  6 time       cpk_enzyme         -0.00935
#>  7 creatinine ejection_fraction  -0.0113 
#>  8 sodium     ejection_fraction   0.176  
#>  9 time       ejection_fraction   0.0417 
#> 10 creatinine platelets          -0.0412 
#> 11 sodium     platelets           0.0621 
#> 12 time       platelets           0.0105 
#> 13 sodium     creatinine         -0.189  
#> 14 time       creatinine         -0.149  
#> 15 creatinine sodium             -0.189  
#> 16 time       sodium              0.0876 
#> 17 creatinine time               -0.149  
#> 18 sodium     time                0.0876 

# Non-parametric correlation coefficient by spearman method
heartfailure %>%
 correlate(creatinine, sodium, method = "spearman")

#> # A tibble: 12 x 3
#>    var1       var2              coef_corr
#>    <fct>      <fct>                 <dbl>
#>  1 creatinine age                  0.271 
#>  2 sodium     age                 -0.101 
#>  3 creatinine cpk_enzyme          -0.0499
#>  4 sodium     cpk_enzyme           0.0169
#>  5 creatinine ejection_fraction   -0.178 
#>  6 sodium     ejection_fraction    0.162 
#>  7 creatinine platelets           -0.0510
#>  8 sodium     platelets            0.0495
#>  9 sodium     creatinine          -0.300 
#> 10 creatinine sodium              -0.300 
#> 11 creatinine time                -0.161 
#> 12 sodium     time                 0.0864

# ---------------------------------------------
# Correlation coefficient
# that eliminates redundant combination of variables
heartfailure %>%
 correlate() %>%
 filter(as.integer(var1) > as.integer(var2))

#> # A tibble: 21 x 3
#>    var1              var2       coef_corr
#>    <fct>             <fct>          <dbl>
#>  1 cpk_enzyme        age          -0.0814
#>  2 ejection_fraction age           0.0602
#>  3 platelets         age          -0.0525
#>  4 creatinine        age           0.159 
#>  5 sodium            age          -0.0459
#>  6 time              age          -0.224 
#>  7 ejection_fraction cpk_enzyme   -0.0441
#>  8 platelets         cpk_enzyme    0.0245
#>  9 creatinine        cpk_enzyme   -0.0164
#> 10 sodium            cpk_enzyme    0.0596
#> # … with 11 more rows

heartfailure %>%
  correlate(creatinine, sodium) %>%
  filter(as.integer(var1) > as.integer(var2))

#> # A tibble: 9 x 3
#>   var1       var2              coef_corr
#>   <fct>      <fct>                 <dbl>
#> 1 creatinine age                  0.159 
#> 2 sodium     age                 -0.0459
#> 3 creatinine cpk_enzyme          -0.0164
#> 4 sodium     cpk_enzyme           0.0596
#> 5 creatinine ejection_fraction   -0.0113
#> 6 sodium     ejection_fraction    0.176 
#> 7 creatinine platelets           -0.0412
#> 8 sodium     platelets            0.0621
#> 9 sodium     creatinine          -0.189 

# Using pipes & dplyr -------------------------
# Compute the correlation coefficient of Sales variable by 'smoking'
# and 'death_event' variables. And extract only those with absolute
# value of correlation coefficient is greater than 0.2
heartfailure %>%
 group_by(smoking, death_event) %>%
 correlate(creatinine) %>%
 filter(abs(coef_corr) >= 0.2)

#> # A tibble: 7 x 5
#>   smoking death_event var1       var2              coef_corr
#>   <fct>   <fct>       <fct>      <fct>                 <dbl>
#> 1 No      Yes         creatinine ejection_fraction     0.298
#> 2 Yes     No          creatinine ejection_fraction    -0.201
#> 3 Yes     No          creatinine sodium               -0.290
#> 4 Yes     No          creatinine time                  0.246
#> 5 Yes     Yes         creatinine age                   0.255
#> 6 Yes     Yes         creatinine sodium               -0.286
#> 7 Yes     Yes         creatinine time                 -0.201

# extract only those with 'smoking' variable level is "Yes",
# and compute the correlation coefficient of 'Sales' variable
# by 'hblood_pressure' and 'death_event' variables.
# And the correlation coefficient is negative and smaller than 0.5
heartfailure %>%
  filter(smoking == "Yes") %>%
  group_by(hblood_pressure, death_event) %>%
  correlate(creatinine) %>%
  filter(coef_corr < 0) %>%
  filter(abs(coef_corr) > 0.5)

#> # A tibble: 1 x 5
#>   hblood_pressure death_event var1       var2   coef_corr
#>   <fct>           <fct>       <fct>      <fct>      <dbl>
#> 1 Yes             Yes         creatinine sodium    -0.561

# }