Compute descriptive statistic

The describe() compute descriptive statistic of numerical(INTEGER, NUMBER, etc.) column of the DBMS table through tbl_dbi for exploratory data analysis.

# S3 method for tbl_dbi
describe(.data, ..., in_database = FALSE, collect_size = Inf)

Arguments

.data
a tbl_dbi.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. If the first expression is negative, describe() will automatically start with all variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

in_database
Specifies whether to perform in-database operations. If TRUE, most operations are performed in the DBMS. if FALSE, table data is taken in R and operated in-memory. Not yet supported in_database = TRUE.

collect_size
a integer. The number of data samples from the DBMS to R. Applies only if in_database = FALSE.

See vignette("EDA") for an introduction to these concepts.

Value

An object of the same class as .data.

Details

This function is useful when used with the group_by function of the dplyr package. If you want to calculate the statistic by level of the categorical data you are interested in, rather than the whole statistic, you can use grouped_df as the group_by() function.

Descriptive statistic information

The information derived from the numerical data describe is as follows.

n : number of observations excluding missing values

na : number of missing values

mean : arithmetic average

sd : standard deviation

se_mean : standard error mean. sd/sqrt(n)

IQR : interquartile range (Q3-Q1)

skewness : skewness

kurtosis : kurtosis

p25 : Q1. 25% percentile

p50 : median. 50% percentile

p75 : Q3. 75% percentile

p01, p05, p10, p20, p30 : 1%, 5%, 20%, 30% percentiles

p40, p60, p70, p80 : 40%, 60%, 70%, 80% percentiles

p90, p95, p99, p100 : 90%, 95%, 99%, 100% percentiles

See also

Examples

library(dplyr)

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# connect DBMS
con_sqlite <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# copy carseats to the DBMS with a table named TB_CARSEATS
copy_to(con_sqlite, carseats, name = "TB_CARSEATS", overwrite = TRUE)

# Using pipes ---------------------------------
# Positive values select variables
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  describe(Sales, CompPrice, Income)

#> # A tibble: 3 x 26
#>   variable     n    na   mean    sd se_mean   IQR skewness kurtosis   p00    p01
#>   <chr>    <int> <int>  <dbl> <dbl>   <dbl> <dbl>    <dbl>    <dbl> <dbl>  <dbl>
#> 1 Sales      400     0   7.50  2.82   0.141  3.93   0.186   -0.0809     0  0.906
#> 2 CompPri…   400     0 125.   15.3    0.767 20     -0.0428   0.0417    77 89.0  
#> 3 Income     380    20  68.8  27.9    1.43  47      0.0354  -1.09      21 21.8  
#> # … with 15 more variables: p05 <dbl>, p10 <dbl>, p20 <dbl>, p25 <dbl>,
#> #   p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>, p70 <dbl>, p75 <dbl>,
#> #   p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>, p100 <dbl>

# Negative values to drop variables, and In-memory mode and collect size is 200
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  describe(-Sales, -CompPrice, -Income, collect_size = 200)

#> # A tibble: 5 x 26
#>   variable     n    na   mean     sd se_mean   IQR skewness kurtosis   p00   p01
#>   <chr>    <int> <int>  <dbl>  <dbl>   <dbl> <dbl>    <dbl>    <dbl> <dbl> <dbl>
#> 1 Adverti…   200     0   5.88   6.07   0.429  11     0.648    -0.667     0   0  
#> 2 Populat…   200     0 255.   149.    10.6   251.    0.0241   -1.22     10  16  
#> 3 Price      200     0 114.    23.7    1.68   31.2  -0.107     1.08     24  54.9
#> 4 Age        200     0  54.6   15.9    1.13   24    -0.245    -1.01     25  25  
#> 5 Educati…   200     0  14.1    2.67   0.189   5    -0.0592   -1.34     10  10  
#> # … with 15 more variables: p05 <dbl>, p10 <dbl>, p20 <dbl>, p25 <dbl>,
#> #   p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>, p70 <dbl>, p75 <dbl>,
#> #   p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>, p100 <dbl>

# Using pipes & dplyr -------------------------
# Find the statistic of all numerical variables by 'ShelveLoc' and 'US',
# and extract only those with 'ShelveLoc' variable level is "Good".
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  group_by(ShelveLoc, US) %>%
  describe() %>%
  filter(ShelveLoc == "Good")

#> # A tibble: 8 x 27
#>   variable ShelveLoc     n    na   mean     sd se_mean    IQR skewness kurtosis
#>   <chr>    <chr>     <dbl> <dbl>  <dbl>  <dbl>   <dbl>  <dbl>    <dbl>    <dbl>
#> 1 Sales    Good         85     0  10.2    2.50   0.271   3.63  -0.0759   -0.261
#> 2 CompPri… Good         85     0 126.    15.0    1.62   22      0.0141   -0.490
#> 3 Income   Good         81     4  68.1   28.7    3.18   52     -0.0443   -1.29 
#> 4 Adverti… Good         85     0   7.35   6.80   0.738  12      0.424    -0.901
#> 5 Populat… Good         85     0 267.   127.    13.8   177     -0.142    -0.714
#> 6 Price    Good         85     0 118.    25.1    2.73   29     -0.227    -0.125
#> 7 Age      Good         85     0  52.6   15.4    1.67   23     -0.147    -1.10 
#> 8 Educati… Good         85     0  13.8    2.58   0.280   5     -0.0102   -1.24 
#> # … with 17 more variables: p00 <dbl>, p01 <dbl>, p05 <dbl>, p10 <dbl>,
#> #   p20 <dbl>, p25 <dbl>, p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>,
#> #   p70 <dbl>, p75 <dbl>, p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>,
#> #   p100 <dbl>

# extract only those with 'Urban' variable level is "Yes",
# and find 'Sales' statistics by 'ShelveLoc' and 'US'
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  filter(Urban == "Yes") %>%
  group_by(ShelveLoc, US) %>%
  describe(Sales)

#> # A tibble: 3 x 27
#>   variable ShelveLoc     n    na  mean    sd se_mean   IQR skewness kurtosis
#>   <chr>    <chr>     <dbl> <dbl> <dbl> <dbl>   <dbl> <dbl>    <dbl>    <dbl>
#> 1 Sales    Bad          73     0  5.55  2.37   0.277  3.4     0.116   0.0293
#> 2 Sales    Good         56     0 10.3   2.65   0.355  3.96   -0.210  -0.216 
#> 3 Sales    Medium      150     0  7.32  2.16   0.176  3.05    0.241  -0.280 
#> # … with 17 more variables: p00 <dbl>, p01 <dbl>, p05 <dbl>, p10 <dbl>,
#> #   p20 <dbl>, p25 <dbl>, p30 <dbl>, p40 <dbl>, p50 <dbl>, p60 <dbl>,
#> #   p70 <dbl>, p75 <dbl>, p80 <dbl>, p90 <dbl>, p95 <dbl>, p99 <dbl>,
#> #   p100 <dbl>