Diagnose data quality of variables in the DBMS

The diagnose() produces information for diagnosing the quality of the column of the DBMS table through tbl_dbi.

# S3 method for tbl_dbi
diagnose(.data, ..., in_database = TRUE, collect_size = Inf)

Arguments

.data
a tbl_dbi.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. If the first expression is negative, diagnose() will automatically start with all variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

in_database
a logical. Specifies whether to perform in-database operations. If TRUE, most operations are performed in the DBMS. if FALSE, table data is taken in R and operated in-memory.

collect_size
a integer. The number of data samples from the DBMS to R. Applies only if in_database = FALSE.

Value

An object of tbl_df.

Details

The scope of data quality diagnosis is information on missing values and unique value information. Data quality diagnosis can determine variables that require missing value processing. Also, the unique value information can determine the variable to be removed from the data analysis.

Diagnostic information

The information derived from the data diagnosis is as follows.:

variables : column names

types : data type of the variable or to select a variable to be corrected or removed through data diagnosis.

integer, numeric, factor, ordered, character, etc.

missing_count : number of missing values

missing_percent : percentage of missing values

unique_count : number of unique values

unique_rate : ratio of unique values. unique_count / number of observation

See vignette("diagonosis") for an introduction to these concepts.

See also

Examples

library(dplyr)

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# connect DBMS
con_sqlite <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# copy carseats to the DBMS with a table named TB_CARSEATS
copy_to(con_sqlite, carseats, name = "TB_CARSEATS", overwrite = TRUE)

# Using pipes ---------------------------------
# Diagnosis of all columns
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose()

#> # A tibble: 11 x 6
#>    variables   types     missing_count missing_percent unique_count unique_rate
#>    <chr>       <chr>             <dbl>           <dbl>        <int>       <dbl>
#>  1 Sales       double                0            0             336      0.84  
#>  2 CompPrice   double                0            0              73      0.182 
#>  3 Income      double               20            5              99      0.248 
#>  4 Advertising double                0            0              28      0.07  
#>  5 Population  double                0            0             275      0.688 
#>  6 Price       double                0            0             101      0.252 
#>  7 ShelveLoc   character             0            0               3      0.0075
#>  8 Age         double                0            0              56      0.14  
#>  9 Education   double                0            0               9      0.0225
#> 10 Urban       character             5            1.25            3      0.0075
#> 11 US          character             0            0               2      0.005 

# Positive values select columns
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose(Sales, Income, Age)

#> # A tibble: 3 x 6
#>   variables types  missing_count missing_percent unique_count unique_rate
#>   <chr>     <chr>          <dbl>           <dbl>        <int>       <dbl>
#> 1 Sales     double             0               0          336       0.84 
#> 2 Income    double            20               5           99       0.248
#> 3 Age       double             0               0           56       0.14 

# Negative values to drop columns
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose(-Sales, -Income, -Age)

#> # A tibble: 8 x 6
#>   variables   types     missing_count missing_percent unique_count unique_rate
#>   <chr>       <chr>             <dbl>           <dbl>        <int>       <dbl>
#> 1 CompPrice   double                0            0              73      0.182 
#> 2 Advertising double                0            0              28      0.07  
#> 3 Population  double                0            0             275      0.688 
#> 4 Price       double                0            0             101      0.252 
#> 5 ShelveLoc   character             0            0               3      0.0075
#> 6 Education   double                0            0               9      0.0225
#> 7 Urban       character             5            1.25            3      0.0075
#> 8 US          character             0            0               2      0.005 

# Positions values select columns, and In-memory mode
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose(1, 3, 8, in_database = FALSE)

#> # A tibble: 3 x 6
#>   variables types   missing_count missing_percent unique_count unique_rate
#>   <chr>     <chr>           <int>           <dbl>        <int>       <dbl>
#> 1 Sales     numeric             0               0          336       0.84 
#> 2 Income    numeric            20               5           99       0.248
#> 3 Age       numeric             0               0           56       0.14 

# Positions values select columns, and In-memory mode and collect size is 200
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose(-8, -9, -10, in_database = FALSE, collect_size = 200)

#> # A tibble: 8 x 6
#>   variables   types     missing_count missing_percent unique_count unique_rate
#>   <chr>       <chr>             <int>           <dbl>        <int>       <dbl>
#> 1 Sales       numeric               0             0            182       0.91 
#> 2 CompPrice   numeric               0             0             65       0.325
#> 3 Income      numeric              11             5.5           82       0.41 
#> 4 Advertising numeric               0             0             23       0.115
#> 5 Population  numeric               0             0            162       0.81 
#> 6 Price       numeric               0             0             82       0.41 
#> 7 ShelveLoc   character             0             0              3       0.015
#> 8 US          character             0             0              2       0.01 

# Using pipes & dplyr -------------------------
# Diagnosis of missing variables
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose() %>%
  filter(missing_count > 0)

#> # A tibble: 2 x 6
#>   variables types     missing_count missing_percent unique_count unique_rate
#>   <chr>     <chr>             <dbl>           <dbl>        <int>       <dbl>
#> 1 Income    double               20            5              99      0.248 
#> 2 Urban     character             5            1.25            3      0.0075