Diagnose data quality of categorical variables in the DBMS

The diagnose_category() produces information for diagnosing the quality of the character(CHAR, VARCHAR, VARCHAR2, etc.) column of the DBMS table through tbl_dbi.

# S3 method for tbl_dbi
diagnose_category(.data, ..., top = 10, in_database = TRUE, collect_size = Inf)

Arguments

.data
a tbl_dbi.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. If the first expression is negative, diagnose_category() will automatically start with all variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

top
an integer. Specifies the upper top rank to extract. Default is 10.

in_database
Specifies whether to perform in-database operations. If TRUE, most operations are performed in the DBMS. if FALSE, table data is taken in R and operated in-memory.

collect_size
a integer. The number of data samples from the DBMS to R. Applies only if in_database = FALSE.

Value

an object of tbl_df.

Details

The scope of the diagnosis is the occupancy status of the levels in categorical data. If a certain level of occupancy is close to 100 then the removal of this variable in the forecast model will have to be considered. Also, if the occupancy of all levels is close to 0 variable is likely to be an identifier.

Categorical diagnostic information

The information derived from the categorical data diagnosis is as follows.

variables : variable names

levels: level names

N : number of observation

freq : number of observation at the levels

ratio : percentage of observation at the levels

rank : rank of occupancy ratio of levels

See vignette("diagonosis") for an introduction to these concepts.

See also

Examples

library(dplyr)

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# connect DBMS
con_sqlite <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# copy carseats to the DBMS with a table named TB_CARSEATS
copy_to(con_sqlite, carseats, name = "TB_CARSEATS", overwrite = TRUE)

# Using pipes ---------------------------------
# Diagnosis of all categorical variables
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose_category()

#> # A tibble: 8 x 6
#>   variables levels     N  freq ratio  rank
#>   <chr>     <chr>  <int> <int> <dbl> <int>
#> 1 ShelveLoc Medium   400   219 54.8      1
#> 2 ShelveLoc Bad      400    96 24        2
#> 3 ShelveLoc Good     400    85 21.2      3
#> 4 Urban     Yes      400   278 69.5      1
#> 5 Urban     No       400   117 29.2      2
#> 6 Urban     NA       400     5  1.25     3
#> 7 US        Yes      400   258 64.5      1
#> 8 US        No       400   142 35.5      2

# Positive values select variables
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose_category(Urban, US)

#> # A tibble: 5 x 6
#>   variables levels     N  freq ratio  rank
#>   <chr>     <chr>  <int> <int> <dbl> <int>
#> 1 Urban     Yes      400   278 69.5      1
#> 2 Urban     No       400   117 29.2      2
#> 3 Urban     NA       400     5  1.25     3
#> 4 US        Yes      400   258 64.5      1
#> 5 US        No       400   142 35.5      2

# Negative values to drop variables, and In-memory mode
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose_category(-Urban, -US, in_database = FALSE)

#> # A tibble: 3 x 6
#>   variables levels     N  freq ratio  rank
#> * <chr>     <chr>  <int> <int> <dbl> <int>
#> 1 ShelveLoc Medium   400   219  54.8     1
#> 2 ShelveLoc Bad      400    96  24       2
#> 3 ShelveLoc Good     400    85  21.2     3

# Positions values select variables, and In-memory mode and collect size is 200
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose_category(7, in_database = FALSE, collect_size = 200)

#> # A tibble: 3 x 6
#>   variables levels     N  freq ratio  rank
#> * <chr>     <chr>  <int> <int> <dbl> <int>
#> 1 ShelveLoc Medium   200   113  56.5     1
#> 2 ShelveLoc Bad      200    47  23.5     2
#> 3 ShelveLoc Good     200    40  20       3

# Positions values select variables
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose_category(-7)

#> # A tibble: 5 x 6
#>   variables levels     N  freq ratio  rank
#>   <chr>     <chr>  <int> <int> <dbl> <int>
#> 1 Urban     Yes      400   278 69.5      1
#> 2 Urban     No       400   117 29.2      2
#> 3 Urban     NA       400     5  1.25     3
#> 4 US        Yes      400   258 64.5      1
#> 5 US        No       400   142 35.5      2

# Top rank levels with top argument
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose_category(top = 2)

#> # A tibble: 6 x 6
#>   variables levels     N  freq ratio  rank
#>   <chr>     <chr>  <int> <int> <dbl> <int>
#> 1 ShelveLoc Bad      400    96 24        2
#> 2 ShelveLoc Good     400    85 21.2      3
#> 3 Urban     No       400   117 29.2      2
#> 4 Urban     NA       400     5  1.25     3
#> 5 US        Yes      400   258 64.5      1
#> 6 US        No       400   142 35.5      2

# Using pipes & dplyr -------------------------
# Extraction of level that is more than 60% of categorical data
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose_category()  %>%
  filter(ratio >= 60)

#> # A tibble: 2 x 6
#>   variables levels     N  freq ratio  rank
#>   <chr>     <chr>  <int> <int> <dbl> <int>
#> 1 Urban     Yes      400   278  69.5     1
#> 2 US        Yes      400   258  64.5     1