Diagnose data quality of numerical variables

The diagnose_numeric() produces information for diagnosing the quality of the numerical data.

diagnose_numeric(.data, ...)

# S3 method for data.frame
diagnose_numeric(.data, ...)

Arguments

.data
a data.frame or a tbl_df.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. If the first expression is negative, diagnose_numeric() will automatically start with all variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

Value

an object of tbl_df.

Details

The scope of the diagnosis is the calculate a statistic that can be used to understand the distribution of numerical data. min, Q1, mean, median, Q3, max can be used to estimate the distribution of data. If the number of zero or minus is large, it is necessary to suspect the error of the data. If the number of outliers is large, a strategy of eliminating or replacing outliers is needed.

Numerical diagnostic information

The information derived from the numerical data diagnosis is as follows.

variables : variable names

min : minimum

Q1 : 25 percentile

mean : arithmetic average

median : median. 50 percentile

Q3 : 75 percentile

max : maximum

zero : count of zero values

minus : count of minus values

outlier : count of outliers

See vignette("diagonosis") for an introduction to these concepts.

See also

Examples

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# Diagnosis of numerical variables
diagnose_numeric(carseats)

#> # A tibble: 8 x 10
#>   variables     min     Q1   mean median     Q3   max  zero minus outlier
#>   <chr>       <dbl>  <dbl>  <dbl>  <dbl>  <dbl> <dbl> <int> <int>   <int>
#> 1 Sales           0   5.39   7.50   7.49   9.32  16.3     1     0       2
#> 2 CompPrice      77 115    125.   125    135    175       0     0       2
#> 3 Income         21  42     69.0   69     91.2  120       0     0       0
#> 4 Advertising     0   0      6.64   5     12     29     144     0       0
#> 5 Population     10 139    265.   272    398.   509       0     0       0
#> 6 Price          24 100    116.   117    131    191       0     0       5
#> 7 Age            25  39.8   53.3   54.5   66     80       0     0       0
#> 8 Education      10  12     13.9   14     16     18       0     0       0

# Select the variable to diagnose
diagnose_numeric(carseats, Sales, Income)

#> # A tibble: 2 x 10
#>   variables   min    Q1  mean median    Q3   max  zero minus outlier
#>   <chr>     <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <int> <int>   <int>
#> 1 Sales         0  5.39  7.50   7.49  9.32  16.3     1     0       2
#> 2 Income       21 42    69.0   69    91.2  120       0     0       0

diagnose_numeric(carseats, -Sales, -Income)

#> # A tibble: 6 x 10
#>   variables     min    Q1   mean median    Q3   max  zero minus outlier
#>   <chr>       <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <int> <int>   <int>
#> 1 CompPrice      77 115   125.    125    135    175     0     0       2
#> 2 Advertising     0   0     6.64    5     12     29   144     0       0
#> 3 Population     10 139   265.    272    398.   509     0     0       0
#> 4 Price          24 100   116.    117    131    191     0     0       5
#> 5 Age            25  39.8  53.3    54.5   66     80     0     0       0
#> 6 Education      10  12    13.9    14     16     18     0     0       0

diagnose_numeric(carseats, "Sales", "Income")

#> # A tibble: 2 x 10
#>   variables   min    Q1  mean median    Q3   max  zero minus outlier
#>   <chr>     <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <int> <int>   <int>
#> 1 Sales         0  5.39  7.50   7.49  9.32  16.3     1     0       2
#> 2 Income       21 42    69.0   69    91.2  120       0     0       0

diagnose_numeric(carseats, 5)

#> # A tibble: 1 x 10
#>   variables    min    Q1  mean median    Q3   max  zero minus outlier
#> * <chr>      <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <int> <int>   <int>
#> 1 Population    10   139  265.    272  398.   509     0     0       0

# Using pipes ---------------------------------
library(dplyr)

# Diagnosis of all numerical variables
carseats %>%
  diagnose_numeric()

#> # A tibble: 8 x 10
#>   variables     min     Q1   mean median     Q3   max  zero minus outlier
#>   <chr>       <dbl>  <dbl>  <dbl>  <dbl>  <dbl> <dbl> <int> <int>   <int>
#> 1 Sales           0   5.39   7.50   7.49   9.32  16.3     1     0       2
#> 2 CompPrice      77 115    125.   125    135    175       0     0       2
#> 3 Income         21  42     69.0   69     91.2  120       0     0       0
#> 4 Advertising     0   0      6.64   5     12     29     144     0       0
#> 5 Population     10 139    265.   272    398.   509       0     0       0
#> 6 Price          24 100    116.   117    131    191       0     0       5
#> 7 Age            25  39.8   53.3   54.5   66     80       0     0       0
#> 8 Education      10  12     13.9   14     16     18       0     0       0

# Positive values select variables
carseats %>%
  diagnose_numeric(Sales, Income)

#> # A tibble: 2 x 10
#>   variables   min    Q1  mean median    Q3   max  zero minus outlier
#>   <chr>     <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <int> <int>   <int>
#> 1 Sales         0  5.39  7.50   7.49  9.32  16.3     1     0       2
#> 2 Income       21 42    69.0   69    91.2  120       0     0       0

# Negative values to drop variables
carseats %>%
  diagnose_numeric(-Sales, -Income)

#> # A tibble: 6 x 10
#>   variables     min    Q1   mean median    Q3   max  zero minus outlier
#>   <chr>       <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <int> <int>   <int>
#> 1 CompPrice      77 115   125.    125    135    175     0     0       2
#> 2 Advertising     0   0     6.64    5     12     29   144     0       0
#> 3 Population     10 139   265.    272    398.   509     0     0       0
#> 4 Price          24 100   116.    117    131    191     0     0       5
#> 5 Age            25  39.8  53.3    54.5   66     80     0     0       0
#> 6 Education      10  12    13.9    14     16     18     0     0       0

# Positions values select variables
carseats %>%
  diagnose_numeric(5)

#> # A tibble: 1 x 10
#>   variables    min    Q1  mean median    Q3   max  zero minus outlier
#> * <chr>      <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <int> <int>   <int>
#> 1 Population    10   139  265.    272  398.   509     0     0       0

# Positions values select variables
carseats %>%
  diagnose_numeric(-1, -5)

#> # A tibble: 6 x 10
#>   variables     min    Q1   mean median    Q3   max  zero minus outlier
#>   <chr>       <dbl> <dbl>  <dbl>  <dbl> <dbl> <dbl> <int> <int>   <int>
#> 1 CompPrice      77 115   125.    125   135     175     0     0       2
#> 2 Income         21  42    69.0    69    91.2   120     0     0       0
#> 3 Advertising     0   0     6.64    5    12      29   144     0       0
#> 4 Price          24 100   116.    117   131     191     0     0       5
#> 5 Age            25  39.8  53.3    54.5  66      80     0     0       0
#> 6 Education      10  12    13.9    14    16      18     0     0       0

# Using pipes & dplyr -------------------------
# Information records of zero variable more than 0
carseats %>%
  diagnose_numeric()  %>%
  filter(zero > 0)

#> # A tibble: 2 x 10
#>   variables     min    Q1  mean median    Q3   max  zero minus outlier
#>   <chr>       <dbl> <dbl> <dbl>  <dbl> <dbl> <dbl> <int> <int>   <int>
#> 1 Sales           0  5.39  7.50   7.49  9.32  16.3     1     0       2
#> 2 Advertising     0  0     6.64   5    12     29     144     0       0