Diagnose outlier of numerical variables

The diagnose_outlier() produces outlier information for diagnosing the quality of the numerical data.

diagnose_outlier(.data, ...)

# S3 method for data.frame
diagnose_outlier(.data, ...)

Arguments

.data
a data.frame or a tbl_df.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. If the first expression is negative, diagnose_outlier() will automatically start with all variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

Value

an object of tbl_df.

Details

The scope of the diagnosis is the provide a outlier information. If the number of outliers is small and the difference between the averages including outliers and the averages not including them is large, it is necessary to eliminate or replace the outliers.

Outlier Diagnostic information

The information derived from the numerical data diagnosis is as follows.

variables : variable names

outliers_cnt : number of outliers

outliers_ratio : percent of outliers

outliers_mean : arithmetic average of outliers

with_mean : arithmetic average of with outliers

without_mean : arithmetic average of without outliers

See vignette("diagonosis") for an introduction to these concepts.

See also

Examples

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# Diagnosis of numerical variables
diagnose_outlier(carseats)

#>     variables outliers_cnt outliers_ratio outliers_mean  with_mean without_mean
#> 1       Sales            2           0.50         15.95   7.496325     7.453844
#> 2   CompPrice            2           0.50        126.00 124.975000   124.969849
#> 3      Income            0           0.00           NaN  68.507895    68.507895
#> 4 Advertising            0           0.00           NaN   6.635000     6.635000
#> 5  Population            0           0.00           NaN 264.840000   264.840000
#> 6       Price            5           1.25        100.40 115.795000   115.989873
#> 7         Age            0           0.00           NaN  53.322500    53.322500
#> 8   Education            0           0.00           NaN  13.900000    13.900000

# Select the variable to diagnose
diagnose_outlier(carseats, Sales, Income)

#>        variables outliers_cnt outliers_ratio outliers_mean with_mean
#> Sales      Sales            2            0.5         15.95  7.496325
#> Income    Income            0            0.0           NaN 68.507895
#>        without_mean
#> Sales      7.453844
#> Income    68.507895

diagnose_outlier(carseats, -Sales, -Income)

#>               variables outliers_cnt outliers_ratio outliers_mean with_mean
#> CompPrice     CompPrice            2           0.50         126.0  124.9750
#> Advertising Advertising            0           0.00           NaN    6.6350
#> Population   Population            0           0.00           NaN  264.8400
#> Price             Price            5           1.25         100.4  115.7950
#> Age                 Age            0           0.00           NaN   53.3225
#> Education     Education            0           0.00           NaN   13.9000
#>             without_mean
#> CompPrice       124.9698
#> Advertising       6.6350
#> Population      264.8400
#> Price           115.9899
#> Age              53.3225
#> Education        13.9000

diagnose_outlier(carseats, "Sales", "Income")

#>        variables outliers_cnt outliers_ratio outliers_mean with_mean
#> Sales      Sales            2            0.5         15.95  7.496325
#> Income    Income            0            0.0           NaN 68.507895
#>        without_mean
#> Sales      7.453844
#> Income    68.507895

diagnose_outlier(carseats, 5)

#> # A tibble: 1 x 6
#>   variables  outliers_cnt outliers_ratio outliers_mean with_mean without_mean
#> * <chr>             <int>          <dbl>         <dbl>     <dbl>        <dbl>
#> 1 Population            0              0           NaN      265.         265.

# Using pipes ---------------------------------
library(dplyr)

# Diagnosis of all numerical variables
carseats %>%
  diagnose_outlier()

#>     variables outliers_cnt outliers_ratio outliers_mean  with_mean without_mean
#> 1       Sales            2           0.50         15.95   7.496325     7.453844
#> 2   CompPrice            2           0.50        126.00 124.975000   124.969849
#> 3      Income            0           0.00           NaN  68.507895    68.507895
#> 4 Advertising            0           0.00           NaN   6.635000     6.635000
#> 5  Population            0           0.00           NaN 264.840000   264.840000
#> 6       Price            5           1.25        100.40 115.795000   115.989873
#> 7         Age            0           0.00           NaN  53.322500    53.322500
#> 8   Education            0           0.00           NaN  13.900000    13.900000

# Positive values select variables
carseats %>%
  diagnose_outlier(Sales, Income)

#>        variables outliers_cnt outliers_ratio outliers_mean with_mean
#> Sales      Sales            2            0.5         15.95  7.496325
#> Income    Income            0            0.0           NaN 68.507895
#>        without_mean
#> Sales      7.453844
#> Income    68.507895

# Negative values to drop variables
carseats %>%
  diagnose_outlier(-Sales, -Income)

#>               variables outliers_cnt outliers_ratio outliers_mean with_mean
#> CompPrice     CompPrice            2           0.50         126.0  124.9750
#> Advertising Advertising            0           0.00           NaN    6.6350
#> Population   Population            0           0.00           NaN  264.8400
#> Price             Price            5           1.25         100.4  115.7950
#> Age                 Age            0           0.00           NaN   53.3225
#> Education     Education            0           0.00           NaN   13.9000
#>             without_mean
#> CompPrice       124.9698
#> Advertising       6.6350
#> Population      264.8400
#> Price           115.9899
#> Age              53.3225
#> Education        13.9000

# Positions values select variables
carseats %>%
  diagnose_outlier(5)

#> # A tibble: 1 x 6
#>   variables  outliers_cnt outliers_ratio outliers_mean with_mean without_mean
#> * <chr>             <int>          <dbl>         <dbl>     <dbl>        <dbl>
#> 1 Population            0              0           NaN      265.         265.

# Positions values select variables
carseats %>%
  diagnose_outlier(-1, -5)

#>               variables outliers_cnt outliers_ratio outliers_mean with_mean
#> CompPrice     CompPrice            2           0.50         126.0 124.97500
#> Income           Income            0           0.00           NaN  68.50789
#> Advertising Advertising            0           0.00           NaN   6.63500
#> Price             Price            5           1.25         100.4 115.79500
#> Age                 Age            0           0.00           NaN  53.32250
#> Education     Education            0           0.00           NaN  13.90000
#>             without_mean
#> CompPrice      124.96985
#> Income          68.50789
#> Advertising      6.63500
#> Price          115.98987
#> Age             53.32250
#> Education       13.90000

# Using pipes & dplyr -------------------------
# outlier_ratio is more than 1%
carseats %>%
  diagnose_outlier()  %>%
  filter(outliers_ratio > 1)

#>   variables outliers_cnt outliers_ratio outliers_mean with_mean without_mean
#> 1     Price            5           1.25         100.4   115.795     115.9899