Reporting the information of data diagnosis for table of the DBMS

The diagnose_report() report the information for diagnosing the quality of the DBMS table through tbl_dbi

# S3 method for tbl_dbi
diagnose_report(
  .data,
  output_format = c("pdf", "html"),
  output_file = NULL,
  output_dir = tempdir(),
  font_family = NULL,
  in_database = FALSE,
  collect_size = Inf,
  ...
)

Arguments

.data
a tbl_dbi.

output_format
report output type. Choose either "pdf" and "html". "pdf" create pdf file by knitr::knit(). "html" create html file by rmarkdown::render().

output_file
name of generated file. default is NULL.

output_dir
name of directory to generate report file. default is tempdir().

font_family
character. font family name for figure in pdf.

in_database
Specifies whether to perform in-database operations. If TRUE, most operations are performed in the DBMS. if FALSE, table data is taken in R and operated in-memory. Not yet supported in_database = TRUE.

collect_size
a integer. The number of data samples from the DBMS to R. Applies only if in_database = FALSE.

...
arguments to be passed to methods.

Details

Generate generalized data diagnostic reports automatically. You can choose to output to pdf and html files. This is useful for diagnosing a data frame with a large number of variables than data with a small number of variables. For pdf output, Korean Gothic font must be installed in Korean operating system.

Reported information

Reported from the data diagnosis is as follows.

Diagnose Data

Overview of Diagnosis

List of all variables quality

Diagnosis of missing data

Diagnosis of unique data(Text and Category)

Diagnosis of unique data(Numerical)

Detailed data diagnosis

Diagnosis of categorical variables

Diagnosis of numerical variables

List of numerical diagnosis (zero)

List of numerical diagnosis (minus)

Diagnose Outliers

Overview of Diagnosis

Diagnosis of numerical variable outliers

Detailed outliers diagnosis

See vignette("diagonosis") for an introduction to these concepts.

See also

Examples

# \donttest{
library(dplyr)

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# connect DBMS
con_sqlite <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# copy carseats to the DBMS with a table named TB_CARSEATS
copy_to(con_sqlite, carseats, name = "TB_CARSEATS", overwrite = TRUE)

# reporting the diagnosis information -------------------------
# create pdf file. file name is DataDiagnosis_Report.pdf
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose_report()

#> 
#> 
#> processing file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/DataDiagnosis_Report.Rnw

#>   |                                                                              |                                                                      |   0%  |                                                                              |.......................                                               |  33%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................                       |  67%
#> label: child-section-application (with options) 
#> List of 1
#>  $ child: chr "01_Diagnose.Rnw"
#> 

#> 
#> 
#> processing file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/01_Diagnose.Rnw

#>   |                                                                              |                                                                      |   0%  |                                                                              |...                                                                   |   4%
#>   ordinary text without R code
#> 
#>   |                                                                              |......                                                                |   8%
#> label: environment (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 

#> 
#> Attaching package: 'magrittr'

#> The following object is masked from 'package:dlookr':
#> 
#>     extract

#> 
#> Attaching package: 'kableExtra'

#> The following object is masked from 'package:dplyr':
#> 
#>     group_rows

#>   |                                                                              |........                                                              |  12%
#>   ordinary text without R code
#> 
#>   |                                                                              |...........                                                           |  16%
#> label: overview (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                              |..............                                                        |  20%
#>   ordinary text without R code
#> 
#>   |                                                                              |.................                                                     |  24%
#> label: missing_data (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                              |....................                                                  |  28%
#>   ordinary text without R code
#> 
#>   |                                                                              |......................                                                |  32%
#> label: unique_date_category (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                              |.........................                                             |  36%
#>   ordinary text without R code
#> 
#>   |                                                                              |............................                                          |  40%
#> label: unique_data_numeric (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                              |...............................                                       |  44%
#>   ordinary text without R code
#> 
#>   |                                                                              |..................................                                    |  48%
#> label: diagnose_catagory (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                              |....................................                                  |  52%
#>   ordinary text without R code
#> 
#>   |                                                                              |.......................................                               |  56%
#> label: diagnose_numeric (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                              |..........................................                            |  60%
#>   ordinary text without R code
#> 
#>   |                                                                              |.............................................                         |  64%
#> label: numeric_zero (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                              |................................................                      |  68%
#>   ordinary text without R code
#> 
#>   |                                                                              |..................................................                    |  72%
#> label: numeric_minus (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                              |.....................................................                 |  76%
#>   ordinary text without R code
#> 
#>   |                                                                              |........................................................              |  80%
#> label: outliers (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                              |...........................................................           |  84%
#>   ordinary text without R code
#> 
#>   |                                                                              |..............................................................        |  88%
#> label: detail_outliers (with options) 
#> List of 5
#>  $ tidy     : logi TRUE
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#>  $ echo     : logi FALSE
#>  $ results  : chr "asis"
#> 

#>   |                                                                              |................................................................      |  92%
#>   ordinary text without R code
#> 
#>   |                                                                              |...................................................................   |  96%
#> label: option_undo (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#>   |                                                                              |......................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/DataDiagnosis_Report.tex

# create pdf file. file name is Diagn.pdf, and collect size is 350
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose_report(collect_size = 350, output_file = "Diagn.pdf")

#> 
#> 
#> processing file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/DataDiagnosis_Report.Rnw

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |...........................                                                     |  33%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.....................................................                           |  67%
#> label: child-section-application (with options) 
#> List of 1
#>  $ child: chr "01_Diagnose.Rnw"
#> 

#> 
#> 
#> processing file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/01_Diagnose.Rnw

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |...                                                                             |   4%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......                                                                          |   8%
#> label: environment (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |..........                                                                      |  12%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............                                                                   |  16%
#> label: overview (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                                        |................                                                                |  20%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...................                                                             |  24%
#> label: missing_data (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                                        |......................                                                          |  28%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..........................                                                      |  32%
#> label: unique_date_category (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                                        |.............................                                                   |  36%
#>   ordinary text without R code
#> 
#>   |                                                                                        |................................                                                |  40%
#> label: unique_data_numeric (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                                        |...................................                                             |  44%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......................................                                          |  48%
#> label: diagnose_catagory (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                                        |..........................................                                      |  52%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................                                   |  56%
#> label: diagnose_numeric (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                                        |................................................                                |  60%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...................................................                             |  64%
#> label: numeric_zero (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                                        |......................................................                          |  68%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..........................................................                      |  72%
#> label: numeric_minus (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                                        |.............................................................                   |  76%
#>   ordinary text without R code
#> 
#>   |                                                                                        |................................................................                |  80%
#> label: outliers (with options) 
#> List of 2
#>  $ echo   : logi FALSE
#>  $ results: chr "asis"
#> 
#>   |                                                                                        |...................................................................             |  84%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......................................................................          |  88%
#> label: detail_outliers (with options) 
#> List of 5
#>  $ tidy     : logi TRUE
#>  $ tidy.opts: language list(width.cutoff = 65)
#>  $ size     : chr "small"
#>  $ echo     : logi FALSE
#>  $ results  : chr "asis"
#> 

#>   |                                                                                        |..........................................................................      |  92%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................................................   |  96%
#> label: option_undo (with options) 
#> List of 1
#>  $ echo: logi FALSE
#> 
#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 
#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/Diagn.tex

# create html file. file name is Diagnosis_Report.html
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose_report(output_format = "html")

#> 
#> 
#> processing file: Diagnosis_Report.Rmd

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |...                                                                             |   4%
#>    inline R code fragments
#> 
#>   |                                                                                        |......                                                                          |   8%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                                        |..........                                                                      |  12%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............                                                                   |  16%
#> label: load_packages (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |................                                                                |  20%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...................                                                             |  24%
#> label: overview (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |......................                                                          |  28%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..........................                                                      |  32%
#> label: missing_data (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |.............................                                                   |  36%
#>   ordinary text without R code
#> 
#>   |                                                                                        |................................                                                |  40%
#> label: unique_date_category (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |...................................                                             |  44%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......................................                                          |  48%
#> label: unique_data_numeric (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |..........................................                                      |  52%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................                                   |  56%
#> label: diagnose_catagory (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |................................................                                |  60%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...................................................                             |  64%
#> label: diagnose_numeric (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |......................................................                          |  68%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..........................................................                      |  72%
#> label: numeric_zero (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |.............................................................                   |  76%
#>   ordinary text without R code
#> 
#>   |                                                                                        |................................................................                |  80%
#> label: numeric_minus (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |...................................................................             |  84%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......................................................................          |  88%
#> label: outliers (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |..........................................................................      |  92%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................................................   |  96%
#> label: detail_outliers (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 4
#>  $ fig.width : num 6
#>  $ results   : chr "asis"
#> 

#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: Diagnosis_Report.knit.md

#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS Diagnosis_Report.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/Diagnosis_Report.html --lua-filter /Library/Frameworks/R.framework/Versions/4.0/Resources/library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Library/Frameworks/R.framework/Versions/4.0/Resources/library/rmarkdown/rmarkdown/lua/latex-div.lua --self-contained --standalone --section-divs --table-of-contents --toc-depth 3 --template /Library/Frameworks/R.framework/Versions/4.0/Resources/library/prettydoc/resources/templates/cayman.html --highlight-style pygments --number-sections --include-in-header /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/rmarkdown-str134381e4f8416.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --css /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/Diagnosis_Report_files/style.css 

#> 
#> Output created: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/Diagnosis_Report.html

# create html file. file name is Diagn.html
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  diagnose_report(output_format = "html", output_file = "Diagn.html")

#> 
#> 
#> processing file: Diagnosis_Report.Rmd

#>   |                                                                                        |                                                                                |   0%  |                                                                                        |...                                                                             |   4%
#>    inline R code fragments
#> 
#>   |                                                                                        |......                                                                          |   8%
#> label: setup (with options) 
#> List of 1
#>  $ include: logi FALSE
#> 
#>   |                                                                                        |..........                                                                      |  12%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............                                                                   |  16%
#> label: load_packages (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |................                                                                |  20%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...................                                                             |  24%
#> label: overview (with options) 
#> List of 3
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#> 
#>   |                                                                                        |......................                                                          |  28%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..........................                                                      |  32%
#> label: missing_data (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |.............................                                                   |  36%
#>   ordinary text without R code
#> 
#>   |                                                                                        |................................                                                |  40%
#> label: unique_date_category (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |...................................                                             |  44%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......................................                                          |  48%
#> label: unique_data_numeric (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |..........................................                                      |  52%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................                                   |  56%
#> label: diagnose_catagory (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |................................................                                |  60%
#>   ordinary text without R code
#> 
#>   |                                                                                        |...................................................                             |  64%
#> label: diagnose_numeric (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |......................................................                          |  68%
#>   ordinary text without R code
#> 
#>   |                                                                                        |..........................................................                      |  72%
#> label: numeric_zero (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |.............................................................                   |  76%
#>   ordinary text without R code
#> 
#>   |                                                                                        |................................................................                |  80%
#> label: numeric_minus (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |...................................................................             |  84%
#>   ordinary text without R code
#> 
#>   |                                                                                        |......................................................................          |  88%
#> label: outliers (with options) 
#> List of 4
#>  $ echo   : logi FALSE
#>  $ warning: logi FALSE
#>  $ message: logi FALSE
#>  $ comment: chr ""
#> 
#>   |                                                                                        |..........................................................................      |  92%
#>   ordinary text without R code
#> 
#>   |                                                                                        |.............................................................................   |  96%
#> label: detail_outliers (with options) 
#> List of 7
#>  $ echo      : logi FALSE
#>  $ warning   : logi FALSE
#>  $ message   : logi FALSE
#>  $ comment   : chr ""
#>  $ fig.height: num 4
#>  $ fig.width : num 6
#>  $ results   : chr "asis"
#> 

#>   |                                                                                        |................................................................................| 100%
#>   ordinary text without R code
#> 
#> 

#> output file: Diagnosis_Report.knit.md

#> /Applications/RStudio.app/Contents/MacOS/pandoc/pandoc +RTS -K512m -RTS Diagnosis_Report.utf8.md --to html4 --from markdown+autolink_bare_uris+tex_math_single_backslash --output /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/Diagn.html --lua-filter /Library/Frameworks/R.framework/Versions/4.0/Resources/library/rmarkdown/rmarkdown/lua/pagebreak.lua --lua-filter /Library/Frameworks/R.framework/Versions/4.0/Resources/library/rmarkdown/rmarkdown/lua/latex-div.lua --self-contained --standalone --section-divs --table-of-contents --toc-depth 3 --template /Library/Frameworks/R.framework/Versions/4.0/Resources/library/prettydoc/resources/templates/cayman.html --highlight-style pygments --number-sections --include-in-header /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/rmarkdown-str134387f146980.html --mathjax --variable 'mathjax-url:https://mathjax.rstudio.com/latest/MathJax.js?config=TeX-AMS-MML_HTMLorMML' --css /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/Diagn_files/style.css 

#> 
#> Output created: /var/folders/2g/sk4df3sn3bdbxh2rb9kdt_5c0000gn/T//Rtmp4A1Sys/Diagn.html

# }