Reporting the information of EDA

The eda_report() report the information of exploratory data analysis for object inheriting from data.frame.

eda_report(.data, ...)

# S3 method for data.frame
eda_report(
  .data,
  target = NULL,
  output_format = c("pdf", "html"),
  output_file = NULL,
  output_dir = tempdir(),
  font_family = NULL,
  browse = TRUE,
  ...
)

Arguments

.data
a data.frame or a tbl_df.

...
arguments to be passed to methods.

target
target variable.

output_format
character. report output type. Choose either "pdf" and "html". "pdf" create pdf file by knitr::knit(). "html" create html file by rmarkdown::render().

output_file
character. name of generated file. default is NULL.

output_dir
character. name of directory to generate report file. default is tempdir().

font_family
character. font family name for figure in pdf.

browse
logical. choose whether to output the report results to the browser.

Details

Generate generalized EDA report automatically. You can choose to output as pdf and html files. This feature is useful for EDA of data with many variables, rather than data with fewer variables. For pdf output, Korean Gothic font must be installed in Korean operating system.

Reported information

The EDA process will report the following information:

Introduction

Information of Dataset

Information of Variables

About EDA Report

Univariate Analysis

Descriptive Statistics

Normality Test of Numerical Variables

Statistics and Visualization of (Sample) Data

Relationship Between Variables

Correlation Coefficient

Correlation Coefficient by Variable Combination

Correlation Plot of Numerical Variables

Target based Analysis

Grouped Descriptive Statistics

Grouped Numerical Variables

Grouped Categorical Variables

Grouped Relationship Between Variables

Grouped Correlation Coefficient

Grouped Correlation Plot of Numerical Variables

See vignette("EDA") for an introduction to these concepts.

Examples