Impute Missing Values

Missing values are imputed with some representative values and statistical methods.

imputate_na(.data, xvar, yvar, method, seed, print_flag, no_attrs)

Arguments

.data

a data.frame or a tbl_df.

xvar

variable name to replace missing value.

yvar

target variable.

method

method of missing values imputation.

seed

integer. the random seed used in mice. only used "mice" method.

print_flag

logical. If TRUE, mice will print running log on console. Use print_flag=FALSE for silent computation. Used only when method is "mice".

no_attrs

logical. If TRUE, return numerical variable or categorical variable. else If FALSE, imputation class.

Value

An object of imputation class. or numerical variable or categorical variable. if no_attrs is FALSE then return imputation class, else no_attrs is TRUE then return numerical vector or factor. Attributes of imputation class is as follows.

var_type : the data type of predictor to replace missing value.

method : method of missing value imputation.

predictor is numerical variable.

"mean" : arithmetic mean.

"median" : median.

"mode" : mode.

"knn" : K-nearest neighbors.

"rpart" : Recursive Partitioning and Regression Trees.

"mice" : Multivariate Imputation by Chained Equations.

predictor is categorical variable.

"mode" : mode.

"rpart" : Recursive Partitioning and Regression Trees.

"mice" : Multivariate Imputation by Chained Equations.

na_pos : position of missing value in predictor.

seed : the random seed used in mice. only used "mice" method.

type : "missing values". type of imputation.

message : a message tells you if the result was successful.

success : Whether the imputation was successful.

Details

imputate_na() creates an imputation class. The `imputation` class includes missing value position, imputed value, and method of missing value imputation, etc. The `imputation` class compares the imputed value with the original value to help determine whether the imputed value is used in the analysis.

See vignette("transformation") for an introduction to these concepts.

See also

Examples

# \donttest{
# Generate data for the example
heartfailure2 <- heartfailure
heartfailure2[sample(seq(NROW(heartfailure2)), 20), "platelets"] <- NA
heartfailure2[sample(seq(NROW(heartfailure2)), 5), "smoking"] <- NA

# Replace the missing value of the platelets variable with median
imputate_na(heartfailure2, platelets, method = "median")
#>   [1] 265000 263358 262000 210000 327000 204000 127000 454000 263358 262000
#>  [11] 368000 253000 136000 276000 427000  47000 262000 166000 237000  87000
#>  [21] 276000 297000 289000 368000 263358 149000 196000 284000 153000 200000
#>  [31] 263358 360000 319000 302000 188000 228000 226000 321000 305000 329000
#>  [41] 263358 153000 185000 262000 194000 310000 271000 451000 140000 395000
#>  [51] 166000 418000 263358 351000 255000 262000 223000 216000 319000 254000
#>  [61] 390000 216000 254000 385000 263358 119000 213000 274000 244000 497000
#>  [71] 374000 122000 243000 149000 262000 204000 317000 237000 283000 324000
#>  [81] 293000 263358 196000 172000 302000 406000 262000 304000 235000 181000
#>  [91] 249000 297000 263358 210000 327000 219000 254000 255000 318000 221000
#> [101] 298000 263358 149000 226000 286000 621000 263000 226000 304000 850000
#> [111] 306000 228000 252000 351000 262000 164000 262000 507000 203000 263358
#> [121] 262000 162000 228000 127000 217000 237000 271000 300000 267000 227000
#> [131] 249000 262000 263358 295000 231000 263358 172000 305000 221000 211000
#> [141] 263358 348000 329000 229000 338000 266000 218000 242000 225000 228000
#> [151] 235000 244000 184000 263358 235000 194000 277000 262000 235000 362000
#> [161] 242000 174000 448000  75000 334000 192000 220000  70000 270000 305000
#> [171] 263358 325000 176000 262000 281000 337000 105000 132000 267000 279000
#> [181] 303000 221000 265000 224000 219000 389000 153000 365000 201000 275000
#> [191] 350000 309000 260000 160000 262000 223000 263358 259000 279000 263358
#> [201]  73000 377000 220000 212000 277000 362000 226000 186000 283000 268000
#> [211] 389000 147000 481000 244000 290000 203000 358000 151000 271000 371000
#> [221] 263358 194000 365000 130000 504000 265000 189000 141000 262000 262000
#> [231]  62000 262000 255000 330000 305000 262000 248000 173000 257000 263358
#> [241] 533000 249000 255000 220000 264000 282000 314000 246000 262000 223000
#> [251] 404000 231000 274000 236000 263358 334000 294000 253000 233000 308000
#> [261] 203000 283000 198000 208000 147000 362000 263358 133000 302000 222000
#> [271] 263358 221000 215000 189000 150000 262000 327000  25100 232000 451000
#> [281] 241000  51000 215000 263358 279000 336000 279000 543000 263358 390000
#> [291] 222000 262000 382000 179000 155000 262000 742000 140000 395000
#> attr(,"var_type")
#> [1] "numerical"
#> attr(,"method")
#> [1] "median"
#> attr(,"na_pos")
#>  [1]   3  10  44  56  75  87 115 117 121 132 174 195 229 230 232 236 249 276 292
#> [20] 296
#> attr(,"type")
#> [1] "missing values"
#> attr(,"message")
#> [1] "complete imputation"
#> attr(,"success")
#> [1] TRUE
#> attr(,"class")
#> [1] "imputation" "numeric"   

# Replace the missing value of the platelets variable with rpart
# The target variable is death_event.
# Require rpart package
# imputate_na(heartfailure2, platelets, death_event, method = "rpart")

# Replace the missing value of the smoking variable with mode
# imputate_na(heartfailure2, smoking, method = "mode")

# Replace the missing value of the smoking variable with mice
# The target variable is death_event.
# imputate_na(heartfailure2, smoking, death_event, method = "mice")

## using dplyr -------------------------------------
library(dplyr)

# The mean before and after the imputation of the platelets variable
heartfailure2 %>%
  mutate(platelets_imp = imputate_na(heartfailure2, platelets, death_event, 
                                     method = "knn", no_attrs = TRUE)) %>%
  group_by(death_event) %>%
  summarise(orig = mean(platelets, na.rm = TRUE),
            imputation = mean(platelets_imp))
#> # A tibble: 2 × 3
#>   death_event    orig imputation
#>   <fct>         <dbl>      <dbl>
#> 1 No          267364.    266064.
#> 2 Yes         255050.    254483.

# If the variable of interest is a numerical variable
# Require rpart package
# platelets <- imputate_na(heartfailure2, platelets, death_event, method = "rpart")
# platelets
# summary(platelets)

# plot(platelets)

# If the variable of interest is a categorical variable
# smoking <- imputate_na(heartfailure2, smoking, death_event, method = "mice")
# smoking
# summary(smoking)

# plot(smoking)
# }