Job Change of Data Scientists

A dataset containing the gender and other attributes of almost 20000 cases.

data(jobchange)

Format

A data frame with 19158 rows and 14 variables. The variables are as follows:

enrollee_id

unique ID for candidate

city

city code.

city_dev_index

developement index of the city (scaled).

gender

gender of candidate.

relevent_experience

relevant experience of candidate

enrolled_university

type of University course enrolled if any.

education_level

education level of candidate.

major_discipline

education major discipline of candidate.

experience

candidate total experience in years.

company_size

number of employees in current employer's company.

company_type

type of current employer.

last_new_job

difference in years between previous job and current job.

training_hours

training hours completed.

job_chnge

if looking for a job change (boolean), Yes, No.

Source

"HR Analytics: Job Change of Data Scientists" in Kaggle <https://www.kaggle.com/arashnic/hr-analytics-job-change-of-data-scientists>, License : CC0(Public Domain

Details

This dataset designed to understand the factors that lead a person to leave current job for HR researches too.