Performs the Shapiro-Wilk test of normality

The normality() performs Shapiro-Wilk test of normality of numerical(INTEGER, NUMBER, etc.) column of the DBMS table through tbl_dbi.

# S3 method for tbl_dbi
normality(.data, ..., sample = 5000, in_database = FALSE, collect_size = Inf)

Arguments

.data
a tbl_dbi.

...
one or more unquoted expressions separated by commas. You can treat variable names like they are positions. Positive values select variables; negative values to drop variables. If the first expression is negative, normality() will automatically start with all variables. These arguments are automatically quoted and evaluated in a context where column names represent column positions. They support unquoting and splicing.

sample
the number of samples to perform the test.

in_database
Specifies whether to perform in-database operations. If TRUE, most operations are performed in the DBMS. if FALSE, table data is taken in R and operated in-memory. Not yet supported in_database = TRUE.

collect_size
a integer. The number of data samples from the DBMS to R. Applies only if in_database = FALSE.

See vignette("EDA") for an introduction to these concepts.

Value

An object of the same class as .data.

Details

This function is useful when used with the group_by function of the dplyr package. If you want to test by level of the categorical data you are interested in, rather than the whole observation, you can use group_tf as the group_by function. This function is computed shapiro.test function.

Normality test information

The information derived from the numerical data test is as follows.

statistic : the value of the Shapiro-Wilk statistic.

p_value : an approximate p-value for the test. This is said in Roystion(1995) to be adequate for p_value < 0.1.

sample : the numer of samples to perform the test. The number of observations supported by the stats::shapiro.test function is 3 to 5000.

See also

Examples

# \donttest{
library(dplyr)

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

# connect DBMS
con_sqlite <- DBI::dbConnect(RSQLite::SQLite(), ":memory:")

# copy carseats to the DBMS with a table named TB_CARSEATS
copy_to(con_sqlite, carseats, name = "TB_CARSEATS", overwrite = TRUE)

# Using pipes ---------------------------------
# Normality test of all numerical variables
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  normality()

#> # A tibble: 8 x 4
#>   vars        statistic  p_value sample
#>   <chr>           <dbl>    <dbl>  <dbl>
#> 1 Sales           0.995 2.54e- 1    400
#> 2 CompPrice       0.998 9.77e- 1    400
#> 3 Income          0.962 2.37e- 8    400
#> 4 Advertising     0.874 1.49e-17    400
#> 5 Population      0.952 4.08e-10    400
#> 6 Price           0.996 3.90e- 1    400
#> 7 Age             0.957 1.86e- 9    400
#> 8 Education       0.924 2.43e-13    400

# Positive values select variables, and In-memory mode and collect size is 200
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  normality(Sales, Price, collect_size  = 200)

#> # A tibble: 2 x 4
#>   vars  statistic p_value sample
#>   <chr>     <dbl>   <dbl>  <dbl>
#> 1 Sales     0.994   0.550    200
#> 2 Price     0.989   0.113    200

# Positions values select variables
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  normality(1)

#> # A tibble: 1 x 4
#>   vars  statistic p_value sample
#>   <chr>     <dbl>   <dbl>  <dbl>
#> 1 Sales     0.995   0.254    400

# Using pipes & dplyr -------------------------
# Test all numerical variables by 'ShelveLoc' and 'US',
# and extract only those with 'ShelveLoc' variable level is "Good".
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  group_by(ShelveLoc, US) %>%
  normality() %>%
  filter(ShelveLoc == "Good")

#> # A tibble: 16 x 6
#>    variable    ShelveLoc US    statistic  p_value sample
#>    <chr>       <chr>     <chr>     <dbl>    <dbl>  <dbl>
#>  1 Sales       Good      No        0.955 3.42e- 1     24
#>  2 Sales       Good      Yes       0.983 5.67e- 1     61
#>  3 CompPrice   Good      No        0.970 6.58e- 1     24
#>  4 CompPrice   Good      Yes       0.984 5.98e- 1     61
#>  5 Income      Good      No        0.900 2.18e- 2     24
#>  6 Income      Good      Yes       0.957 4.28e- 2     61
#>  7 Advertising Good      No        0.209 2.19e-10     24
#>  8 Advertising Good      Yes       0.976 2.65e- 1     61
#>  9 Population  Good      No        0.976 8.14e- 1     24
#> 10 Population  Good      Yes       0.971 1.56e- 1     61
#> 11 Price       Good      No        0.963 5.12e- 1     24
#> 12 Price       Good      Yes       0.978 3.46e- 1     61
#> 13 Age         Good      No        0.944 2.01e- 1     24
#> 14 Age         Good      Yes       0.944 7.92e- 3     61
#> 15 Education   Good      No        0.935 1.29e- 1     24
#> 16 Education   Good      Yes       0.907 2.18e- 4     61

# extract only those with 'Urban' variable level is "Yes",
# and test 'Sales' by 'ShelveLoc' and 'US'
con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  filter(Urban == "Yes") %>%
  group_by(ShelveLoc, US) %>%
  normality(Sales)

#> # A tibble: 6 x 6
#>   variable ShelveLoc US    statistic p_value sample
#>   <chr>    <chr>     <chr>     <dbl>   <dbl>  <dbl>
#> 1 Sales    Bad       No        0.985   0.968     23
#> 2 Sales    Bad       Yes       0.985   0.747     51
#> 3 Sales    Good      No        0.959   0.576     18
#> 4 Sales    Good      Yes       0.968   0.341     38
#> 5 Sales    Medium    No        0.980   0.514     53
#> 6 Sales    Medium    Yes       0.992   0.814     96

# Test log(Income) variables by 'ShelveLoc' and 'US',
# and extract only p.value greater than 0.01.

# SQLite extension functions for log
RSQLite::initExtension(con_sqlite)

con_sqlite %>%
  tbl("TB_CARSEATS") %>%
  mutate(log_income = log(Income)) %>%
  group_by(ShelveLoc, US) %>%
  normality(log_income) %>%
  filter(p_value > 0.01)

#> # A tibble: 1 x 6
#>   variable   ShelveLoc US    statistic p_value sample
#>   <chr>      <chr>     <chr>     <dbl>   <dbl>  <dbl>
#> 1 log_income Bad       No        0.946   0.104     34

# }