Diagnose Performance Binned Variable

The performance_bin() calculates metrics to evaluate the performance of binned variable for binomial classification model.

performance_bin(y, x, na.rm = FALSE)

Arguments

y
character or numeric, integer, factor. a binary response variable (0, 1). The variable must contain only the integers 0 and 1 as element. However, in the case of factor/character having two levels, it is performed while type conversion is performed in the calculation process.

x
integer or factor, character. At least 2 different values. and Inf is not allowed.

na.rm
logical. a logical indicating whether missing values should be removed.

Value

an object of "performance_bin" class. vaue of data.frame is as follows.

Bin : character. bins.

CntRec : integer. frequency by bins.

CntPos : integer. frequency of positive by bins.

CntNeg : integer. frequency of negative by bins.

CntCumPos : integer. cumulate frequency of positive by bins.

CntCumNeg : integer. cumulate frequency of negative by bins.

RatePos : integer. relative frequency of positive by bins.

RateNeg : integer. relative frequency of negative by bins.

RateCumPos : numeric. cumulate relative frequency of positive by bins.

RateCumNeg : numeric. cumulate relative frequency of negative by bins.

Odds : numeric. odd ratio.

LnOdds : numeric. loged odd ratio.

WoE : numeric. weight of evidence.

IV : numeric. Jeffrey's Information Value.

JSD : numeric. Jensen-Shannon Divergence.

AUC : numeric. AUC. area under curve.

Attributes of "performance_bin" class is as follows.

names : character. variable name of data.frame with "Binning Table".

class : character. name of class. "performance_bin" "data.frame".

row.names : character. row name of data.frame with "Binning Table".

IV : numeric. Jeffrey's Information Value.

JSD : numeric. Jensen-Shannon Divergence.

KS : numeric. Kolmogorov-Smirnov Statistics.

gini : numeric. Gini index.

HHI : numeric. Herfindahl-Hirschman Index.

HHI_norm : numeric.normalized Herfindahl-Hirschman Index.

Cramer_V : numeric. Cramer's V Statistics.

chisq_test : data.frame. table of significance tests. name is as follows.

Bin A : character. first bins.

Bin B : character. second bins.

statistics : numeric. statistics of Chi-square test.

p_value : numeric. p-value of Chi-square test.

Details

This function is useful when used with the mutate/transmute function of the dplyr package.

See also

Examples

# Generate data for the example
carseats <- ISLR::Carseats

set.seed(123)
carseats[sample(seq(NROW(carseats)), 20), "Advertising"] <- NA

# Change the target variable to 0(negative) and 1(positive).
carseats$US_2 <- ifelse(carseats$US %in% "Yes", 1, 0)

# Binnig from Advertising to Advertising_bin.
breaks <- c(-1,  0,  6, 29)
carseats$Advertising_bin <- cut(carseats$Advertising, breaks)

# Diagnose performance binned variable
perf <- performance_bin(carseats$US_2, carseats$Advertising_bin)
perf

#>      Bin CntRec CntPos CntNeg CntCumPos CntCumNeg RatePos RateNeg RateCumPos RateCumNeg
#> 1 (-1,0]    133     19    114        19       114 0.07364 0.80282    0.07364    0.80282
#> 2  (0,6]     66     52     14        71       128 0.20155 0.09859    0.27519    0.90141
#> 3 (6,29]    181    179      2       250       130 0.69380 0.01408    0.96899    0.91549
#> 4   <NA>     20      8     12       258       142 0.03101 0.08451    1.00000    1.00000
#> 5  Total    400    258    142        NA        NA 1.00000 1.00000         NA         NA
#>       Odds   LnOdds      WoE      IV     JSD     AUC
#> 1  0.16667 -1.79176 -2.38889 1.74192 0.17733 0.02956
#> 2  3.71429  1.31219  0.71505 0.07362 0.00901 0.01720
#> 3 89.50000  4.49424  3.89711 2.64892 0.21078 0.00876
#> 4  0.66667 -0.40547 -1.00260 0.05364 0.00644 0.08320
#> 5  1.81690  0.59713       NA 4.51809 0.40356 0.13872

summary(perf)

#> ── Binning Table ──────────────────────── Several Metrics ──
#>      Bin CntRec CntPos CntNeg RatePos RateNeg     Odds      WoE      IV     JSD     AUC
#> 1 (-1,0]    133     19    114 0.07364 0.80282  0.16667 -2.38889 1.74192 0.17733 0.02956
#> 2  (0,6]     66     52     14 0.20155 0.09859  3.71429  0.71505 0.07362 0.00901 0.01720
#> 3 (6,29]    181    179      2 0.69380 0.01408 89.50000  3.89711 2.64892 0.21078 0.00876
#> 4   <NA>     20      8     12 0.03101 0.08451  0.66667 -1.00260 0.05364 0.00644 0.08320
#> 5  Total    400    258    142 1.00000 1.00000  1.81690       NA 4.51809 0.40356 0.13872
#> 
#> ── General Metrics ─────────────────────────────────────────
#> ● Gini index                       :  -0.72257
#> ● IV (Jeffrey)                     :  4.51809
#> ● JS (Jensen-Shannon) Divergence   :  0.40356
#> ● Kolmogorov-Smirnov Statistics    :  0.72917
#> ● HHI (Herfindahl-Hirschman Index) :  0.34504
#> ● HHI (normalized)                 :  0.12672
#> ● Cramer's V                       :  0.81086
#> 
#> ── Significance Tests ──────────────────── Chisquare Test ──
#>    Bin A  Bin B statistics      p_value
#> 1 (-1,0]  (0,6]   79.97033 3.800735e-19
#> 2  (0,6] (6,29]   32.27665 1.337104e-08
#> 

# plot(perf)

# Diagnose performance binned variable without NA
perf <- performance_bin(carseats$US_2, carseats$Advertising_bin, na.rm = TRUE)
perf

#>      Bin CntRec CntPos CntNeg CntCumPos CntCumNeg RatePos RateNeg RateCumPos RateCumNeg
#> 1 (-1,0]    133     19    114        19       114   0.076 0.87692      0.076    0.87692
#> 2  (0,6]     66     52     14        71       128   0.208 0.10769      0.284    0.98462
#> 3 (6,29]    181    179      2       250       130   0.716 0.01538      1.000    1.00000
#> 4  Total    380    250    130        NA        NA   1.000 1.00000         NA         NA
#>       Odds   LnOdds      WoE      IV     JSD     AUC
#> 1  0.16667 -1.79176 -2.44569 1.95881 0.19772 0.03332
#> 2  3.71429  1.31219  0.65826 0.06603 0.00811 0.01938
#> 3 89.50000  4.49424  3.84031 2.69058 0.21616 0.00988
#> 4  1.92308  0.65393       NA 4.71542 0.42199 0.06258

summary(perf)

#> ── Binning Table ──────────────────────── Several Metrics ──
#>      Bin CntRec CntPos CntNeg RatePos RateNeg     Odds      WoE      IV     JSD     AUC
#> 1 (-1,0]    133     19    114   0.076 0.87692  0.16667 -2.44569 1.95881 0.19772 0.03332
#> 2  (0,6]     66     52     14   0.208 0.10769  3.71429  0.65826 0.06603 0.00811 0.01938
#> 3 (6,29]    181    179      2   0.716 0.01538 89.50000  3.84031 2.69058 0.21616 0.00988
#> 4  Total    380    250    130   1.000 1.00000  1.92308       NA 4.71542 0.42199 0.06258
#> 
#> ── General Metrics ─────────────────────────────────────────
#> ● Gini index                       :  -0.87483
#> ● IV (Jeffrey)                     :  4.71542
#> ● JS (Jensen-Shannon) Divergence   :  0.42199
#> ● Kolmogorov-Smirnov Statistics    :  0.80092
#> ● HHI (Herfindahl-Hirschman Index) :  0.37954
#> ● HHI (normalized)                 :  0.06931
#> ● Cramer's V                       :  0.81086
#> 
#> ── Significance Tests ──────────────────── Chisquare Test ──
#>    Bin A  Bin B statistics      p_value
#> 1 (-1,0]  (0,6]   79.97033 3.800735e-19
#> 2  (0,6] (6,29]   32.27665 1.337104e-08
#> 

# plot(perf)