Diagnose Performance Binned Variable

The performance_bin() calculates metrics to evaluate the performance of binned variable for binomial classification model.

performance_bin(y, x, na.rm = FALSE)

Arguments

y
character or numeric, integer, factor. a binary response variable (0, 1). The variable must contain only the integers 0 and 1 as element. However, in the case of factor/character having two levels, it is performed while type conversion is performed in the calculation process.

x
integer or factor, character. At least 2 different values. and Inf is not allowed.

na.rm
logical. a logical indicating whether missing values should be removed.

Value

an object of "performance_bin" class. vaue of data.frame is as follows.

Bin : character. bins.

CntRec : integer. frequency by bins.

CntPos : integer. frequency of positive by bins.

CntNeg : integer. frequency of negative by bins.

CntCumPos : integer. cumulate frequency of positive by bins.

CntCumNeg : integer. cumulate frequency of negative by bins.

RatePos : integer. relative frequency of positive by bins.

RateNeg : integer. relative frequency of negative by bins.

RateCumPos : numeric. cumulate relative frequency of positive by bins.

RateCumNeg : numeric. cumulate relative frequency of negative by bins.

Odds : numeric. odd ratio.

LnOdds : numeric. loged odd ratio.

WoE : numeric. weight of evidence.

IV : numeric. Jeffrey's Information Value.

JSD : numeric. Jensen-Shannon Divergence.

AUC : numeric. AUC. area under curve.

Attributes of "performance_bin" class is as follows.

names : character. variable name of data.frame with "Binning Table".

class : character. name of class. "performance_bin" "data.frame".

row.names : character. row name of data.frame with "Binning Table".

IV : numeric. Jeffrey's Information Value.

JSD : numeric. Jensen-Shannon Divergence.

KS : numeric. Kolmogorov-Smirnov Statistics.

gini : numeric. Gini index.

HHI : numeric. Herfindahl-Hirschman Index.

HHI_norm : numeric.normalized Herfindahl-Hirschman Index.

Cramer_V : numeric. Cramer's V Statistics.

chisq_test : data.frame. table of significance tests. name is as follows.

Bin A : character. first bins.

Bin B : character. second bins.

statistics : numeric. statistics of Chi-square test.

p_value : numeric. p-value of Chi-square test.

Details

This function is useful when used with the mutate/transmute function of the dplyr package.

See also

Examples

# Generate data for the example
heartfailure2 <- heartfailure

set.seed(123)
heartfailure2[sample(seq(NROW(heartfailure2)), 5), "creatinine"] <- NA

# Change the target variable to 0(negative) and 1(positive).
heartfailure2$death_event_2 <- ifelse(heartfailure2$death_event %in% "Yes", 1, 0)

# Binnig from creatinine to platelets_bin.
breaks <- c(0,  1,  2, 10)
heartfailure2$creatinine_bin <- cut(heartfailure2$creatinine, breaks)

# Diagnose performance binned variable
perf <- performance_bin(heartfailure2$death_event_2, heartfailure2$creatinine_bin)
perf

#>      Bin CntRec CntPos CntNeg CntCumPos CntCumNeg RatePos RateNeg RateCumPos
#> 1  (0,1]    128     24    104        24       104 0.25000 0.51232    0.25000
#> 2  (1,2]    132     51     81        75       185 0.53125 0.39901    0.78125
#> 3 (2,10]     34     21     13        96       198 0.21875 0.06404    1.00000
#> 4   <NA>      5      0      5        96       203 0.00000 0.02463    1.00000
#> 5  Total    299     96    203        NA        NA 1.00000 1.00000         NA
#>   RateCumNeg    Odds   LnOdds      WoE      IV     JSD     AUC
#> 1    0.51232 0.23077 -1.46634 -0.71748 0.18821 0.02303 0.06404
#> 2    0.91133 0.62963 -0.46262  0.28623 0.03785 0.00472 0.20574
#> 3    0.97537 1.61538  0.47957  1.22843 0.19005 0.02237 0.05704
#> 4    1.00000 0.00000     -Inf       NA      NA      NA 0.02463
#> 5         NA 0.47291 -0.74886       NA      NA      NA 0.35145

summary(perf)

#> ── Binning Table ──────────────────────── Several Metrics ── 
#>      Bin CntRec CntPos CntNeg RatePos RateNeg    Odds      WoE      IV     JSD
#> 1  (0,1]    128     24    104 0.25000 0.51232 0.23077 -0.71748 0.18821 0.02303
#> 2  (1,2]    132     51     81 0.53125 0.39901 0.62963  0.28623 0.03785 0.00472
#> 3 (2,10]     34     21     13 0.21875 0.06404 1.61538  1.22843 0.19005 0.02237
#> 4   <NA>      5      0      5 0.00000 0.02463 0.00000       NA      NA      NA
#> 5  Total    299     96    203 1.00000 1.00000 0.47291       NA      NA      NA
#>       AUC
#> 1 0.06404
#> 2 0.20574
#> 3 0.05704
#> 4 0.02463
#> 5 0.35145
#> 
#> ── General Metrics ───────────────────────────────────────── 
#> • Gini index                       :  -0.29711
#> • IV (Jeffrey)                     :  NA
#> • JS (Jensen-Shannon) Divergence   :  NA
#> • Kolmogorov-Smirnov Statistics    :  0.26232
#> • HHI (Herfindahl-Hirschman Index) :  0.39137
#> • HHI (normalized)                 :  0.1885
#> • Cramer's V                       :  0.30024 
#> 
#> ── Significance Tests ──────────────────── Chisquare Test ── 
#>   Bin A  Bin B statistics      p_value
#> 1 (0,1]  (1,2]  12.520885 0.0004024283
#> 2 (1,2] (2,10]   5.888296 0.0152418462
#> 

# plot(perf)

# Diagnose performance binned variable without NA
# perf <- performance_bin(heartfailure2$death_event_2, heartfailure2$creatinine_bin, na.rm = TRUE) 
# perf
# summary(perf)

# plot(perf)