Summarizing Binned Variable

summary method for "bins" and "optimal_bins".

# S3 method for bins
summary(object, ...)

# S3 method for bins
print(x, ...)

Arguments

object
an object of "bins" and "optimal_bins", usually, a result of a call to binning().

...
further arguments passed to or from other methods.

x
an object of class "bins" and "optimal_bins", usually, a result of a call to binning().

Value

The function summary.bins() computes and returns a data.frame of summary statistics of the binned given in object. Variables of data frame is as follows.

levels : levels of factor.

freq : frequency of levels.

rate : relative frequency of levels. it is not percentage.

Details

print.bins() prints the information of "bins" and "optimal_bins" objects nicely. This includes frequency of bins, binned type, and number of bins. summary.bins() returns data.frame including frequency and relative frequency for each levels(bins).

See vignette("transformation") for an introduction to these concepts.

See also

Examples

# Generate data for the example
heartfailure2 <- heartfailure
heartfailure2[sample(seq(NROW(heartfailure2)), 20), "platelets"] <- NA

# Binning the carat variable. default type argument is "quantile"
bin <- binning(heartfailure2$platelets)

# Print bins class object
bin

#> binned type: quantile
#> number of bins: 10
#> x
#> [2.51e+04,1.53e+05] (1.53e+05,1.96e+05] (1.96e+05,2.21e+05] (2.21e+05,2.37e+05] 
#>                  29                  28                  29                  29 
#> (2.37e+05,2.62e+05] (2.62e+05,2.64e+05] (2.64e+05,2.83e+05] (2.83e+05,3.19e+05] 
#>                  26                  27                  29                  28 
#> (3.19e+05,3.81e+05]  (3.81e+05,8.5e+05]                <NA> 
#>                  26                  28                  20 

# Summarize bins class object
summary(bin)

#>                 levels freq       rate
#> 1  [2.51e+04,1.53e+05]   29 0.09698997
#> 2  (1.53e+05,1.96e+05]   28 0.09364548
#> 3  (1.96e+05,2.21e+05]   29 0.09698997
#> 4  (2.21e+05,2.37e+05]   29 0.09698997
#> 5  (2.37e+05,2.62e+05]   26 0.08695652
#> 6  (2.62e+05,2.64e+05]   27 0.09030100
#> 7  (2.64e+05,2.83e+05]   29 0.09698997
#> 8  (2.83e+05,3.19e+05]   28 0.09364548
#> 9  (3.19e+05,3.81e+05]   26 0.08695652
#> 10  (3.81e+05,8.5e+05]   28 0.09364548
#> 11                <NA>   20 0.06688963