Summarizing compare_category information

print and summary method for "compare_category" class.

# S3 method for compare_category
summary(
  object,
  method = c("all", "table", "relative", "chisq"),
  pos = NULL,
  na.rm = TRUE,
  marginal = FALSE,
  verbose = TRUE,
  ...
)

# S3 method for compare_category
print(x, ...)

Arguments

object
an object of class "compare_category", usually, a result of a call to compare_category().

method
character. Specifies the type of information to be aggregated. "table" create contingency table, "relative" create relative contingency table, and "chisq" create information of chi-square test. and "all" aggregates all information. The default is "all"

pos
integer. Specifies the pair of variables to be summarized by index. The default is NULL, which aggregates all variable pairs.

na.rm
logical. Specifies whether to include NA when counting the contingency tables or performing a chi-square test. The default is TRUE, where NA is removed and aggregated.

marginal
logical. Specifies whether to add marginal values to the contingency table. The default value is FALSE, so no marginal value is added.

verbose
logical. Specifies whether to output additional information during the calculation process. The default is to output information as TRUE. In this case, the function returns the value with invisible(). If FALSE, the value is returned by return().

...
further arguments passed to or from other methods.

x
an object of class "compare_category", usually, a result of a call to compare_category().

Details

print.compare_category() displays only the information compared between the variables included in compare_category. The "type", "variables" and "combination" attributes are not displayed. When using summary.compare_category(), it is advantageous to set the verbose argument to TRUE if the user is only viewing information from the console. It is also advantageous to specify FALSE if you want to manipulate the results.

See also

Examples

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

library(dplyr)

# Compare the all categorical variables
all_var <- compare_category(carseats)

# Print compare class object
all_var

#> $`ShelveLoc vs Urban`
#> # A tibble: 9 x 6
#>   ShelveLoc Urban     n   rate var1_rate var2_rate
#>   <fct>     <fct> <int>  <dbl>     <dbl>     <dbl>
#> 1 Bad       No       21 0.0525   0.219       0.179
#> 2 Bad       Yes      73 0.182    0.760       0.263
#> 3 Bad       NA        2 0.005    0.0208      0.4  
#> 4 Good      No       28 0.07     0.329       0.239
#> 5 Good      Yes      56 0.14     0.659       0.201
#> 6 Good      NA        1 0.0025   0.0118      0.2  
#> 7 Medium    No       68 0.17     0.311       0.581
#> 8 Medium    Yes     149 0.372    0.680       0.536
#> 9 Medium    NA        2 0.005    0.00913     0.4  
#> 
#> $`ShelveLoc vs US`
#> # A tibble: 6 x 6
#>   ShelveLoc US        n  rate var1_rate var2_rate
#>   <fct>     <fct> <int> <dbl>     <dbl>     <dbl>
#> 1 Bad       No       34 0.085     0.354     0.239
#> 2 Bad       Yes      62 0.155     0.646     0.240
#> 3 Good      No       24 0.06      0.282     0.169
#> 4 Good      Yes      61 0.152     0.718     0.236
#> 5 Medium    No       84 0.21      0.384     0.592
#> 6 Medium    Yes     135 0.338     0.616     0.523
#> 
#> $`Urban vs US`
#> # A tibble: 6 x 6
#>   Urban US        n   rate var1_rate var2_rate
#>   <fct> <fct> <int>  <dbl>     <dbl>     <dbl>
#> 1 No    No       45 0.112      0.385   0.317  
#> 2 No    Yes      72 0.18       0.615   0.279  
#> 3 Yes   No       96 0.24       0.345   0.676  
#> 4 Yes   Yes     182 0.455      0.655   0.705  
#> 5 NA    No        1 0.0025     0.2     0.00704
#> 6 NA    Yes       4 0.01       0.8     0.0155 
#> 

# Compare the two categorical variables
two_var <- compare_category(carseats, ShelveLoc, Urban)

# Print compare class object
two_var

#> $`ShelveLoc vs Urban`
#> # A tibble: 9 x 6
#>   ShelveLoc Urban     n   rate var1_rate var2_rate
#>   <fct>     <fct> <int>  <dbl>     <dbl>     <dbl>
#> 1 Bad       No       21 0.0525   0.219       0.179
#> 2 Bad       Yes      73 0.182    0.760       0.263
#> 3 Bad       NA        2 0.005    0.0208      0.4  
#> 4 Good      No       28 0.07     0.329       0.239
#> 5 Good      Yes      56 0.14     0.659       0.201
#> 6 Good      NA        1 0.0025   0.0118      0.2  
#> 7 Medium    No       68 0.17     0.311       0.581
#> 8 Medium    Yes     149 0.372    0.680       0.536
#> 9 Medium    NA        2 0.005    0.00913     0.4  
#> 

# Summary the all case : Return a invisible copy of an object.
stat <- summary(all_var)

#> ── Contingency tables ───────────────────────────── Number of table is 3 ──
#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes
#>    Bad     21  73
#>    Good    28  56
#>    Medium  68 149
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes
#>    Bad     34  62
#>    Good    24  61
#>    Medium  84 135
#> 
#> $`Urban vs US`
#>      US
#> Urban  No Yes
#>   No   45  72
#>   Yes  96 182
#> 
#> ── Relative contingency tables ──────────────────── Number of table is 3 ──
#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc         No        Yes
#>    Bad    0.05316456 0.18481013
#>    Good   0.07088608 0.14177215
#>    Medium 0.17215190 0.37721519
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc     No    Yes
#>    Bad    0.0850 0.1550
#>    Good   0.0600 0.1525
#>    Medium 0.2100 0.3375
#> 
#> $`Urban vs US`
#>      US
#> Urban        No       Yes
#>   No  0.1139241 0.1822785
#>   Yes 0.2430380 0.4607595
#> 
#> ── Chi-squared contingency table tests ──────────── Number of table is 3 ──
#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 3.2517549 0.1967390  2
#> 2  ShelveLoc         US 2.7396674 0.2541492  2
#> 3      Urban         US 0.3958796 0.5292249  1

# Summary by returned object
stat

#> $table
#> $table$`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes
#>    Bad     21  73
#>    Good    28  56
#>    Medium  68 149
#> 
#> $table$`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes
#>    Bad     34  62
#>    Good    24  61
#>    Medium  84 135
#> 
#> $table$`Urban vs US`
#>      US
#> Urban  No Yes
#>   No   45  72
#>   Yes  96 182
#> 
#> 
#> $relative
#> $relative$`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc         No        Yes
#>    Bad    0.05316456 0.18481013
#>    Good   0.07088608 0.14177215
#>    Medium 0.17215190 0.37721519
#> 
#> $relative$`ShelveLoc vs US`
#>          US
#> ShelveLoc     No    Yes
#>    Bad    0.0850 0.1550
#>    Good   0.0600 0.1525
#>    Medium 0.2100 0.3375
#> 
#> $relative$`Urban vs US`
#>      US
#> Urban        No       Yes
#>   No  0.1139241 0.1822785
#>   Yes 0.2430380 0.4607595
#> 
#> 
#> $chisq
#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 3.2517549 0.1967390  2
#> 2  ShelveLoc         US 2.7396674 0.2541492  2
#> 3      Urban         US 0.3958796 0.5292249  1
#> 

# component of table 
stat$table

#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes
#>    Bad     21  73
#>    Good    28  56
#>    Medium  68 149
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes
#>    Bad     34  62
#>    Good    24  61
#>    Medium  84 135
#> 
#> $`Urban vs US`
#>      US
#> Urban  No Yes
#>   No   45  72
#>   Yes  96 182
#> 

# component of chi-square test 
stat$chisq

#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 3.2517549 0.1967390  2
#> 2  ShelveLoc         US 2.7396674 0.2541492  2
#> 3      Urban         US 0.3958796 0.5292249  1

# component of chi-square test 
summary(all_var, "chisq")

#> ── Chi-squared contingency table tests ──────────── Number of table is 3 ──
#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 3.2517549 0.1967390  2
#> 2  ShelveLoc         US 2.7396674 0.2541492  2
#> 3      Urban         US 0.3958796 0.5292249  1

# component of chi-square test (first, third case)
summary(all_var, "chisq", pos = c(1, 3))

#> ── Chi-squared contingency table tests ──────────── Number of table is 2 ──
#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 3.2517549 0.1967390  2
#> 2      Urban         US 0.3958796 0.5292249  1

# component of relative frequency table 
summary(all_var, "relative")

#> ── Relative contingency tables ──────────────────── Number of table is 3 ──
#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc         No        Yes
#>    Bad    0.05316456 0.18481013
#>    Good   0.07088608 0.14177215
#>    Medium 0.17215190 0.37721519
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc     No    Yes
#>    Bad    0.0850 0.1550
#>    Good   0.0600 0.1525
#>    Medium 0.2100 0.3375
#> 
#> $`Urban vs US`
#>      US
#> Urban        No       Yes
#>   No  0.1139241 0.1822785
#>   Yes 0.2430380 0.4607595
#> 

# component of table without missing values 
summary(all_var, "table", na.rm = TRUE)

#> ── Contingency tables ───────────────────────────── Number of table is 3 ──
#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes
#>    Bad     21  73
#>    Good    28  56
#>    Medium  68 149
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes
#>    Bad     34  62
#>    Good    24  61
#>    Medium  84 135
#> 
#> $`Urban vs US`
#>      US
#> Urban  No Yes
#>   No   45  72
#>   Yes  96 182
#> 

# component of table include marginal value 
margin <- summary(all_var, "table", marginal = TRUE)

#> ── Contingency tables ───────────────────────────── Number of table is 3 ──
#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes <Total>
#>   Bad      21  73      94
#>   Good     28  56      84
#>   Medium   68 149     217
#>   <Total> 117 278     395
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes <Total>
#>   Bad      34  62      96
#>   Good     24  61      85
#>   Medium   84 135     219
#>   <Total> 142 258     400
#> 
#> $`Urban vs US`
#>          US
#> Urban      No Yes <Total>
#>   No       45  72     117
#>   Yes      96 182     278
#>   <Total> 141 254     395
#> 

margin

#> $`ShelveLoc vs Urban`
#>          Urban
#> ShelveLoc  No Yes <Total>
#>   Bad      21  73      94
#>   Good     28  56      84
#>   Medium   68 149     217
#>   <Total> 117 278     395
#> 
#> $`ShelveLoc vs US`
#>          US
#> ShelveLoc  No Yes <Total>
#>   Bad      34  62      96
#>   Good     24  61      85
#>   Medium   84 135     219
#>   <Total> 142 258     400
#> 
#> $`Urban vs US`
#>          US
#> Urban      No Yes <Total>
#>   No       45  72     117
#>   Yes      96 182     278
#>   <Total> 141 254     395
#> 

# component of chi-square test 
summary(two_var, method = "chisq")

#> ── Chi-squared contingency table tests ──────────── Number of table is 1 ──
#>   variable_1 variable_2 statistic  p.value df
#> 1  ShelveLoc      Urban  3.251755 0.196739  2

# verbose is FALSE 
summary(all_var, "chisq", verbose = FALSE)

#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban 3.2517549 0.1967390  2
#> 2  ShelveLoc         US 2.7396674 0.2541492  2
#> 3      Urban         US 0.3958796 0.5292249  1

#' # Using pipes & dplyr -------------------------
# If you want to use dplyr, set verbose to FALSE
summary(all_var, "chisq", verbose = FALSE) %>%
  filter(p.value < 0.26)

#>   variable_1 variable_2 statistic   p.value df
#> 1  ShelveLoc      Urban  3.251755 0.1967390  2
#> 2  ShelveLoc         US  2.739667 0.2541492  2

# Extract component from list by index
summary(all_var, "table", na.rm = TRUE, verbose = FALSE) %>%
  "[["(1)

#>          Urban
#> ShelveLoc  No Yes
#>    Bad     21  73
#>    Good    28  56
#>    Medium  68 149

# Extract component from list by name
  summary(all_var, "table", na.rm = TRUE, verbose = FALSE) %>%
  "[["("ShelveLoc vs Urban")

#>          Urban
#> ShelveLoc  No Yes
#>    Bad     21  73
#>    Good    28  56
#>    Medium  68 149