Summarizing compare_numeric information

print and summary method for "compare_numeric" class.

# S3 method for compare_numeric
summary(
  object,
  method = c("all", "correlation", "linear"),
  thres_corr = 0.3,
  thres_rs = 0.1,
  verbose = TRUE,
  ...
)

# S3 method for compare_numeric
print(x, ...)

Arguments

object
an object of class "compare_numeric", usually, a result of a call to compare_numeric().

method
character. Select statistics to be aggregated. "correlation" calculates the Pearson's correlation coefficient, and "linear" returns the aggregation of the linear model. "all" returns both information. However, the difference between summary.compare_numeric() and compare_numeric() is that only cases that are greater than the specified threshold are returned. "correlation" returns only cases with a correlation coefficient greater than the thres_corr argument value. "linear" returns only cases with R^2 greater than the thres_rs argument.

thres_corr
numeric. This is the correlation coefficient threshold of the correlation coefficient information to be returned. The default is 0.3.

thres_rs
numeric. R^2 threshold of linear model summaries information to return. The default is 0.1.

verbose
logical. Specifies whether to output additional information during the calculation process. The default is to output information as TRUE. In this case, the function returns the value with invisible(). If FALSE, the value is returned by return().

...
further arguments passed to or from other methods.

x
an object of class "compare_numeric", usually, a result of a call to compare_numeric().

Value

An object of the class as compare based list. The information to examine the relationship between numerical variables is as follows each components. - correlation component : Pearson's correlation coefficient.

var1 : factor. The level of the first variable to compare. 'var1' is the name of the first variable to be compared.

var2 : factor. The level of the second variable to compare. 'var2' is the name of the second variable to be compared.

coef_corr : double. Pearson's correlation coefficient.

- linear component : linear model summaries

var1 : factor. The level of the first variable to compare. 'var1' is the name of the first variable to be compared.

var2 : factor. The level of the second variable to compare. 'var2' is the name of the second variable to be compared.

r.squared : double. The percent of variance explained by the model.

adj.r.squared : double. r.squared adjusted based on the degrees of freedom.

sigma : double. The square root of the estimated residual variance.

statistic : double. F-statistic.

p.value : double. p-value from the F test, describing whether the full regression is significant.

df : integer degrees of freedom.

logLik : double. the log-likelihood of data under the model.

AIC : double. the Akaike Information Criterion.

BIC : double. the Bayesian Information Criterion.

deviance : double. deviance.

df.residual : integer residual degrees of freedom.

Details

print.compare_numeric() displays only the information compared between the variables included in compare_numeric. When using summary.compare_numeric(), it is advantageous to set the verbose argument to TRUE if the user is only viewing information from the console. It is also advantageous to specify FALSE if you want to manipulate the results.

See also

Examples

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

library(dplyr)

# Compare the all numerical variables
all_var <- compare_numeric(carseats)

# Print compare class object
all_var

#> $correlation
#> # A tibble: 28 x 3
#>    var1      var2        coef_corr
#>    <chr>     <chr>           <dbl>
#>  1 Sales     CompPrice      0.0641
#>  2 Sales     Income         0.143 
#>  3 Sales     Advertising    0.270 
#>  4 Sales     Population     0.0505
#>  5 Sales     Price         -0.445 
#>  6 Sales     Age           -0.232 
#>  7 Sales     Education     -0.0520
#>  8 CompPrice Income        -0.0868
#>  9 CompPrice Advertising   -0.0242
#> 10 CompPrice Population    -0.0947
#> # … with 18 more rows
#> 
#> $linear
#> # A tibble: 28 x 14
#>    var1  var2  r.squared adj.r.squared sigma statistic  p.value    df logLik   AIC   BIC
#>    <chr> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl> <dbl> <dbl>
#>  1 Sales Comp…  0.00411      0.00160    2.82     1.64  2.01e- 1     1  -982. 1969. 1981.
#>  2 Sales Inco…  0.0204       0.0178     2.77     7.87  5.28e- 3     1  -926. 1858. 1870.
#>  3 Sales Adve…  0.0726       0.0703     2.72    31.2   4.38e- 8     1  -967. 1941. 1953.
#>  4 Sales Popu…  0.00255      0.0000412  2.82     1.02  3.14e- 1     1  -982. 1970. 1982.
#>  5 Sales Price  0.198        0.196      2.53    98.2   7.62e-21     1  -938. 1882. 1894.
#>  6 Sales Age    0.0537       0.0514     2.75    22.6   2.79e- 6     1  -971. 1949. 1961.
#>  7 Sales Educ…  0.00270      0.000194   2.82     1.08  3.00e- 1     1  -982. 1970. 1982.
#>  8 Comp… Inco…  0.00754      0.00491   15.1      2.87  9.10e- 2     1 -1571. 3148. 3160.
#>  9 Comp… Adve…  0.000586    -0.00193   15.3      0.233 6.29e- 1     1 -1659. 3324. 3336.
#> 10 Comp… Popu…  0.00897      0.00648   15.3      3.60  5.84e- 2     1 -1657. 3321. 3333.
#> # … with 18 more rows, and 3 more variables: deviance <dbl>, df.residual <int>,
#> #   nobs <int>
#> 

# Compare the correlation that case of joint the Price variable
all_var %>%
  "$"(correlation) %>%
  filter(var1 == "Price" | var2 == "Price") %>%
  arrange(desc(abs(coef_corr)))

#> # A tibble: 7 x 3
#>   var1        var2      coef_corr
#>   <chr>       <chr>         <dbl>
#> 1 CompPrice   Price        0.585 
#> 2 Sales       Price       -0.445 
#> 3 Price       Age         -0.102 
#> 4 Income      Price       -0.0514
#> 5 Advertising Price        0.0445
#> 6 Population  Price       -0.0121
#> 7 Price       Education    0.0117

# Compare the correlation that case of abs(coef_corr) > 0.3
all_var %>%
  "$"(correlation) %>%
  filter(abs(coef_corr) > 0.3)

#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 Sales     Price    -0.445
#> 2 CompPrice Price     0.585

# Compare the linear model that case of joint the Price variable  
all_var %>%
  "$"(linear) %>%
  filter(var1 == "Price" | var2 == "Price") %>%
  arrange(desc(r.squared))

#> # A tibble: 7 x 14
#>   var1  var2  r.squared adj.r.squared  sigma statistic  p.value    df logLik   AIC   BIC
#>   <chr> <chr>     <dbl>         <dbl>  <dbl>     <dbl>    <dbl> <dbl>  <dbl> <dbl> <dbl>
#> 1 Comp… Price  0.342       0.340       12.5   207.     4.50e-38     1 -1575. 3157. 3169.
#> 2 Sales Price  0.198       0.196        2.53   98.2    7.62e-21     1  -938. 1882. 1894.
#> 3 Price Age    0.0104      0.00795     23.6     4.20   4.11e- 2     1 -1831. 3668. 3680.
#> 4 Inco… Price  0.00264     0.00000634  27.8     1.00   3.17e- 1     1 -1802. 3610. 3622.
#> 5 Adve… Price  0.00198    -0.000524     6.65    0.791  3.74e- 1     1 -1325. 2655. 2667.
#> 6 Popu… Price  0.000147   -0.00236    148.      0.0587 8.09e- 1     1 -2564. 5134. 5146.
#> 7 Price Educ…  0.000138   -0.00237     23.7     0.0549 8.15e- 1     1 -1833. 3672. 3684.
#> # … with 3 more variables: deviance <dbl>, df.residual <int>, nobs <int>

# Compare the two numerical variables
two_var <- compare_numeric(carseats, Price, CompPrice)

# Print compare class object
two_var

#> $correlation
#> # A tibble: 1 x 3
#>   var1  var2      coef_corr
#>   <chr> <chr>         <dbl>
#> 1 Price CompPrice     0.585
#> 
#> $linear
#> # A tibble: 1 x 14
#>   var1  var2  r.squared adj.r.squared sigma statistic  p.value    df logLik   AIC   BIC
#>   <chr> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl> <dbl> <dbl>
#> 1 Price Comp…     0.342         0.340  19.2      207. 4.50e-38     1 -1749. 3504. 3516.
#> # … with 3 more variables: deviance <dbl>, df.residual <int>, nobs <int>
#> 

# Summary the all case : Return a invisible copy of an object.
stat <- summary(all_var)

#> ── Correlation check : abs(r) > 0.3 ──────────── Number of pairs is 2/28 ──
#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Price     0.585
#> 2 Sales     Price    -0.445
#> ── R.squared check : R^2 > 0.1 ───────────────── Number of pairs is 2/28 ──
#> # A tibble: 2 x 14
#>   var1  var2  r.squared adj.r.squared sigma statistic  p.value    df logLik   AIC   BIC
#>   <chr> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl> <dbl> <dbl>
#> 1 Comp… Price     0.342         0.340 12.5      207.  4.50e-38     1 -1575. 3157. 3169.
#> 2 Sales Price     0.198         0.196  2.53      98.2 7.62e-21     1  -938. 1882. 1894.
#> # … with 3 more variables: deviance <dbl>, df.residual <int>, nobs <int>

# Just correlation
summary(all_var, method = "correlation")

#> ── Correlation check : abs(r) > 0.3 ──────────── Number of pairs is 2/28 ──
#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Price     0.585
#> 2 Sales     Price    -0.445

# Just correlation condition by r > 0.2
summary(all_var, method = "correlation", thres_corr = 0.2)

#> ── Correlation check : abs(r) > 0.2 ──────────── Number of pairs is 5/28 ──
#> # A tibble: 5 x 3
#>   var1        var2        coef_corr
#>   <chr>       <chr>           <dbl>
#> 1 CompPrice   Price           0.585
#> 2 Sales       Price          -0.445
#> 3 Sales       Advertising     0.270
#> 4 Advertising Population      0.266
#> 5 Sales       Age            -0.232

# linear model summries condition by R^2 > 0.05
summary(all_var, thres_rs = 0.05)

#> ── Correlation check : abs(r) > 0.3 ──────────── Number of pairs is 2/28 ──
#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Price     0.585
#> 2 Sales     Price    -0.445
#> ── R.squared check : R^2 > 0.05 ──────────────── Number of pairs is 5/28 ──
#> # A tibble: 5 x 14
#>   var1  var2  r.squared adj.r.squared sigma statistic  p.value    df logLik   AIC   BIC
#>   <chr> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl> <dbl> <dbl>
#> 1 Comp… Price    0.342         0.340  12.5      207.  4.50e-38     1 -1575. 3157. 3169.
#> 2 Sales Price    0.198         0.196   2.53      98.2 7.62e-21     1  -938. 1882. 1894.
#> 3 Sales Adve…    0.0726        0.0703  2.72      31.2 4.38e- 8     1  -967. 1941. 1953.
#> 4 Adve… Popu…    0.0706        0.0682  6.42      30.2 6.90e- 8     1 -1310. 2627. 2639.
#> 5 Sales Age      0.0537        0.0514  2.75      22.6 2.79e- 6     1  -971. 1949. 1961.
#> # … with 3 more variables: deviance <dbl>, df.residual <int>, nobs <int>

# verbose is FALSE 
summary(all_var, verbose = FALSE)

#> $correlation
#> # A tibble: 2 x 3
#>   var1      var2  coef_corr
#>   <chr>     <chr>     <dbl>
#> 1 CompPrice Price     0.585
#> 2 Sales     Price    -0.445
#> 
#> $linear
#> # A tibble: 2 x 14
#>   var1  var2  r.squared adj.r.squared sigma statistic  p.value    df logLik   AIC   BIC
#>   <chr> <chr>     <dbl>         <dbl> <dbl>     <dbl>    <dbl> <dbl>  <dbl> <dbl> <dbl>
#> 1 Comp… Price     0.342         0.340 12.5      207.  4.50e-38     1 -1575. 3157. 3169.
#> 2 Sales Price     0.198         0.196  2.53      98.2 7.62e-21     1  -938. 1882. 1894.
#> # … with 3 more variables: deviance <dbl>, df.residual <int>, nobs <int>
#>