Summarizing Performance for Optimal Bins

summary method for "optimal_bins". summary metrics to evaluate the performance of binomial classification model.

# S3 method for optimal_bins
summary(object, ...)

Arguments

object
an object of class "optimal_bins", usually, a result of a call to binning_by().

...
further arguments to be passed from or to other methods.

Value

NULL.

Details

print() to print only binning table information of "optimal_bins" objects. summary.performance_bin() includes general metrics and result of significance tests life follows.:

Binning Table : Metrics by bins.

CntRec, CntPos, CntNeg, RatePos, RateNeg, Odds, WoE, IV, JSD, AUC.

General Metrics.

Gini index.

Jeffrey's Information Value.

Jensen-Shannon Divergence.

Kolmogorov-Smirnov Statistics.

Herfindahl-Hirschman Index.

normalized Herfindahl-Hirschman Index.

Cramer's V Statistics.

Table of Significance Tests.

See also

Examples

library(dplyr)

# Generate data for the example
heartfailure2 <- heartfailure
heartfailure2[sample(seq(NROW(heartfailure2)), 5), "creatinine"] <- NA

# optimal binning
bin <- binning_by(heartfailure2, "death_event", "creatinine")

#> Warning: The factor y has been changed to a numeric vector consisting of 0 and 1.
#> 'Yes' changed to 1 (positive) and 'No' changed to 0 (negative).

bin

#> binned type: optimal
#> number of bins: 3
#> x
#> [0.5,0.9] (0.9,1.8] (1.8,9.4]      <NA> 
#>        79       169        46         5 

# summary optimal_bins class
summary(bin)

#> ── Binning Table ──────────────────────── Several Metrics ──
#>         Bin CntRec CntPos CntNeg RatePos RateNeg    Odds      WoE      IV
#> 1 [0.5,0.9]     79      9     70 0.09375 0.34483 0.12857 -1.30241 0.32701
#> 2 (0.9,1.8]    169     52    117 0.54167 0.57635 0.44444 -0.06207 0.00215
#> 3 (1.8,9.4]     46     33     13 0.34375 0.06404 2.53846  1.68042 0.47003
#> 4      <NA>      5      2      3 0.02083 0.01478 0.66667  0.34339 0.00208
#> 5     Total    299     96    203 1.00000 1.00000 0.47291       NA 0.80127
#>       JSD     AUC
#> 1 0.03821 0.01616
#> 2 0.00027 0.21013
#> 3 0.05269 0.05170
#> 4 0.00026 0.01462
#> 5 0.09143 0.29262
#> 
#> ── General Metrics ─────────────────────────────────────────
#> ● Gini index                       :  -0.41477
#> ● IV (Jeffrey)                     :  0.80127
#> ● JS (Jensen-Shannon) Divergence   :  0.09143
#> ● Kolmogorov-Smirnov Statistics    :  0.28577
#> ● HHI (Herfindahl-Hirschman Index) :  0.41323
#> ● HHI (normalized)                 :  0.21764
#> ● Cramer's V                       :  0.408
#> 
#> ── Significance Tests ──────────────────── Chisquare Test ──
#>       Bin A     Bin B statistics      p_value
#> 1 [0.5,0.9] (0.9,1.8]   10.89831 9.625185e-04
#> 2 (0.9,1.8] (1.8,9.4]   25.38929 4.685125e-07
#> 

# performance table
attr(bin, "performance")

#>         Bin CntRec CntPos CntNeg CntCumPos CntCumNeg RatePos RateNeg RateCumPos
#> 1 [0.5,0.9]     79      9     70         9        70 0.09375 0.34483    0.09375
#> 2 (0.9,1.8]    169     52    117        61       187 0.54167 0.57635    0.63542
#> 3 (1.8,9.4]     46     33     13        94       200 0.34375 0.06404    0.97917
#> 4      <NA>      5      2      3        96       203 0.02083 0.01478    1.00000
#> 5     Total    299     96    203        NA        NA 1.00000 1.00000         NA
#>   RateCumNeg    Odds   LnOdds      WoE      IV     JSD     AUC
#> 1    0.34483 0.12857 -2.05127 -1.30241 0.32701 0.03821 0.01616
#> 2    0.92118 0.44444 -0.81093 -0.06207 0.00215 0.00027 0.21013
#> 3    0.98522 2.53846  0.93156  1.68042 0.47003 0.05269 0.05170
#> 4    1.00000 0.66667 -0.40547  0.34339 0.00208 0.00026 0.01462
#> 5         NA 0.47291 -0.74886       NA 0.80127 0.09143 0.29262

# visualize all information for optimal_bins class
# plot(bin)

# visualize WoE information for optimal_bins class
# plot(bin, type = "WoE")

# visualize all information without typographic
# plot(bin, typographic = FALSE)

# extract binned results
extract(bin) %>%
  head(20)

#>  [1] (1.8,9.4] (0.9,1.8] (0.9,1.8] (1.8,9.4] <NA>      (1.8,9.4] (0.9,1.8]
#>  [8] (0.9,1.8] (0.9,1.8] (1.8,9.4] (1.8,9.4] [0.5,0.9] (0.9,1.8] (0.9,1.8]
#> [15] (0.9,1.8] (0.9,1.8] [0.5,0.9] [0.5,0.9] (0.9,1.8] (1.8,9.4]
#> Levels: [0.5,0.9] < (0.9,1.8] < (1.8,9.4]