Summarizing univar_numeric information

print and summary method for "univar_numeric" class.

# S3 method for univar_numeric
summary(object, stand = c("robust", "minmax", "zscore"), ...)

# S3 method for univar_numeric
print(x, ...)

Arguments

object
an object of class "univar_numeric", usually, a result of a call to univar_numeric().

stand
character Describe how to standardize the original data. "robust" normalizes the raw data through transformation calculated by IQR and median. "minmax" normalizes the original data using minmax transformation. "zscore" standardizes the original data using z-Score transformation. The default is "robust".

...
further arguments passed to or from other methods.

x
an object of class "univar_numeric", usually, a result of a call to univar_numeric().

Value

An object of the class as indivisual variabes based list. The statistics returned by summary.univar_numeric() are different from the statistics returned by univar_numeric(). univar_numeric() is the statistics for the original data, but summary. univar_numeric() is the statistics for the standardized data. A component named "statistics" is a tibble object with the following statistics.:

variable : factor. The level of the variable. 'variable' is the name of the variable.

n : number of observations excluding missing values

na : number of missing values

mean : arithmetic average

sd : standard deviation

se_mean : standard error mean. sd/sqrt(n)

IQR : interquartile range (Q3-Q1)

skewness : skewness

kurtosis : kurtosis

median : median. 50% percentile

Details

print.univar_numeric() displays only the information of variables included in univar_numeric The "variables" attribute is not displayed.

See also

Examples

# Generate data for the example
carseats <- ISLR::Carseats
carseats[sample(seq(NROW(carseats)), 20), "Income"] <- NA
carseats[sample(seq(NROW(carseats)), 5), "Urban"] <- NA

library(dplyr)

# Calculates the all categorical variavels
all_var <- univar_numeric(carseats)

# Print univar_numeric class object
all_var

#> $statistics
#> # A tibble: 8 x 10
#>   variable        n    na   mean     sd se_mean    IQR skewness kurtosis median
#>   <chr>       <int> <int>  <dbl>  <dbl>   <dbl>  <dbl>    <dbl>    <dbl>  <dbl>
#> 1 Sales         400     0   7.50   2.82   0.141   3.93   0.186   -0.0809   7.49
#> 2 CompPrice     400     0 125.    15.3    0.767  20     -0.0428   0.0417 125   
#> 3 Income        380    20  68.6   27.9    1.43   46.2    0.0585  -1.08    69   
#> 4 Advertising   400     0   6.64   6.65   0.333  12      0.640   -0.545    5   
#> 5 Population    400     0 265.   147.     7.37  260.    -0.0512  -1.20   272   
#> 6 Price         400     0 116.    23.7    1.18   31     -0.125    0.452  117   
#> 7 Age           400     0  53.3   16.2    0.810  26.2   -0.0772  -1.13    54.5 
#> 8 Education     400     0  13.9    2.62   0.131   4      0.0440  -1.30    14   
#> 

# Calculates the Price, CompPrice variable
univar_numeric(carseats, Price, CompPrice)

#> $statistics
#> # A tibble: 2 x 10
#>   variable      n    na  mean    sd se_mean   IQR skewness kurtosis median
#>   <chr>     <int> <int> <dbl> <dbl>   <dbl> <dbl>    <dbl>    <dbl>  <dbl>
#> 1 Price       400     0  116.  23.7   1.18     31  -0.125    0.452     117
#> 2 CompPrice   400     0  125.  15.3   0.767    20  -0.0428   0.0417    125
#> 

# Summary the all case : Return a invisible copy of an object.
stat <- summary(all_var)

# Summary by returned object
stat

#> # A tibble: 8 x 8
#>   variable        mean    sd se_mean   IQR skewness kurtosis median
#>   <chr>          <dbl> <dbl>   <dbl> <dbl>    <dbl>    <dbl>  <dbl>
#> 1 Sales        0.00161 0.719  0.0359     1   0.186   -0.0809      0
#> 2 CompPrice   -0.00125 0.767  0.0383     1  -0.0428   0.0417      0
#> 3 Income      -0.00785 0.603  0.0310     1   0.0585  -1.08        0
#> 4 Advertising  0.136   0.554  0.0277     1   0.640   -0.545       0
#> 5 Population  -0.0276  0.568  0.0284     1  -0.0512  -1.20        0
#> 6 Price       -0.0389  0.764  0.0382     1  -0.125    0.452       0
#> 7 Age         -0.0449  0.617  0.0309     1  -0.0772  -1.13        0
#> 8 Education   -0.025   0.655  0.0328     1   0.0440  -1.30        0

# Statistics of numerical variables normalized by Min-Max method
summary(all_var, stand = "minmax")

#> # A tibble: 8 x 8
#>   variable     mean    sd se_mean   IQR skewness kurtosis median
#>   <chr>       <dbl> <dbl>   <dbl> <dbl>    <dbl>    <dbl>  <dbl>
#> 1 Sales       0.461 0.174 0.00868 0.242   0.186   -0.0809  0.460
#> 2 CompPrice   0.490 0.156 0.00782 0.204  -0.0428   0.0417  0.490
#> 3 Income      0.481 0.282 0.0145  0.467   0.0585  -1.08    0.485
#> 4 Advertising 0.229 0.229 0.0115  0.414   0.640   -0.545   0.172
#> 5 Population  0.511 0.295 0.0148  0.520  -0.0512  -1.20    0.525
#> 6 Price       0.550 0.142 0.00709 0.186  -0.125    0.452   0.557
#> 7 Age         0.515 0.295 0.0147  0.477  -0.0772  -1.13    0.536
#> 8 Education   0.488 0.328 0.0164  0.5     0.0440  -1.30    0.5  

# Statistics of numerical variables standardized by Z-score method
summary(all_var, stand = "zscore")

#> # A tibble: 8 x 8
#>   variable         mean    sd se_mean   IQR skewness kurtosis   median
#>   <chr>           <dbl> <dbl>   <dbl> <dbl>    <dbl>    <dbl>    <dbl>
#> 1 Sales        1.15e-16  1.00  0.0500  1.39   0.186   -0.0809 -0.00224
#> 2 CompPrice    3.67e-16  1     0.05    1.30  -0.0428   0.0417  0.00163
#> 3 Income       6.96e-17  1     0.0513  1.66   0.0585  -1.08    0.0130 
#> 4 Advertising  2.03e-17  1     0.05    1.80   0.640   -0.545  -0.246  
#> 5 Population   1.67e-16  1.00  0.0500  1.76  -0.0512  -1.20    0.0486 
#> 6 Price       -7.39e-17  1     0.05    1.31  -0.125    0.452   0.0509 
#> 7 Age          1.12e-16  1.00  0.0500  1.62  -0.0772  -1.13    0.0727 
#> 8 Education   -1.38e-16  1     0.05    1.53   0.0440  -1.30    0.0382